perf(background): slim background i18n to section subset (#505/#506)
Split locale resources into per-section modules and introduce
core/i18n/background.ts: a translator over only the message sections
reachable from background-bundled modules (content/background/tool/prompt).
Background-side modules now import the slim slice, keeping the sidepanel
UI tree (sidepanel/app/manifest/locale/pet/common) out of the service
worker bundle. The background key type is narrowed so an unavailable key
fails compilation instead of throwing at runtime.

Result (chrome-mv3 background.js):
- raw 815390 -> 720903 (-94587 B, -11.6%); gzip 234596 -> 209156
- #505/#506 target (<790000 raw / <225000 gzip) met; backgroundRawBytesMax
  stays 820000 with real headroom restored.

Also:
- core/i18n/runtime.ts: resources-free resolution core shared by the full
  and slim translators (single logic truth).
- core/i18n/store.ts now imports the resource-free runtime.
- tests/i18n-background-slice.test.ts: guards every background-reachable
  translate key against the slim resource set.
- scripts/i18n-coverage-audit.mjs: allowlist covers the new zh-CN section
  modules; locale.ts line rule follows the split.
- scripts/sidepanel-chunk-budget.mjs: re-baseline initial shell
  (375456/112742 -> 375672/114387) and firstChatScreen cap
  (405800/123100 -> 406000/125000): the refactor regrouped rollup chunks
  (+216 raw, +0.06%; ChatPage itself unchanged). Background budget is not
  touched.
- tests/tool-provider-import-boundary.test.ts: runtime.ts now imports the
  slim i18n entry.
