+++ core/i18n/resources/en.ts
@@ -1,1528 +1,24 @@
 import type { LocaleMessages } from './zh-CN';
+import { common } from './en/common';
+import { manifest } from './en/manifest';
+import { app } from './en/app';
+import { locale } from './en/locale';
+import { sidepanel } from './en/sidepanel';
+import { content } from './en/content';
+import { background } from './en/background';
+import { tool } from './en/tool';
+import { prompt } from './en/prompt';
+import { pet } from './en/pet';
 
 export const en = {
-  common: {
-    add: 'Add',
-    all: 'All',
-    auto: 'Auto',
-    cancel: 'Cancel',
-    clear: 'Clear',
-    close: 'Close',
-    confirm: 'Confirm',
-    delete: 'Delete',
-    deactivate: 'Disable',
-    disabled: 'Disabled',
-    edit: 'Edit',
-    enable: 'Enable',
-    enabled: 'Enabled',
-    error: 'Error',
-    loading: 'Loading...',
-    manual: 'Manual',
-    none: 'None',
-    open: 'Open',
-    preview: 'Preview',
-    refresh: 'Refresh',
-    remove: 'Remove',
-    retry: 'Retry',
-    save: 'Save',
-    saveChanges: 'Save changes',
-    search: 'Search',
-    status: 'Status',
-    success: 'Success',
-    sync: 'Sync',
-    test: 'Test',
-    update: 'Update',
-  },
-  manifest: {
-    name: 'DeepSeek++',
-    description: 'Agentic memory, skills, execution, automation, and MCP tools for DeepSeek',
-    actionTitle: 'DeepSeek++',
-  },
-  app: {
-    sideNavLabel: 'Sidebar navigation',
-    version: 'v{version}',
-    tabs: {
-      chat: 'Chat',
-      library: 'Library',
-      memory: 'Memory',
-      projects: 'Projects',
-      saved: 'Saved',
-      capabilities: 'Capabilities',
-      preset: 'Presets',
-      automation: 'Automation',
-      settings: 'Settings',
-    },
-  },
-  locale: {
-    settingTitle: 'Language',
-    settingDescription: 'Choose the language for DeepSeek++ UI and model behavior.',
-    auto: 'Follow browser',
-    zhCN: '简体中文',
-    en: 'English',
-  },
-  sidepanel: {
-    subTabs: {
-      scrollLeft: 'Scroll tabs left',
-      scrollRight: 'Scroll tabs right',
-    },
-    libraryPage: {
-      navLabel: 'Library navigation',
-      tabs: {
-        memory: 'Memory',
-        saved: 'Saved',
-      },
-    },
-    memory: {
-      actions: {
-        pin: 'Pin',
-        unpin: 'Unpin',
-      },
-      age: {
-        justNow: 'Just now',
-        minutesAgo: '{count}m ago',
-        hoursAgo: '{count}h ago',
-        daysAgo: '{count}d ago',
-      },
-      form: {
-        namePlaceholder: 'Title',
-        contentPlaceholder: 'Content',
-        tagsPlaceholder: 'Tags, comma-separated',
-      },
-      types: {
-        user: 'User',
-        feedback: 'Feedback',
-        topic: 'Topic',
-        reference: 'Reference',
-      },
-    },
-    preset: {
-      activeBadge: 'Active',
-      form: {
-        namePlaceholder: 'Preset name, e.g. Code assistant or Translation expert',
-        contentLabel: 'System prompt content',
-        contentPlaceholder: 'You are a professional...\n\n## Core principles\n- ...',
-      },
-    },
-    skill: {
-      disabledBadge: 'Disabled',
-      memoryEnabledBadge: 'Memory injection',
-      unknownLicense: 'Unknown license',
-      actions: {
-        enableSkill: 'Enable {name}',
-        disableSkill: 'Disable {name}',
-        editSkill: 'Edit {name}',
-        deleteSkill: 'Delete {name}',
-        updateSkill: 'Update local Skill (re-read index)',
-      },
-      autoActivation: {
-        title: 'Auto-activation',
-        description: 'When on, DeepSeek++ matches and activates local Skills automatically from the conversation. When off, a Skill activates only when you explicitly call /skill-name.',
-        firstMessage: 'Auto-activate on first message of new chat',
-        firstMessageDescription: 'Only match and activate a local Skill on the first message of a new chat.',
-        everyMessage: 'Auto-activate on every message of current chat',
-        everyMessageDescription: 'Match and activate on every message of the current chat; enabling this also enables “first message”.',
-      },
-      form: {
-        namePlaceholder: 'Name, e.g. my-skill',
-        triggerCommand: 'Trigger command:',
-        descriptionPlaceholder: 'Description, when to use this skill',
-        instructionsLabel: 'Instructions (Markdown, tell AI how to execute)',
-        instructionsPlaceholder: 'You are a...\n\n## Core principles\n- ...',
-        memoryInjectionLabel: 'Enable memory injection',
-      },
-      sources: {
-        builtin: 'Built-in',
-        official: 'Third-party',
-        thirdParty: 'Third-party',
-        custom: 'Custom',
-        remote: 'GitHub',
-        local: 'Local',
-      },
-    },
-    scenario: {
-      title: 'Context menu scenarios',
-      description: 'Select text and right-click to send it to sidepanel chat',
-      customTitle: 'Custom scenarios',
-      namePlaceholder: 'Scenario name',
-      templatePlaceholder: 'Prompt template with {text}',
-      operationFailed: 'Scenario action failed: {error}',
-      backgroundUnavailable: 'Background context menu refresh failed.',
-      savedButReloadFailed: 'Scenario saved, but the list could not refresh: {error}',
-      savedButMenuFailed: 'Scenario saved, but the background context menu could not refresh: {error}',
-    },
-    githubSkillImport: {
-      title: 'Import Skill from GitHub',
-      description: 'Supports repositories, directories, and single SKILL.md links. DeepSeek++ previews content before import and never overwrites local custom Skills.',
-      urlPlaceholder: 'https://github.com/owner/repo or .../SKILL.md',
-      permissionError: 'GitHub API access permission is required to read repository Skills',
-      previewFailed: 'Preview failed',
-      importFailed: 'Import failed',
-      importedMessage: 'Imported {count} Skills',
-      selectAll: 'Select all',
-      clearSelection: 'Clear selection',
-      selectedSummary: 'Selected {selected} / {total} · {bytes}',
-      importSelected: 'Import selected Skills',
-      openRepository: 'Open repository',
-      repoRoot: 'repo root',
-      unknownLicense: 'Unknown',
-      warningOverflow: '{count} more warnings',
-      renamedBadge: 'Renamed',
-      resourceCount: '{count} resources',
-      omittedCount: '{count} omitted',
-      renamedNotice: '{count} Skills were renamed automatically because of naming conflicts.',
-      meta: {
-        license: 'License',
-        version: 'Version',
-        skill: 'Skill',
-        defaultBranch: 'Default',
-      },
-    },
-    localSkillImport: {
-      title: 'Import Local Skill',
-      description: 'Reads a local Skill folder through Shell MCP. DeepSeek++ keeps the folder reference and never executes scripts during import.',
-      pathPlaceholder: '/Users/me/.codex/skills/my-skill',
-      pickFolder: 'Choose',
-      pickFailed: 'Folder selection failed',
-      previewFailed: 'Preview failed',
-      importFailed: 'Import failed',
-      importedMessage: 'Imported {count} local Skills',
-      selectAll: 'Select all',
-      clearSelection: 'Clear selection',
-      selectedSummary: 'Selected {selected} / {total} · {bytes}',
-      importSelected: 'Import selected Skills',
-      referencedMode: 'Referenced folder',
-      warningOverflow: '{count} more warnings',
-      renamedBadge: 'Renamed',
-      resourceCount: '{count} bundled',
-      scriptCount: '{count} scripts',
-      omittedCount: '{count} on demand',
-      notBundledCount: '{count} not bundled',
-      omittedExplanation: '{count} supporting files were not bundled into context because a count or size limit was reached. They were not deleted and remain available in the local Skill folder for on-demand reads through Shell MCP.',
-      omittedUnavailableExplanation: '{count} supporting files were not bundled into context. They were not deleted and remain in the local Skill folder, but on-demand reads are currently unavailable. Fix the affected Shell MCP reader before importing those Skills.',
-      readerUnavailable: 'On-demand reader unavailable',
-      readerTechnicalDetail: 'Details: {detail}',
-      readerActions: {
-        updateHost: 'Reinstall Shell Native Host from MCP > Shell Local, restart the browser, then preview again.',
-        enableReader: 'Set Shell Local execution mode to Auto and allow local_file_read, then preview again.',
-        checkConnection: 'Check the Shell Local connection and preview again.',
-      },
-      renamedNotice: '{count} Skills were renamed automatically because of naming conflicts.',
-      meta: {
-        skill: 'Skill',
-        mode: 'Mode',
-      },
-    },
-    projectsPage: {
-      title: 'Projects',
-      description: 'Add DeepSeek conversations to projects. Project instructions and project memories are injected for those conversations.',
-      summary: '{projects} projects, {conversations} project conversations',
-      createTitle: 'Create project',
-      namePlaceholder: 'Project name',
-      descriptionPlaceholder: 'Project description (optional)',
-      instructionsPlaceholder: 'Project instructions',
-      createProject: 'Create project',
-      projectCreated: 'Project “{name}” created',
-      empty: 'No projects',
-      emptyHelp: 'Create a project, then add the current DeepSeek conversation or assign the next new conversation to it.',
-      deleteProject: 'Delete',
-      deleteConfirm: 'Delete project "{name}"?',
-      conversationCount: '{count} conversations',
-      pendingBadge: 'Next',
-      detailMeta: '{conversations} conversations, {memories} project memories',
-      currentConversation: 'Current conversation',
-      noCurrentConversation: 'The active tab is not a DeepSeek conversation.',
-      addCurrentConversation: 'Add current conversation',
-      moveCurrentConversation: 'Move to this project',
-      useNextConversation: 'Use this project for the next new conversation',
-      cancelNextConversation: 'Cancel next conversation',
-      conversationsTitle: 'Project conversations',
-      emptyConversations: 'No conversations belong to this project yet.',
-      memoriesTitle: 'Project memories',
-      emptyMemories: 'No project memories yet.',
-      backendUnavailable: 'Project backend is unavailable. Reload the extension and try again.',
-      operationFailed: 'Project action failed: {error}',
-    },
-    savedPage: {
-      title: 'Saved items',
-      description: 'Save reusable prompts, answer snippets, and reference links. Search them, insert them into chat, or export a backup.',
-      count: '{count} saved',
-      snippet: 'Snippet',
-      bookmark: 'Bookmark',
-      titlePlaceholder: 'Title',
-      contentPlaceholder: 'Prompt snippet, note, or reusable text',
-      tagsPlaceholder: 'Tags, comma-separated',
-      save: 'Save item',
-      searchPlaceholder: 'Search saved items',
-      empty: 'No saved items',
-      emptyHelp: 'Save your first snippet to reuse fixed instructions, workflows, or reference material.',
-      insertPrompt: 'Insert into chat',
-      exportMarkdown: 'Export Markdown',
-      exportJson: 'Export JSON',
-      saved: 'Saved',
-      inserted: 'Inserted into the current DeepSeek chat',
-      insertFailed: 'Could not insert into chat: {error}',
-      deleteConfirm: 'Delete this saved item?',
-      backendUnavailable: 'Saved items backend is unavailable. Reload the extension and try again.',
-      operationFailed: 'Saved item action failed: {error}',
-    },
-    whatsNew: {
-      title: "What's new",
-      subtitle: 'This release adds Browser Control and tightens side-panel feedback, third-party Skill management, and permission copy.',
-      versionBadge: 'v{version}',
-      dismiss: 'Got it',
-      items: {
-        browserControl: 'Browser Control can select a target tab from the side panel and perform visible web actions for DeepSeek++.',
-        browserControlBoundary: 'Browser Control stays off by default and only adds text snapshots and browser action tools to new conversations after it is enabled.',
-        sidepanelFeedback: 'Saved items, projects, and prompt controls now show clearer success, failure, and unavailable-state feedback.',
-        thirdPartySkills: 'Third-party Skills are grouped by source and can be enabled, disabled, checked for updates, or synced independently.',
-      },
-    },
-    memoryPage: {
-      title: 'Memory',
-      description: 'Manage the preferences, feedback, topic context, and references that influence future DeepSeek++ replies.',
-      emptyAll: 'No memories yet. They will accumulate automatically during conversations.',
-      emptyFiltered: 'No memories in this category',
-      emptyHelp: 'DeepSeek++ saves durable preferences and project context from conversation, and you can add entries manually.',
-      count: '{count} memories total',
-      deleteConfirm: 'Delete this memory?',
-      backendUnavailable: 'Memory storage is unavailable. Reload the extension and try again.',
-      operationFailed: 'Memory operation failed: {error}',
-    },
-    presetPage: {
-      title: 'System prompt presets',
-      description: 'Prepare fixed system roles or behavior boundaries for new chats. The active preset is injected before the first message.',
-      import: 'Import',
-      create: 'New',
-      activeMeta: 'Preset active',
-      inactiveMeta: 'No active preset',
-      empty: 'No presets',
-      emptyHelp: 'Create a system prompt preset, then select it to inject automatically before the first message of each new conversation.',
-      activeHelp: 'When a preset is enabled, its prompt is injected before the first message of each new conversation. Only one preset can be active at a time.',
-      deleteConfirm: 'Delete this preset?',
-      backendUnavailable: 'Preset storage is unavailable. Reload the extension and try again.',
-      operationFailed: 'Preset operation failed: {error}',
-    },
-    skillPage: {
-      title: 'Available Skills',
-      description: 'Skills are reusable expert workflows. Type a /skill command to bring task-specific rules into the current chat.',
-      createCustom: 'Custom',
-      importLocal: 'Local',
-      summary: '{total} Skills, {enabled} enabled',
-      sectionGithub: 'GitHub imports',
-      sectionBuiltin: 'Built-in',
-      sectionOfficial: 'Third-party',
-      sectionThirdParty: 'Third-party',
-      sectionCustom: 'Custom',
-      bundledThirdPartySource: 'Bundled with the extension, disabled by default',
-      unknownThirdPartySource: 'Unknown third-party source',
-      enabledSkillCount: '{enabled}/{total} enabled',
-      enableSourceSkills: 'Enable all',
-      disableSourceSkills: 'Disable all',
-      expandSource: 'Expand {source}',
-      collapseSource: 'Collapse {source}',
-      usagePrefix: 'Type',
-      usageTrigger: '/skill-name arguments',
-      usageSuffix: 'in the DeepSeek input to trigger a Skill. Example:',
-      usageExample: '/frontend-design build a login page',
-      githubSourceTitle: 'GitHub sources',
-      localReferencedSource: 'Local referenced folder',
-      repoRoot: 'repo root',
-      check: 'Check',
-      checking: 'Checking...',
-      sync: 'Sync',
-      syncing: 'Syncing...',
-      remove: 'Remove',
-      skillCount: '{count} Skills',
-      unknownLicense: 'Unknown license',
-      syncedAt: 'Synced {time}',
-      checkedAt: 'Checked {time}',
-      checkPermissionError: 'GitHub API access permission is required to check for updates',
-      syncPermissionError: 'GitHub API access permission is required to sync updates',
-      checkFailed: 'Update check failed',
-      syncFailed: 'Sync failed',
-      syncedSkills: 'Synced {count} Skills',
-      deleteSourceConfirm: 'Remove {count} Skills imported from {repository}?',
-      noUpdates: 'No upstream updates found',
-      changedUpdates: '{count} imported Skills may have updates',
-      newSkills: '{count} new Skills',
-      missingSkills: '{count} imported Skills disappeared upstream',
-      updatesFound: 'Upstream updates found',
-      backendUnavailable: 'Skill storage is unavailable. Reload the extension and try again.',
-      operationFailed: 'Skill operation failed: {error}',
-      relocatePrompt: 'The original folder was moved or no longer exists. Please reselect the local Skill folder.',
-    },
-    settings: {
-      title: 'Settings',
-      description: 'Adjust model mode, sidepanel chat, prompt injection, appearance, sync, and local data management.',
-      tabs: {
-        general: 'General',
-        api: 'API',
-        prompt: 'Prompt',
-        voice: 'Voice',
-        appearance: 'Appearance',
-        usage: 'Usage',
-        data: 'Data',
-        projectFiles: 'Project Folders',
-        about: 'About',
-      },
-      navLabel: 'Settings sub-navigation',
-      generalDescription: 'Interface language and webpage behavior toggles.',
-      apiDescription: 'API keys and request URLs for sidepanel chat and multimodal analysis.',
-      promptDescription: 'Control how memory, system prompt and presets are injected.',
-      voiceDescription: 'Voice input and read-aloud for replies.',
-      appearanceDescription: 'Background and floating pet on DeepSeek pages.',
-      usageDescription: 'Review recent model tokens, activity, and model distribution.',
-      dataDescription: 'Cloud sync plus local memory import, export and clear.',
-      projectFilesDescription: 'Authorize a local project folder for @ file references in chat.',
-      projectFilesSection: 'Trusted folder',
-      projectFilesSectionDescription: 'The folder is scanned locally only. Local paths are never uploaded; file content is read on demand when sent.',
-      projectFilesChoose: 'Choose folder',
-      projectFilesChooseAgain: 'Choose again',
-      projectFilesRemove: 'Remove authorization',
-      projectFilesRemoveConfirm: 'The files will no longer be referenceable with @ in chat. Remove authorization?',
-      projectFilesAuthorized: 'Authorized',
-      projectFilesSessionExpired: 'Authorization expired, choose the folder again',
-      projectFilesFiles: '{count} files',
-      projectFilesSkipped: '{count} skipped',
-      projectFilesTruncated: 'Index limit reached, keeping the first {count} files',
-      projectFilesImagesReady: '{count} images available for @ references',
-      projectFilesTextPending: '{count} text files, content references come in a later version',
-      projectFilesOther: '{count} other',
-      projectFilesPickFailed: 'Failed to read the folder, please retry.',
-      interfaceSection: 'Interface',
-      interfaceLanguage: 'Interface language',
-      interfaceLanguageDescription: 'Follow Chrome or browser language by default, or pin the UI manually.',
-      languageAuto: 'Follow',
-      languageChinese: 'Simplified Chinese',
-      languageEnglish: 'English',
-      languageCurrent: 'Current: {language}',
-      apiKeyRequired: 'Enter a DeepSeek API Key',
-      saveFailed: 'Save failed',
-      clearFailed: 'Clear failed',
-      apiKeySaved: 'Saved. Context menu scenarios can now run on regular webpages.',
-      apiKeyCleared: 'Cleared. Context menu scenarios are limited to DeepSeek pages again.',
-      webDavPermissionDenied: 'Permission is required to connect to the WebDAV server',
-      operationFailed: 'Operation failed',
-      syncConfigChanged: 'Sync settings changed. Review the current target and try again.',
-      syncRemoteCommittedWarning: 'The remote operation completed, but local sync bookkeeping could not be confirmed. Do not retry immediately; the authoritative sync config has been reloaded.',
-      syncCounts: '{memories} memories, {skills} Skills, {presets} presets, {projects} projects, {projectConversations} project conversations, {savedItems} saved items',
-      connectionSuccess: 'Connection succeeded',
-      connectionFailed: 'Connection failed',
-      uploadConfirm: 'Overwrite remote data with local memories, Skills, presets, projects, and saved items?',
-      uploadSuccess: 'Upload complete. Remote data was overwritten. {counts}',
-      uploadFailed: 'Upload failed',
-      downloadConfirm: 'Overwrite local memories, Skills, presets, projects, and saved items with remote data? This cannot be undone.',
-      downloadSuccess: 'Download complete. Local data was overwritten. {counts}',
-      downloadFailed: 'Download failed',
-      importMemoryArrayError: 'Import file must be a memory array',
-      importSuccess: 'Imported {count} memories',
-      jsonFormatError: 'Invalid JSON format',
-      clearAllConfirm: 'Clear all memories? This cannot be undone.',
-      neverSynced: 'Never synced',
-      modelSection: 'Chat settings',
-      modelMode: 'Web model mode',
-      modelModeDescription: 'Choose the official mode used by chat.deepseek.com conversations. Sidepanel API chat still chooses the model on the Chat page.',
-      modelDefault: 'Default',
-      modelExpert: 'Expert',
-      modelVision: 'Vision',
-      sidepanelChat: 'Sidepanel chat',
-      sidepanelChatDescription: 'Show the chat tab in the sidepanel with webpage login or official API Key support',
-      apiKeyDescription: 'After configuration, context menu scenarios can run on regular webpages',
-      configured: 'Configured',
-      notConfigured: 'Not configured',
-      apiKeyReplacePlaceholder: 'Enter a new key to replace it',
-      multimodalApi: 'Multimodal API',
-      multimodalApiDescription: 'Configure OpenAI image analysis, Gemini video analysis, and custom request URLs for the built-in Multimodal MCP.',
-      openaiKeyReplacePlaceholder: 'Enter a new OpenAI key to replace it',
-      geminiKeyReplacePlaceholder: 'Enter a new Gemini key to replace it',
-      openaiImageModel: 'OpenAI image model',
-      geminiVideoModel: 'Gemini video model',
-      openaiBaseUrl: 'OpenAI request URL',
-      geminiBaseUrl: 'Gemini request URL',
-      multimodalBaseUrlInvalid: 'OpenAI / Gemini request URLs must be valid http(s) URLs',
-      multimodalSaved: 'Multimodal API settings saved',
-      multimodalCleared: 'Multimodal API settings cleared',
-      clearMultimodalApi: 'Clear multimodal settings',
-      saving: 'Saving',
-      clearing: 'Clearing',
-      clearApiKey: 'Clear API Key',
-      petWhale: 'DeepSeek whale',
-      petWhaleDescription: 'Show a state-aware pet on DeepSeek pages',
-      backgroundSection: 'Background settings',
-      customBackground: 'Custom background',
-      customBackgroundDescription: 'Set a background image for DeepSeek pages',
-      uploadImage: 'Upload image',
-      imageUrlPlaceholder: 'Paste image URL',
-      backgroundPreviewAlt: 'Background preview',
-      backgroundPreviewOverlay: 'Preview simulation',
-      backgroundOpacity: 'Background opacity',
-      clearBackground: 'Clear background',
-      floatingPetSection: 'Floating pet',
-      positionBottomRight: 'Bottom right',
-      positionBottomLeft: 'Bottom left',
-      positionCustom: 'Custom',
-      size: 'Size',
-      opacity: 'Opacity',
-      petMotion: 'Floating motion',
-      petMotionDescription: 'Turn off to reduce movement',
-      floatingChatSection: 'Global floating chat',
-      floatingChat: 'Floating chat ball',
-      floatingChatDescription: 'Show a draggable chat ball on every website. Click it to open the chat window; long-press to move it.',
-      floatingChatPermissionMissing: 'Website access was not granted, so floating chat remains disabled. Enable it again to retry authorization.',
-      floatingChatContextInvalidated: 'The extension context is invalid. Reload the extension before changing floating chat.',
-      cloudSyncSection: 'Cloud sync',
-      syncProvider: 'Sync method',
-      providerWebdav: 'WebDAV',
-      providerGdrive: 'Google Drive',
-      providerOnedrive: 'OneDrive',
-      webDavUrl: 'WebDAV URL',
-      username: 'Username',
-      password: 'Password',
-      remotePath: 'Remote path',
-      clientId: 'Client ID',
-      clientSecret: 'Client secret',
-      redirectUri: 'Redirect URI',
-      redirectUriUnavailable: 'OAuth redirect URI is not available in this browser context',
-      redirectUriHint: 'Add this exact URI to your OAuth app\'s authorized redirect URIs.',
-      authorize: 'Authorize',
-      authorized: 'Authorized',
-      reauthorize: 'Re-authorize',
-      authorizeSuccess: 'Authorization successful',
-      oauthSetupHint: 'Create an OAuth app in {link} and paste the client ID and secret here.',
-      gdriveConsole: 'Google Cloud Console',
-      onedriveConsole: 'Azure portal',
-      testConnection: 'Test connection',
-      uploadLocal: 'Upload local',
-      downloadRemote: 'Download remote',
-      lastSync: 'Last sync: {time}',
-      dataSection: 'Data management',
-      memoryTotal: 'Total memories',
-      exportMemories: 'Export memories',
-      exportLogs: 'Export diagnostic logs',
-      exportLogsHint: 'Exports recent background tool execution and error diagnostics for troubleshooting shell_exec / MCP / authorization issues.',
-      importMemories: 'Import memories',
-      clearAllMemories: 'Clear all memories',
-      usage: {
-        rangeLabel: 'Time range',
-        last7Days: 'Last 7 days',
-        last30Days: 'Last 30 days',
-        loadFailed: 'Failed to load usage statistics',
-        clearStats: 'Clear stats',
-        clearConfirm: 'Clear local usage statistics? Memories, conversations, and cloud sync data are not affected.',
-        clearSuccess: 'Usage statistics cleared',
-        clearFailed: 'Failed to clear usage statistics',
-        emptyTitle: 'No usage statistics yet',
-        emptyDescription: 'After the next completed DeepSeek webpage response, this page will start tracking tokens, TPS, active days, and model distribution.',
-        totalTokens: 'Tokens used',
-        serverSamples: 'Server samples {server}/{total}',
-        sessions: 'Sessions',
-        turns: '{count} replies',
-        messages: 'Messages',
-        activeDays: '{count} active days',
-        currentStreak: 'Current streak',
-        daysUnit: 'days',
-        mostUsedModel: 'Most used model',
-        noModel: 'No model',
-        share: '{percent} share',
-        heatmapTitle: 'Activity heatmap',
-        less: 'Less',
-        more: 'More',
-        dailyTrendTitle: 'Daily token trend',
-        modelUsageTitle: 'Model usage',
-      },
-      aboutSection: 'About',
-      aboutTagline: 'Agentic memory and Skill system',
-    },
-    promptControls: {
-      title: 'Prompt controls',
-      memory: 'Memory injection',
-      memoryDescription: 'Include selected long-term memories in new requests',
-      systemPrompt: 'DeepSeek++ system prompt',
-      systemPromptDescription: 'Include tool schemas, memory rules, and runtime guidance',
-      presetCadence: 'Preset cadence',
-      cadenceDefault: 'Default cadence',
-      cadenceFirst: 'First message only',
-      cadenceEvery: 'Every message',
-      cadenceOff: 'Off',
-      forceLanguage: 'Force response language',
-      languageAuto: 'Auto',
-      languageZh: 'Simplified Chinese',
-      languageEn: 'English',
-      backendUnavailable: 'Prompt settings backend is unavailable. Reload the extension and try again.',
-      loadFailed: 'Failed to load prompt settings: {error}',
-      saveFailed: 'Failed to save prompt settings: {error}',
-    },
-    voice: {
-      title: 'Voice',
-      input: 'Voice input',
-      inputDescription: 'Use browser speech recognition in sidepanel chat',
-      inputUnsupported: 'Speech recognition is not available in this browser context',
-      readAloud: 'Read responses aloud',
-      readAloudDescription: 'Use browser speech synthesis for sidepanel chat responses',
-      readAloudUnsupported: 'Speech synthesis is not available in this browser context',
-      rate: 'Rate',
-      pitch: 'Pitch',
-    },
-    capabilitiesPage: {
-      navLabel: 'Capabilities navigation',
-      tabs: {
-        skill: 'Skill',
-        mcp: 'MCP',
-        tools: 'Tools',
-        browser: 'Browser',
-        preset: 'Presets',
-        automation: 'Automation',
-      },
-    },
-    browserControlPage: {
-      title: 'Browser Control',
-      description: 'Let DeepSeek++ control a selected Chrome or Edge tab through the browser debugger and Accessibility Tree.',
-      enableTitle: 'Enable browser tools',
-      enableDescription: 'When enabled, browser_* tools are injected into new conversations. The model receives text snapshots, not screenshots.',
-      unsupported: 'Browser Control requires Chromium debugger and tabs APIs. It is not available on this platform.',
-      detach: 'Detach',
-      detaching: 'Detaching...',
-      targetsTitle: 'Target tab',
-      noTargets: 'No browser tabs found.',
-      untitled: 'Untitled',
-      noUrl: 'No URL',
-      group: 'Group: {name}',
-      selected: 'Selected',
-      snapshotTitle: 'Snapshot budget',
-      includeSnapshot: 'Return snapshot after actions',
-      includeSnapshotDescription: 'After click, fill, navigation, and keyboard actions, return a fresh Accessibility Tree snapshot.',
-      maxNodes: 'Max AX nodes',
-      maxBytes: 'Max snapshot text bytes',
-      status: {
-        enabled: 'Enabled',
-        attached: 'Attached',
-        target: 'Target',
-      },
-      messages: {
-        enabled: 'Browser tools enabled. New conversations can receive browser_* tool schemas.',
-        disabled: 'Browser tools disabled. Browser tool schemas will not be injected into new conversations.',
-        detached: 'Detached from the current target.',
-        targetFailed: 'Failed to select target tab.',
-        targetSelected: 'Selected tab {id}.',
-      },
-    },
-    mcpPage: {
-      title: 'MCP',
-      description: 'Connect local or remote MCP services so DeepSeek++ can discover tools, control auto execution, and return results.',
-      summary: '{servers} servers, {enabled} enabled, {tools} auto tools',
-      loading: 'Loading MCP config',
-      empty: 'No MCP servers',
-      emptyHint: 'Connect a remote MCP service, or one-click install the Shell / multimodal native host preset.',
-      emptyCreateAction: 'Add MCP server',
-      emptyInstallShell: 'Install Shell preset',
-      shell: 'Shell',
-      multimodal: 'Multimodal',
-      addServer: 'Add',
-      enabled: 'Enabled',
-      selected: 'Selected',
-      auto: 'Auto',
-      disabled: 'Disabled',
-      success: 'Success',
-      failure: 'Failed',
-      localHost: 'local host',
-      messages: {
-        loadFailed: 'Failed to load MCP config',
-        shellExistsSelected: 'Shell MCP already exists. Existing config selected.',
-        shellCreateFailed: 'Failed to create Shell MCP preset',
-        shellCreated: 'Shell MCP preset created. Run the install command below, then restart the browser.',
-        multimodalExistsSelected: 'Multimodal MCP already exists. Existing config selected.',
-        multimodalCreateFailed: 'Failed to create Multimodal MCP preset',
-        multimodalCreated: 'Multimodal MCP preset created. Run the install command below, then configure Multimodal API on Settings.',
-        saveFailed: 'Failed to save MCP server',
-        deleteConfirm: 'Delete MCP server "{name}"?',
-        permissionGranted: 'Granted {origin}',
-        permissionDenied: 'Permission denied',
-        permissionRequired: 'Permission required for {origin}',
-        connectionSuccess: 'Connected. {tools} tools, {latency}',
-        connectionFailed: 'Connection failed',
-        nativeMessagingUnsupported: 'Native Messaging is not available on this platform.',
-      },
-      transportHints: {
-        streamableHttp: 'Recommended. Compatible with newer MCP HTTP services.',
-        http: 'JSON-RPC over HTTP POST.',
-        sse: 'Legacy MCP SSE transport.',
-        stdioBridge: 'Connect to a local stdio bridge already exposed as MCP HTTP; process launch stays inside the bridge service.',
-        nativeMessaging: 'Access local capabilities through the browser Native Messaging Host.',
-      },
-      form: {
-        createTitle: 'Add MCP server',
-        editTitle: 'Edit MCP server',
-        basic: 'Basic',
-        stdioSection: 'Bridge process info',
-        advanced: 'Advanced',
-        advancedHint: 'Headers, secrets, timeouts, limits and execution policy',
-        name: 'Name',
-        transport: 'Transport',
-        serviceUrl: 'Service URL',
-        bridgeEndpointUrl: 'Bridge endpoint URL',
-        bridgeEndpointHint: 'Use the HTTP MCP endpoint exposed by a stdio bridge service, not a local executable path. For local shell/stdio capabilities, prefer the Shell Native Host preset.',
-        command: 'Command',
-        args: 'Arguments',
-        cwd: 'Working directory',
-        env: 'Environment variables',
-        connectMs: 'Connect ms',
-        requestMs: 'Request ms',
-        discoveryMs: 'Discovery ms',
-        resultBytes: 'Result bytes',
-        toolLimit: 'Tool limit',
-        defaultExecution: 'Default execution',
-        allowInject: 'Allow injection',
-        modeAuto: 'Auto execute',
-        modeManual: 'Manual strategy',
-        modeDisabled: 'Disabled',
-        save: 'Save',
-      },
-      headers: {
-        headerName: 'Header',
-        headerValue: 'Value',
-        secretValue: 'Secret value',
-        secretHeaderName: 'Header name',
-      },
-      validation: {
-        nameRequired: 'Name is required',
-        positiveInteger: '{label} must be a positive integer',
-        nativeHostRequired: 'Native Host is required',
-        nativeHostInvalid: 'Native Host can only contain letters, numbers, dots, underscores, and hyphens',
-        serviceUrlRequired: 'Service URL is required',
-        bridgeEndpointRequired: 'Stdio Bridge requires a bridge endpoint URL',
-        serviceUrlUnsupported: 'Service URL only supports http/https',
-        serviceUrlInvalid: 'Service URL is invalid',
-        stdioCommandRequired: 'Stdio Bridge command is required',
-        headerInvalidName: 'Invalid Header name: {name}',
-        emptyValue: '(empty)',
-        headerInvalidValue: 'Header value cannot contain line breaks: {name}',
-        headerSecretRequired: 'Header Secret requires a valid Header name',
-        envInvalid: 'Invalid environment variable format: {line}',
-      },
-      row: {
-        autoTools: '{active}/{total} auto',
-        testing: 'Testing',
-        refreshTools: 'Refresh tools',
-      },
-      detail: {
-        grant: 'Grant',
-        status: 'Status',
-        latency: 'Latency',
-        lastConnected: 'Last connected',
-        transport: 'Transport',
-        executionPolicy: 'Auto execution policy',
-        injectionSummary: '{count} tools are currently injected. Disabled or manual strategy will not enter the DeepSeek Prompt.',
-        promptExposure: 'Prompt exposure',
-        promptExposureDescription: 'Controls how this server\'s tool schemas enter new prompts. It never changes server permissions or execution policy.',
-        exposureDirect: 'Direct: show all tools',
-        exposureAdaptive: 'Adaptive: show relevant tools',
-        exposureOnDemand: 'On demand: catalog only',
-        exposureAdaptiveHint: '{count} pinned tools are preferred when selecting direct schemas.',
-        exposureOnDemandHint: 'Tool schemas stay out of the prompt until the model discovers and describes a capability.',
-        pin: 'Pin',
-        unpin: 'Unpin',
-        adaptiveBudget: 'Adaptive prompt budget',
-        adaptiveBudgetDescription: 'Shared limit for directly shown schemas from all adaptive MCP servers.',
-        adaptiveMaxTools: 'Direct schemas',
-        adaptiveMaxPromptBytes: 'Schema bytes',
-        adaptiveBudgetApply: 'Apply',
-        adaptiveBudgetInvalid: 'Use 1–64 schemas and 2,000–256,000 bytes.',
-        discoveredTools: 'Discovered tools',
-        noTools: 'No tools discovered yet',
-        recentCalls: 'Recent calls',
-        noHistory: 'No call history',
-        schemaNone: 'Params: none',
-        schemaSummary: 'Params: {props}{required}',
-        schemaRequired: '; required {required}',
-        toggleTools: 'Tools',
-        toggleHistory: 'Call history',
-        hintExpand: 'Show install guide',
-        hintCollapse: 'Hide install guide',
-      },
-      status: {
-        ready: 'ready',
-        error: 'error',
-        disabled: 'disabled',
-        unknown: 'unknown',
-      },
-      endpoint: {
-        nativeMessaging: 'Native Messaging',
-        bridgeUrl: 'Bridge URL',
-        command: 'command',
-      },
-      shellSetup: {
-        localIntro: 'Open a terminal in the project root and run:',
-        publishedIntro: 'Open a terminal and run this command once:',
-        fallbackIntro: 'If you are using the published extension instead of a local source build, run:',
-        detectedExtensionId: 'Detected the {browser} extension ID automatically. Node.js/npm must be installed locally.',
-        firefoxFixedId: 'Firefox uses a fixed extension ID, so no extra extension id is required. Node.js/npm must be installed locally.',
-        installNote: 'The command installs or updates Shell Native Host. OfficeCLI is skipped by default; remove --skip-officecli if you need OfficeCLI.',
-        enableAndTest: 'After installation, enable this service with the switch above, then click "Test" to verify the connection.',
-        restartAndTest: 'After installation, restart the browser and click "Test" here to verify the connection.',
-        forbidden: 'Native Host is installed, but the current extension ID is not authorized. Run the install command again, then restart the browser.',
-        notFound: 'Native Host was not found. Run the install command below and make sure Node.js/npm is installed.',
-        unavailable: 'The current browser does not support Native Messaging. Use Chrome, Edge, or Firefox.',
-        cannotConnect: 'Cannot connect to Native Host. Confirm that the install script ran and the browser was restarted.',
-        ready: 'Connected, {count} tools discovered.',
-        disabled: 'The service has been created but is not enabled. Install Native Host, then enable and test it.',
-        installFirst: 'Install Native Host first, then click "Test" to verify the connection.',
-      },
-      multimodalSetup: {
-        localIntro: 'Open a terminal in the deepseek-pp project root and run:',
-        publishedIntro: 'Open a terminal and run this command once:',
-        fallbackIntro: 'If you are using the published extension instead of a local source build, run:',
-        settingsNote: 'Configure API keys, models, and request URLs under Settings > Multimodal API. Images use OpenAI; videos use Gemini.',
-        enableAndTest: 'After installation, configure Multimodal API on Settings, enable this service with the switch above, then click "Test".',
-        forbidden: 'Multimodal Native Host is installed, but the current extension ID is not authorized. Run the install command again, then restart the browser.',
-        notFound: 'Multimodal Native Host was not found. Run the install command below and make sure Node.js/npm is installed.',
-        cannotConnect: 'Cannot connect to Multimodal Native Host. Confirm that the install script ran and the browser was restarted.',
-        ready: 'Connected, {count} multimodal tools discovered.',
-        disabled: 'The service has been created but is not enabled. Install Native Host, configure Multimodal API, then enable and test it.',
-        installFirst: 'Install Multimodal Native Host first, configure Multimodal API on Settings, then click "Test".',
-      },
-    },
-    chatPage: {
-      authRequired: 'Configure a DeepSeek API Key, or sign in at chat.deepseek.com first',
-      authHint: 'Without an API Key, sidepanel chat depends on the DeepSeek webpage login session',
-      title: 'Chat',
-      description: 'Start tasks from the side panel with memory, presets, Skills, and enabled tools available.',
-      apiDescription: 'Official API mode with model and thinking controls.',
-      webDescription: 'Web session mode with memory, presets, Skills, and enabled tools.',
-      apiProvider: 'API',
-      webProvider: 'Web',
-      newSessionTitle: 'New session',
-      newSession: 'New',
-      newSessionConfirm: 'The current conversation will be cleared. Start a new session?',
-      empty: 'Type a message to start chatting',
-      emptyHelp: 'You can also send selected webpage text from the context menu or insert a saved prompt.',
-      inputPlaceholder: 'Message DeepSeek++',
-      send: 'Send',
-      sendFailed: 'Send failed',
-      uploadImage: 'Upload image',
-      imageAttachments: 'Image attachments',
-      imageUploading: 'Uploading...',
-      imageReady: 'Attached',
-      imageUploadFailed: 'Upload failed',
-      imageUploadMissingId: 'DeepSeek upload completed without returning a file ID.',
-      imageUploadMax: 'Upload up to {count} images per turn.',
-      imageOnly: 'Only image files are supported.',
-      imageEmpty: 'The image file is empty.',
-      imageTooLarge: 'Image exceeds the {limit} limit.',
-      removeImageAttachment: 'Remove image {name}',
-      atPanelTitle: 'Reference files',
-      atPanelHint: 'Type @ then search by file name or path; select an image to reference it',
-      atPanelFileCount: '{count} files',
-      atPanelNoDirectory: 'No authorized project folder: Settings → Project Folders',
-      atPanelNoMatch: 'No matching files',
-      atPanelEmptyDirectory: 'No referenceable files in the folder',
-      atPanelTextPending: 'Text file references come in a later version',
-      atPanelOtherUnsupported: 'This file type cannot be referenced yet',
-      atPanelSelectFile: 'Select {name}',
-      atPanelDeselectFile: 'Deselect {name}',
-      atPanelClose: 'Close reference panel',
-      modelLabel: 'Model',
-      modelFlash: 'Flash',
-      modelPro: 'Pro',
-      thinkingLabel: 'Thinking',
-      thinkingOff: 'Instant',
-      thinkingOn: 'DeepThink',
-      effortLabel: 'Thinking effort',
-      effortHigh: 'Standard',
-      effortMax: 'Max',
-      reasoningTitle: 'Reasoning',
-      reasoningActive: 'Thinking',
-      read: 'Read',
-      readLatest: 'Read latest assistant response',
-      voiceInput: 'Voice input',
-      stopListening: 'Stop listening',
-      mic: 'Mic',
-      stop: 'Stop',
-    },
-    toolsPage: {
-      diagnosticsDefaultQuery: 'DeepSeek latest news',
-      diagnosticsRunning: 'Diagnosing...',
-      diagnosticsRun: 'Diagnose',
-      bytes: '{count} bytes',
-      errorPrefix: 'Error: {error}',
-      toolTitle: 'Tool switches',
-      toolDescription: 'When disabled, the tool is not injected into conversations and AI cannot call it',
-      webSearchName: 'Web search (web_search)',
-      webSearchDescription: 'Search Bing keywords and return titles, URLs, and snippets',
-      webFetchName: 'Fetch web page (web_fetch)',
-      webFetchDescription: 'Download a URL and extract visible text content',
-      pythonTitle: 'Python interpreter (python_exec)',
-      pythonStatusNoShell: 'Shell Native Host has not been created',
-      pythonStatusUnsupported: 'Shell Native Host is not available on this platform',
-      pythonStatusNoCache: 'Tools have not been refreshed',
-      pythonStatusEnabled: 'Enabled',
-      pythonStatusDiscovered: 'Discovered, not enabled',
-      pythonStatusMissing: 'python_exec not found',
-      pythonDescription: 'Run short code with local Python for quick checks, complex calculations, and small data tasks.',
-      pythonCreate: 'Create Shell',
-      pythonCreating: 'Creating',
-      pythonRefresh: 'Refresh tools',
-      pythonRefreshing: 'Refreshing',
-      pythonStatusAvailable: 'python_status available',
-      pythonMissingDetail: 'The current Shell Native Host did not return python_exec. Reinstall Shell Native Host, restart the browser, then refresh again.',
-      shellExists: 'Shell MCP already exists. Refresh tools or reinstall Native Host.',
-      shellCreated: 'Shell MCP was created. Install or reinstall Shell Native Host, then refresh tools.',
-      pythonFound: 'Python interpreter tool found.',
-      pythonMissingAfterRefresh: 'python_exec was not found. Reinstall Shell Native Host, restart the browser, then refresh tools.',
-      pythonMissingBeforeToggle: 'python_exec was not found. Refresh tools first, and reinstall Shell Native Host if needed.',
-      pythonEnabled: 'Python interpreter enabled.',
-      pythonDisabled: 'Python interpreter disabled.',
-      disabledNotice: 'After a tool is disabled, new conversations no longer include its call format. Existing conversations are not affected.',
-      diagnosticTitle: 'Search diagnostics',
-      diagnosticDescription: 'Test whether search works directly, without the AI chat path',
-      permissionTitle: 'web_fetch permission',
-      permissionDescription: 'Fetching pages requires permission for the target site. Enter a URL here and grant access.',
-      grantPermission: 'Grant',
-      permissionGranted: 'Permission granted. This site can be accessed.',
-      permissionDenied: 'Permission denied. Try again or add access manually in chrome://extensions.',
-      permissionInvalidUrl: 'Invalid URL. Enter a full URL such as https://example.com.',
-      allSitesRequesting: 'Requesting...',
-      allSitesGranted: 'All sites granted',
-      allSitesGrant: 'Grant all sites',
-      allSitesHelp: 'Grant the extension access to all sites, so web_fetch can fetch pages without another prompt',
-    },
-    automationPage: {
-      title: 'Automation',
-      description: 'Run fixed tasks in dedicated DeepSeek conversations, either manually or on a Cron / RRULE schedule.',
-      summary: '{total} tasks, {active} active',
-      create: 'New',
-      created: 'Automation created',
-      loadFailed: 'Failed to load automation data',
-      empty: 'No automations',
-      emptyHelp: 'Create a task to track information, review content, or repeat a fixed prompt on schedule.',
-      namePromptRequired: 'Name and prompt are required',
-      expressionRequired: 'Enter a schedule expression',
-      runFailed: 'Run failed',
-      deleteConfirm: 'Delete automation "{name}"?',
-      form: {
-        name: 'Name',
-        namePlaceholder: 'Task name',
-        model: 'Model',
-        defaultModel: 'Default',
-        promptPlaceholder: 'Enter the content to send to DeepSeek on schedule',
-        trigger: 'Trigger',
-        manual: 'Manual',
-        expression: 'Expression',
-        timezone: 'Timezone',
-        search: 'Web search',
-        thinking: 'Deep thinking',
-        create: 'Create',
-      },
-      status: {
-        active: 'Active',
-        paused: 'Paused',
-        queued: 'Queued',
-        running: 'Running',
-        succeeded: 'Succeeded',
-        failed: 'Failed',
-        timeout: 'Timed out',
-        cancelled: 'Cancelled',
-        skipped: 'Skipped',
-      },
-      meta: {
-        next: 'Next',
-        previous: 'Previous',
-        session: 'Session',
-        recent: 'Recent',
-        notCreated: 'Not created',
-        none: 'None',
-      },
-      actions: {
-        openSession: 'Open session',
-        runNow: 'Run now',
-      },
-      attempt: '{count}x',
-    },
-  },
-  content: {
-    conversation: {
-      untitled: 'Untitled conversation',
-    },
-    export: {
-      buttonIdle: 'Export current conversation',
-      buttonRunning: 'Exporting current conversation',
-      emptyConversation: 'This page has no exportable conversation yet.',
-      formatDialogLabel: 'Choose export formats',
-      formatTitle: 'Export formats',
-      scopeLabel: 'Content scope',
-      scopeFull: 'Everything (tool calls, continuations, reasoning, injected context)',
-      scopeInputOutput: 'Inputs + final output only',
-      formatHtml: 'HTML',
-      formatMarkdown: 'Markdown',
-      formatPdf: 'PDF',
-      formatImageManifest: 'Image manifest',
-      submit: 'Export',
-      progress: 'Exporting current conversation...',
-      failed: 'Export failed',
-      cancelled: 'Export cancelled',
-      partialSuccess: 'Export completed, but some sessions could not be read. Check Export Warnings in the exported file.',
-      success: 'Current conversation exported.',
-    },
-    historyOrganizer: {
-      enhancedSearchTitle: 'DeepSeek++ search enhancements',
-      tagFilterLabel: 'Filter results by tag',
-      tagPlaceholder: 'Enter a tag',
-      currentTagsLabel: 'Tag the current chat',
-      currentTagsPlaceholder: 'Comma-separated, e.g. hk stocks, writing',
-      emptySearchStatus: 'DeepSeek++: waiting for official search results',
-      visibleStatus: 'DeepSeek++: {visible}/{total} visible',
-      loadError: 'DeepSeek++: failed to load history tags: {message}',
-      saveError: 'DeepSeek++: failed to save history tags: {message}',
-    },
-    projectSidebar: {
-      title: 'Projects',
-      empty: 'No projects yet',
-      expandProject: 'Expand project: {name}',
-      collapseProject: 'Collapse project: {name}',
-      showMore: 'Show more',
-      showLess: 'Show less',
-      moveCurrentToProject: 'Move current conversation to {name}',
-      removeCurrentFromProject: 'Remove current conversation from {name}',
-      joinProject: 'Add to project',
-      joinProjectNamed: 'Add to project: {name}',
-      moveToProjectNamed: 'Move to project: {name}',
-      currentProjectNamed: 'In project: {name}',
-      removeFromProjectNamed: 'Remove from project: {name}',
-      conversationActions: 'Conversation actions',
-      newConversationInProject: 'Start a new conversation in {name}',
-      useNextConversation: 'Use {name} for the next new conversation',
-      cancelNextConversation: 'Cancel using {name} for the next new conversation',
-      pendingNextConversation: 'The next new conversation will use this project',
-      operationFailed: 'Project action failed: {message}',
-      age: {
-        justNow: 'Just now',
-        minutesAgo: '{count}m ago',
-        hoursAgo: '{count}h ago',
-        daysAgo: '{count}d ago',
-      },
-    },
-    uxPolish: {
-      downloadCode: 'Download',
-      downloadMessageMarkdownButton: 'MD',
-      downloadMessageMarkdownTitle: 'Download message as Markdown',
-      copyMessageButton: 'Copy',
-      copyMessageTitle: 'Copy full message output',
-    },
-    toolBlock: {
-      title: 'Executed tools ({count})',
-      pythonInterpreter: 'Python interpreter',
-      summaries: {
-        saved: 'Saved',
-        updated: 'Updated',
-        deleted: 'Deleted',
-        searched: 'Searched',
-        fetched: 'Fetched',
-        running: 'Running',
-        executed: 'Executed',
-        failed: 'Execution failed',
-        messageFailed: 'Tool message failed',
-        backgroundFailed: 'Background tool execution failed',
-      },
-      missingResultDetail: 'Background did not return a tool result. Refresh the current DeepSeek page and try again.',
-      missingSandboxResultDetail: 'The sandbox tool did not return an execution result. Reload the DeepSeek++ extension and refresh the current DeepSeek page, then try again; it does not depend on MCP or Shell Local.',
-      missingMcpResultDetail: 'Background did not return an MCP tool result. Refresh the current DeepSeek page and try again. If it still fails, retest Shell Local on the MCP page.',
-      invalidResultDetail: 'Background returned an invalid tool result: {preview}',
-    },
-    agent: {
-      stop: 'Stop',
-      stopped: 'Stopped',
-      step: 'Step {index}',
-      streaming: 'streaming...',
-      complete: 'Complete',
-      completeWithTools: 'Complete ({count} tools)',
-      executingTools: 'Executing tools',
-      budgetReached: 'DeepSeek++ paused after {count} automated tool-continuation rounds to avoid presenting incomplete work as a final answer. Send "continue" in this conversation to resume the remaining work.',
-      error: 'Execution error',
-      footerComplete: 'Agent complete ({steps} steps, {tools} tool calls)',
-      process: 'Process',
-      tools: 'Tool calls',
-      results: 'Results',
-      running: 'Agent running · step {step} · {tools} tool calls',
-      footerError: 'Agent error ({steps} steps, {tools} tool calls)',
-      autoSaved: 'Long output was auto-saved as an artifact: ',
-      autoSavedDownload: 'Download',
-      concurrencyGuard: 'An agent is still running. A duplicate task from this message was skipped. To interrupt, click Stop in the agent panel.',
-    },
-    permission: {
-      webFetch: 'DeepSeek++ needs permission to access {origin} so it can fetch that page',
-      deny: 'Deny',
-      grant: 'Allow',
-      requesting: 'Requesting...',
-    },
-    skillPopup: {
-      hint: '↑↓ Navigate · Enter Select · Esc Close',
-    },
-    tokenSpeed: {
-      title: 'Accumulated tokens: {tokens}; token output speed: {speed}{idle}. Token source: {tokenSource}; speed source: {speedSource}',
-      idleSuffix: ' (idle)',
-      sourceServer: 'server',
-      sourceEstimated: 'local estimate',
-    },
-    multimodalMedia: {
-      buttonTitle: 'Add DeepSeek++ multimodal images or videos',
-      removeTitle: 'Remove {name}',
-      pastedFileName: 'Pasted media',
-      added: 'Added {count} multimodal attachment(s)',
-      analyzing: 'Analyzing {count} multimodal attachment(s)...',
-      analyzed: 'Multimodal analysis complete ({count} item(s))',
-      failed: 'Multimodal analysis failed: {message}',
-      tooMany: 'Attach at most {count} multimodal files per turn',
-      unsupported: '{name} is not a supported image or video',
-      imageTooLarge: '{name} exceeds the image limit of {limit}',
-      videoTooLarge: '{name} exceeds the inline video limit of {limit}. Use a public video URL or wait for local-path picker support.',
-      invalidRequest: 'The DeepSeek request body is invalid, so multimodal media cannot be attached.',
-      emptyPrompt: 'Type a message before sending multimodal media.',
-    },
-    extensionReloaded: 'The extension was reloaded. Refresh the current DeepSeek page and try again.',
-  },
-  background: {
-    contextMenus: {
-      sendToChat: 'Send to chat',
-    },
-    auth: {
-      missingDeepSeek: 'Sign in at chat.deepseek.com first, or refresh the DeepSeek page and try again.',
-    },
-    export: {
-      generating: 'Generating export files',
-      cancelled: 'Export cancelled',
-    },
-    sync: {
-      missingWebDav: 'WebDAV is not configured',
-      missingSync: 'Cloud sync is not configured',
-      configChanged: 'Sync settings changed in another extension window. Review the current target and try again.',
-      configCommitIndeterminate: 'Could not confirm whether the sync settings were saved. The current settings will be reloaded.',
-      missingRemoteFile: 'Remote file {file} is missing. Download stopped to avoid overwriting local data.',
-      authorizationNotRequired: 'The current sync method does not require authorization',
-      identityUnavailable: 'OAuth authorization is not available in this browser context',
-      oauthCancelled: 'Authorization was cancelled',
-      oauthProviderError: 'Authorization failed: {error}',
-      oauthMissingCode: 'Authorization failed: no authorization code was returned',
-      oauthRefreshFailed: 'Failed to refresh access token (HTTP {status}){detail}',
-      oauthTokenExchangeFailed: 'Failed to get token (HTTP {status}){detail}',
-      gdriveMissingRefreshToken: 'Google did not return a refresh_token. Revoke access and authorize again.',
-      gdriveMissingAuthorization: 'Google Drive is not authorized. Click Authorize first.',
-      gdriveAuthorizationExpired: 'Google authorization expired. Authorize again.',
-      gdriveConnectFailed: 'Failed to connect to Google Drive (HTTP {status})',
-      gdriveQueryFailed: 'Failed to query {name} (HTTP {status})',
-      gdriveDownloadFailed: 'Failed to download {key} from Google Drive (HTTP {status})',
-      gdriveUploadFailed: 'Failed to upload {name} to Google Drive (HTTP {status})',
-      onedriveMissingRefreshToken: 'Microsoft did not return a refresh_token. Revoke access and authorize again.',
-      onedriveMissingAuthorization: 'OneDrive is not authorized. Click Authorize first.',
-      onedriveAuthorizationExpired: 'Microsoft authorization expired. Authorize again.',
-      onedriveConnectFailed: 'Failed to connect to OneDrive (HTTP {status})',
-      onedriveDownloadFailed: 'Failed to download {key} from OneDrive (HTTP {status})',
-      onedriveUploadFailed: 'Failed to upload {key} to OneDrive (HTTP {status})',
-    },
-    chat: {
-      continueWithToolResults: '[TOOL_RESULTS]\n{toolResults}\n[/TOOL_RESULTS]\n\nContinue answering based on the tool results above.',
-      maxToolSteps: '(Maximum tool-call steps reached; conversation ended)',
-      interrupted: 'The assistant was interrupted. Please send your message again.',
-    },
-  },
-  tool: {
-    runtime: {
-      invalidFormat: 'Invalid tool format',
-      unknownTool: 'Unknown tool',
-      authorizationRejected: 'Tool authorization rejected',
-      authorizationMissingHint: 'The authorization is closed or was never confirmed. Re-run the action and allow this tool in the authorization prompt.',
-      authorizationStaleHint: 'The authorization expired. Re-run the action to refresh it.',
-      authorizationSessionMismatchHint: 'The authorization belongs to another chat session. Re-run the action in this conversation.',
-      authorizationDisabledHint: 'This tool is currently disabled. Enable it in tool settings and retry.',
-      authorizationIdentityMissingHint: 'Identity information required for execution is missing. Re-run the action.',
-      authorizationGenericHint: 'To continue, re-run the action and complete the authorization confirmation.',
-      sessionBindingUnavailable: 'The new conversation did not receive a browser-owned session route in time. Please wait for the chat URL to update and try again.',
-      incomplete: 'The response ended before this tool call completed.',
-    },
-    memory: {
-      providerName: 'DeepSeek++ Memory',
-      saveTitle: 'Save memory',
-      saveDescription: 'Save a new long-term memory',
-      typeDescription: 'Memory type: user=identity, role, or preference; feedback=behavior correction; topic=discussion notes; reference=external resource link',
-      nameDescription: 'Short title',
-      contentDescription: 'Content to save',
-      tagsDescription: 'Tag list',
-      updateTitle: 'Update memory',
-      updateDescription: 'Update an existing memory',
-      idDescription: 'Memory ID',
-      updatedNameDescription: 'Updated title',
-      updatedContentDescription: 'Updated content',
-      deleteTitle: 'Delete memory',
-      deleteDescription: 'Delete memory',
-      unsupported: 'Unsupported memory tool',
-      saveFailed: 'Save failed',
-      saveMissingConfirmation: 'No save confirmation received',
-      saved: 'Saved',
-      invalidPayload: 'Invalid memory payload',
-      invalidType: 'type must be user, feedback, topic, or reference',
-      invalidName: 'name must be a non-empty string',
-      invalidContent: 'content must be a non-empty string',
-      invalidTags: 'tags must be an array of strings',
-      invalidId: 'Invalid ID',
-      notFound: 'Memory not found',
-      notFoundDetail: 'ID {id} does not exist',
-      updated: 'Updated',
-      deleted: 'Deleted',
-    },
-    web: {
-      providerName: 'DeepSeek++ Web Search',
-      searchTitle: 'Search the web',
-      searchDescription: 'Search the web and return matching page titles, URLs, and snippets',
-      queryDescription: 'Search query keywords',
-      topKDescription: 'Number of results to return, default 5',
-      fetchTitle: 'Fetch web page',
-      fetchDescription: 'Download the specified URL and return visible text after removing navigation, scripts, and styles',
-      urlDescription: 'Full URL to fetch (http:// or https://)',
-      unsupported: 'Unsupported web tool',
-      emptyQuery: 'Search query cannot be empty',
-      searchComplete: 'Search completed with {count} results',
-      searchNoResults: 'No search results',
-      searchFailed: 'Search failed',
-      permissionDenied: 'The extension does not have permission to access Bing. Remove and reload the dist/chrome-mv3 extension directory, or confirm cn.bing.com is listed under site access in chrome://extensions -> DeepSeek++ details.',
-      noParseableResults: 'No parseable search results found: {error}',
-      searchFailedDetail: 'Search failed: {error}',
-      emptyUrl: 'URL cannot be empty',
-      invalidUrl: 'Invalid URL',
-      invalidUrlDetail: 'Could not parse URL: {url}',
-      contentType: 'Page type: {contentType}',
-      contentTypeDetail: 'Content-Type: {contentType}\nURL: {url}\n(Content is not text and cannot be displayed)',
-      truncated: '\n\n... [Content truncated, {count} characters total]',
-      fetchComplete: 'Fetched {url}',
-      fetchTruncatedDetail: 'Page length {length} characters, truncated to {maxLength} characters',
-      fetchLengthDetail: 'Page length {length} characters',
-      fetchFailed: 'Failed to fetch page',
-      missingHostPermission: 'Cannot access {url}; host permission is missing.',
-    },
-    artifact: {
-      createTitle: 'Create downloadable file',
-      createDescription: 'Create a downloadable UTF-8 text artifact or runnable code artifact. Use it for HTML previews and JavaScript, TypeScript, or Python runners. Never embed binary or base64 Office, PDF, image, or archive data; keep files produced by Shell or OfficeCLI at their local path and report that path.',
-      bundleTitle: 'Create downloadable project bundle',
-      bundleDescription: 'Create a ZIP bundle from multiple generated text files. Use this for multi-file projects and scaffolds.',
-      filenameDescription: 'Safe relative file name, for example report.md or src/index.ts',
-      bundleFilenameDescription: 'ZIP filename, for example project.zip',
-      contentDescription: 'Complete UTF-8 text content; never binary or base64 file data',
-      mimeTypeDescription: 'MIME type, for example text/markdown or application/json',
-      previewModeDescription: 'Artifact view mode: auto infers from content, html previews a page, code shows a run button, none keeps it download-only.',
-      languageDescription: 'Artifact language: html, javascript, typescript, python, or text. Inferred from filename and MIME type when omitted.',
-      filesDescription: 'Files to include in the bundle. Each item has path, content, and optional mimeType.',
-      failed: 'Artifact failed',
-      unsupported: 'Unsupported artifact tool',
-      fileReady: 'File ready',
-      bundleReady: 'Bundle ready',
-      result: {
-        files: '{count} files',
-        download: 'Download',
-        downloading: 'Downloading...',
-        downloaded: 'Downloaded',
-        downloadFailed: 'Download failed',
-        preview: 'Preview',
-        previewFailed: 'Preview failed',
-        previewAriaLabel: 'Artifact preview',
-        closePreview: 'Close preview',
-        run: 'Run',
-        running: 'Running...',
-        runAgain: 'Run again',
-        failed: 'Failed',
-        done: 'Done',
-        artifactNotFound: 'Artifact not found',
-        codeExecuted: 'Code executed',
-        codeFailed: 'Code failed',
-        runnerUnavailable: 'Code runner did not return a result',
-      },
-    },
-    sandbox: {
-      runTitle: 'Run code in browser sandbox',
-      runDescription: 'Execute JavaScript, TypeScript, Python, or HTML in an isolated extension sandbox, not in the DeepSeek page.',
-      languageDescription: 'Runtime language: javascript, typescript, python, or html',
-      codeDescription: 'Code to run in the isolated browser sandbox',
-      inputDescription: 'Optional input string available to the code',
-      timeoutDescription: 'Timeout in milliseconds, default 5000',
-      executed: 'Sandbox executed',
-      failed: 'Sandbox failed',
-      invalidRequest: 'Invalid sandbox request',
-      offscreenUnavailable: 'Browser sandbox unavailable',
-      offscreenUnavailableDetail: 'This browser does not expose the MV3 offscreen API needed for isolated JavaScript and TypeScript sandbox execution.',
-      runtimeUnavailableDetail: 'The browser sandbox executor is only available from the extension background runtime.',
-      pythonUnavailable: 'Python sandbox unavailable',
-      pythonUnavailableDetail: 'The bundled Pyodide runtime could not start in the isolated browser sandbox.',
-    },
-    skillCreator: {
-      title: 'Draft a Skill',
-      description: 'Create a custom Skill draft for user review. The draft is not saved until the user approves it.',
-      nameDescription: 'Lowercase Skill trigger name',
-      descriptionDescription: 'When this Skill should be used',
-      instructionsDescription: 'Markdown instructions for the Skill',
-      memoryDescription: 'Whether the Skill should receive memory context',
-      draftReady: 'Skill draft ready',
-      invalidDraft: 'Invalid Skill draft',
-      result: {
-        saveSkill: 'Save Skill',
-        saving: 'Saving...',
-        saved: 'Saved',
-        saveFailed: 'Save failed',
-        memoryOn: 'memory on',
-        memoryOff: 'memory off',
-      },
-    },
-    memoryImport: {
-      title: 'Preview memory import',
-      description: 'Convert another AI workflow export or notes into memory drafts. Nothing is saved until the user imports the preview.',
-      contentDescription: 'JSON, Markdown, or plain-text memory source',
-      typeDescription: 'Default memory type: user, feedback, topic, or reference',
-      tagsDescription: 'Tags to add to imported memories',
-      previewReady: 'Memory import preview ready ({count})',
-      previewDetail: '{duplicates} duplicates skipped, {rejected} rejected',
-      failed: 'Memory import failed',
-      result: {
-        metaSummary: '{count} memories · {duplicates} duplicates',
-        importMemories: 'Import memories',
-        importing: 'Importing...',
-        imported: 'Imported {count}',
-        importFailed: 'Import failed',
-      },
-    },
-    mcpCapability: {
-      providerName: 'MCP Capability Catalog',
-      discoverTitle: 'Discover MCP capabilities',
-      discoverDescription: 'Search available MCP tools that were not directly shown. Discover a capability before describing or invoking it.',
-      describeTitle: 'Describe an MCP capability',
-      describeDescription: 'Use a capability from discovery to inspect one MCP tool and its full input schema without executing it.',
-      invokeTitle: 'Invoke a discovered MCP capability',
-      invokeDescription: 'Invoke only a capability returned by mcp_discover. A raw tool name cannot bypass authorization.',
-      discoverQueryDescription: 'Task or capability to search for.',
-      discoverLimitDescription: 'Maximum number of candidates.',
-      capabilityDescription: 'Capability returned by mcp_discover.',
-      pointerDescription: 'Optional JSON Pointer to one schema field.',
-      argumentsDescription: 'Arguments for the resolved MCP tool.',
-    },
-    browser: {
-      providerName: 'DeepSeek++ Browser Control',
-      navigateTitle: 'Navigate',
-      navigateDescription: 'Open a URL in a new tab by default; set newTab=false only to replace the current controlled tab.',
-      goBackTitle: 'Go back',
-      goBackDescription: 'Move the controlled tab back in browser history.',
-      goForwardTitle: 'Go forward',
-      goForwardDescription: 'Move the controlled tab forward in browser history.',
-      refreshTitle: 'Refresh',
-      refreshDescription: 'Reload the controlled tab.',
-      listTabsTitle: 'List tabs',
-      listTabsDescription: 'List controllable and blocked browser tabs.',
-      selectTabTitle: 'Select tab',
-      selectTabDescription: 'Select a tab as the target for later browser tools.',
-      closeTabTitle: 'Close tab',
-      closeTabDescription: 'Close a tab; defaults to the current controlled tab.',
-      snapshotTitle: 'Read page snapshot',
-      snapshotDescription: 'Return an Accessibility Tree text snapshot with interactable element ids.',
-      clickTitle: 'Click element',
-      clickDescription: 'Click a page element by snapshot uid or CSS selector.',
-      hoverTitle: 'Hover element',
-      hoverDescription: 'Move the mouse over a page element by snapshot uid or CSS selector.',
-      fillTitle: 'Fill field',
-      fillDescription: 'Fill one input-like element by snapshot uid or CSS selector.',
-      fillFormTitle: 'Fill form',
-      fillFormDescription: 'Fill multiple input-like elements in one action.',
-      keyTitle: 'Press key',
-      keyDescription: 'Send a keyboard key to the page or a target element.',
-      typeTitle: 'Type text',
-      typeDescription: 'Insert text into the focused element or a target element.',
-      attachFileTitle: 'Attach file',
-      attachFileDescription: 'Set local file paths on a page file input.',
-      waitForTitle: 'Wait for condition',
-      waitForDescription: 'Wait until a selector, text, or JavaScript expression matches.',
-      handleDialogTitle: 'Handle dialog',
-      handleDialogDescription: 'Accept or dismiss the current JavaScript alert/confirm/prompt dialog.',
-      evaluateScriptTitle: 'Evaluate script',
-      evaluateScriptDescription: 'Run JavaScript in the controlled page and return a serializable result.',
-    },
-  },
-  prompt: {
-    toolCallReminderTitle: 'Tool call format reminder:',
-    taskCompleteTag: 'task_complete',
-    memoryEmpty: '(No memories yet)',
-    memoryDisabled: '(Memory injection disabled for this request)',
-    standaloneMemories: '## Existing Memories\n{memories}',
-    forceResponseLanguage: '## Response Language\nReply in {language}.',
-    responseLanguageChinese: 'Simplified Chinese',
-    responseLanguageEnglish: 'English',
-    skillUserInputWrapper: '{instructions}\n\n---\n\nThe following is the user input for this turn. Follow the instructions above when handling it:\n\n{userInput}',
-    localSkillSystemContextHeader: '## Local Skill Activated',
-    localSkillActivationDirective: 'Local Skill "{skillName}" has been auto-activated. Before taking any action, you MUST complete the following prerequisite steps; otherwise you are forbidden from starting the task: use the `local_file_read` tool to read the entire contents of {skillMdPath}; understand its SOP, tool boundaries, and constraints section by section. Only after reading and understanding that file may you strictly follow its instructions to execute the user\'s request; before that, do not call any business tool and do not begin work directly.',
-    inlineAgent: {
-      continuationIntro: 'These are the tool results just executed for the tool-continuation task. Continue like a real agent, using the original task and these tool results to move the work forward.',
-      continuationEnough: 'If the results are enough, output the final answer. Only call more tools when more information, verification, or file changes are truly needed.',
-      continuationNoPseudo: 'Do not ask the user to click continue, and do not output pseudo tool-call JSON. When more action is needed, output only executable XML tool tags.',
-      failureRecovery: 'At least one tool failed. Do not stop because of a recoverable error; read summary/detail/error, fix the parameters or choose a suitable next step, and continue completing the task.',
-      nudgeNoTools: 'The previous reply did not include executable tool XML, so the automated continuation cannot proceed.',
-      nudgeChoice: 'Choose exactly one path based on the original task and tool results:',
-      nudgeNextTool: '1. If the task is still incomplete, this turn MUST directly output the next executable tool XML.',
-      nudgeComplete: '2. If the task is complete, output <task_complete>{"summary":"..."}</task_complete>.',
-      nudgeCount: 'This is no-tool-call correction attempt {count}.',
-    },
-    automation: {
-      continuationIntro: 'These are the MCP tool results just executed for the automation. Continue completing the automation based on these results.',
-      continuationEnough: 'If the results are enough, output the final conclusion. Only call more tools when more information is truly needed.',
-    },
-    systemChat: `## Role
-You are the user's personal AI assistant with long-term cross-conversation memory. You can remember the user's identity, preferences, technical stack, and key context from prior conversations so future replies are personalized and useful.
-
-## Existing Memories
-{memories}
-
-## Tools
-
-You have access to a set of tools. To call a tool, output an XML block with the tool name itself as the tag and a JSON object as the body, exactly like this:
-
-<memory_save>
-{"type": "user", "name": "User role", "content": "Frontend developer", "tags": ["frontend"]}
-</memory_save>
-
-The JSON body MUST be valid JSON on its own. Do NOT add any other text inside the tags, only JSON. Use forward slashes or escaped backslashes for local file paths. You can place tool calls anywhere in your reply (not only at the end).
-The extension only executes direct tool-name tags. Never use wrapper formats such as <invoke name="tool_name">...</invoke> or <tool_call>...</tool_call>.
-The tag name MUST exactly match one of the available tool names.
-If a tool is listed in Available Tools, it is connected through the extension and you can call it by emitting the XML tag. Do NOT say you cannot call listed MCP tools.
-Never output pseudo tool-call JSON such as {"tool":"name","arguments":{...}} in a Markdown code block. That is explanation text, not an executable call.
-Never place executable tool XML in a thinking/reasoning section. Put tool XML in the final assistant answer content so the extension can execute it.
-
-### Extended Capabilities
-
-If the Available Tools list below includes any of these capabilities, you may proactively use them to complete the user's task without asking for permission first — the user has opted in by enabling them:
-
-- Local command execution (running CLI tools, scripts, file operations on the user's machine)
-- Browser control (opening pages, clicking, typing, reading page snapshots)
-- Web search and web fetch (real-time information, news, documentation)
-- Office document processing (.docx / .xlsx / .pptx)
-- Long-term memory (saving durable user context via memory_save)
-
-Only treat a capability as available when its concrete tool actually appears in the Available Tools list. Do not claim or exercise a capability whose tool is not listed, and do not invent tool names or results. Prefer the tools that are already connected; only request the user to enable a capability when the task clearly needs it and the tool is absent.
-
-### Available Tools
-
-{tools}
-
-You MUST strictly follow the tool names and parameter schemas above when invoking tools.
-
-## Memory Saving Rules
-
-When any of the following appears in conversation, you MUST call the memory_save tool:
-- The user mentions their identity, profession, or role
-- The user expresses preferences, habits, or working style
-- The user corrects your answer style or behavior
-- Important technical decisions or architecture choices appear
-- The user explicitly says "remember", "note this", "do not forget", or similar
-
-### Example
-
-User: I am a frontend developer and mainly use React and TypeScript
-Assistant reply:
-
-Got it. React + TypeScript is a common modern frontend stack. Ask me anything related to it.
-
-<memory_save>
-{"type": "user", "name": "User role and tech stack", "content": "Frontend developer, mainly uses React and TypeScript", "tags": ["frontend", "React", "TypeScript"]}
-</memory_save>
-
-### Rules
-- You may call tools anywhere in your reply, not only at the end
-- Tool calls are executed automatically and results are returned to you
-- Save only information with long-term value, not one-off Q&A
-- Do not save information that already exists in Existing Memories
-
-`,
-    systemThinking: `You have long-term memory. Existing memories:
-
-{memories}
-
-## Tools
-
-You have access to a set of tools. To call a tool, output an XML block with the tool name itself as the tag and a JSON object as the body, exactly like this:
-
-<memory_save>
-{"type": "user", "name": "User role", "content": "Frontend developer", "tags": ["frontend"]}
-</memory_save>
-
-The JSON body MUST be valid JSON on its own. Do NOT add any other text inside the tags, only JSON. Use forward slashes or escaped backslashes for local file paths.
-The extension only executes direct tool-name tags. Never use wrapper formats such as <invoke name="tool_name">...</invoke> or <tool_call>...</tool_call>.
-The tag name MUST exactly match one of the available tool names.
-If a tool is listed in Available Tools, it is connected through the extension and you can call it by emitting the XML tag. Do NOT say you cannot call listed MCP tools.
-Never output pseudo tool-call JSON such as {"tool":"name","arguments":{...}} in a Markdown code block. That is explanation text, not an executable call.
-Never place executable tool XML in a thinking/reasoning section. Put tool XML in the final assistant answer content so the extension can execute it.
-
-### Available Tools
-
-{tools}
-
-You MUST strictly follow the tool names and parameter schemas above when invoking tools.
-
-When the user reveals important durable information (identity, preference, behavior correction, or important decision), you MUST call memory_save. You may call tools anywhere in your reply. Save only information with long-term value; do not duplicate existing memories.
-
----
-
-`,
-    webSearchGuidance: `## Web Search Rules
-
-Use the web_search tool when any of these apply:
-- The user asks about real-time information, news, events, exchange rates, weather, or similar
-- The user asks about knowledge you are not sure about and current sources are needed
-- The user explicitly asks you to search or look something up
-- You need to verify facts, data, or cited sources
-
-### Search Flow
-1. First output a web_search tool call
-2. The search runs automatically; results are shown on the page and sent back to you
-3. Read the results, then answer based on them
-
-### Example
-
-User: Who won the 2024 Nobel Prizes?
-Assistant reply:
-
-I will search for the latest information.
-
-<web_search>
-{"query": "2024 Nobel Prize winners"}
-</web_search>
-
-### Rules
-- Use keywords that match the user's language and target sources
-- If one search is not enough, call web_search again with different keywords
-- Do not invent real-time information without searching
-`,
-    toolFormatReminder: `---
-Tool call format reminder:
-Available tool tag names: {names}
-These listed tools are executable by the extension. Do not claim you cannot call a listed MCP tool.
-To call a tool, use ONLY the direct XML tag whose name is the tool name, with valid JSON as the body.
-For MCP tools, prefer the short tag name when it appears in the available names list.
-For local file paths, use forward slashes or escaped backslashes so the JSON body remains valid.
-Do not use <invoke name="...">, <tool_call>, Markdown code fences, {"tool":"...","arguments":{...}}, or any wrapper format.
-Do not put executable tool XML in a thinking/reasoning section; put it in the final assistant answer content.
-`,
-    pythonHintTitle: '### Python Quick Validation Capability',
-    pythonHintExec: 'Use <{execName}> for short Python snippets that verify an idea, perform complex calculations, or transform small data. Treat it as a scratchpad, not as a general local execution environment.',
-    pythonHintStatus: 'Use <{statusName}>{}</{statusName}> when you need to know the Python version or whether numpy, pandas, or sympy are available.',
-    pythonHintAvailability: 'Assume the Python standard library is available. Only use numpy, pandas, or sympy after python_status reports them as available.',
-    pythonHintSafety: 'Do not install packages, access sensitive local files, run long jobs, or use network access through Python. Keep code short and return concise text or JSON.',
-    shellHintTitle: '### Shell MCP Capability',
-    shellHintConnected: 'Shell MCP is connected through the extension. You can execute local CLI commands by emitting the executable XML tool tag; do not say you cannot run commands when this tool is listed.',
-    shellHintExec: 'Use <{execName}> with a JSON body such as {"command":"officecli --version","timeout_ms":60000} to run OfficeCLI or other local CLI tools.',
-    shellHintStatus: 'Use <{statusName}>{}</{statusName}> first when you need host status, shell, PATH, or working-directory context.',
-    shellHintWindows: 'Match command syntax to shell_status.shell. On Windows the Shell Local host uses PowerShell by default, so list files with commands such as Get-ChildItem -LiteralPath "D:\\\\Documents\\\\Downloads\\\\CN" -File | Select-Object -ExpandProperty FullName, and quote paths once inside the command string. Use cmd.exe /c explicitly only when you need CMD syntax such as dir /b.',
-    shellHintSession: 'For multi-step workflows (e.g. OfficeCLI create + open + add, or anything that needs to keep cd/export state) use a persistent session: open it with <shell_session_begin>{"cwd":"/path"}</shell_session_begin>, run each command with <shell_session_exec>{"session_id":"...","command":"..."}</shell_session_exec>, then close with <shell_session_end>{"session_id":"..."}</shell_session_end>. Take the session_id from the begin result. Every shell_session_exec reuses the same shell, so the working directory, exported variables, and resident processes persist across calls; the session auto-closes after roughly 5 minutes idle.',
-    shellHintNames: 'Recognized shell tool names: {names}',
-  },
-  pet: {
-    lines: {
-      thinking: ['Thinking...', 'Reasoning...', 'Working it through'],
-      working: ['Working', 'Crafting', 'Building'],
-      speaking: ['Explaining', 'Unfolding', 'Finding the thread'],
-      idle: ['Idle', 'Tap to check in', 'Taking a breath'],
-      confused: ['Reorienting', 'Sorting it out', 'Finding the path'],
-      success: ['Done', 'Handled', 'Finished'],
-      error: ['Stuck...', 'Something failed', 'Needs a retry'],
-      sleepy: ['Zzz...', 'Sleepy...', 'Taking a nap'],
-    },
-  },
+  common,
+  manifest,
+  app,
+  locale,
+  sidepanel,
+  content,
+  background,
+  tool,
+  prompt,
+  pet,
 } as const satisfies LocaleMessages;
