+++ core/i18n/resources/en/app.ts
@@ -1,0 +1,15 @@
+export const app = {
+  sideNavLabel: 'Sidebar navigation',
+  version: 'v{version}',
+  tabs: {
+    chat: 'Chat',
+    library: 'Library',
+    memory: 'Memory',
+    projects: 'Projects',
+    saved: 'Saved',
+    capabilities: 'Capabilities',
+    preset: 'Presets',
+    automation: 'Automation',
+    settings: 'Settings',
+  },
+} as const;
