+++ core/i18n/resources/en/background.ts
@@ -1,0 +1,44 @@
+export const background = {
+  contextMenus: {
+    sendToChat: 'Send to chat',
+  },
+  auth: {
+    missingDeepSeek: 'Sign in at chat.deepseek.com first, or refresh the DeepSeek page and try again.',
+  },
+  export: {
+    generating: 'Generating export files',
+    cancelled: 'Export cancelled',
+  },
+  sync: {
+    missingWebDav: 'WebDAV is not configured',
+    missingSync: 'Cloud sync is not configured',
+    configChanged: 'Sync settings changed in another extension window. Review the current target and try again.',
+    configCommitIndeterminate: 'Could not confirm whether the sync settings were saved. The current settings will be reloaded.',
+    missingRemoteFile: 'Remote file {file} is missing. Download stopped to avoid overwriting local data.',
+    authorizationNotRequired: 'The current sync method does not require authorization',
+    identityUnavailable: 'OAuth authorization is not available in this browser context',
+    oauthCancelled: 'Authorization was cancelled',
+    oauthProviderError: 'Authorization failed: {error}',
+    oauthMissingCode: 'Authorization failed: no authorization code was returned',
+    oauthRefreshFailed: 'Failed to refresh access token (HTTP {status}){detail}',
+    oauthTokenExchangeFailed: 'Failed to get token (HTTP {status}){detail}',
+    gdriveMissingRefreshToken: 'Google did not return a refresh_token. Revoke access and authorize again.',
+    gdriveMissingAuthorization: 'Google Drive is not authorized. Click Authorize first.',
+    gdriveAuthorizationExpired: 'Google authorization expired. Authorize again.',
+    gdriveConnectFailed: 'Failed to connect to Google Drive (HTTP {status})',
+    gdriveQueryFailed: 'Failed to query {name} (HTTP {status})',
+    gdriveDownloadFailed: 'Failed to download {key} from Google Drive (HTTP {status})',
+    gdriveUploadFailed: 'Failed to upload {name} to Google Drive (HTTP {status})',
+    onedriveMissingRefreshToken: 'Microsoft did not return a refresh_token. Revoke access and authorize again.',
+    onedriveMissingAuthorization: 'OneDrive is not authorized. Click Authorize first.',
+    onedriveAuthorizationExpired: 'Microsoft authorization expired. Authorize again.',
+    onedriveConnectFailed: 'Failed to connect to OneDrive (HTTP {status})',
+    onedriveDownloadFailed: 'Failed to download {key} from OneDrive (HTTP {status})',
+    onedriveUploadFailed: 'Failed to upload {key} to OneDrive (HTTP {status})',
+  },
+  chat: {
+    continueWithToolResults: '[TOOL_RESULTS]\n{toolResults}\n[/TOOL_RESULTS]\n\nContinue answering based on the tool results above.',
+    maxToolSteps: '(Maximum tool-call steps reached; conversation ended)',
+    interrupted: 'The assistant was interrupted. Please send your message again.',
+  },
+} as const;
