+++ core/i18n/resources/en/common.ts
@@ -1,0 +1,32 @@
+export const common = {
+  add: 'Add',
+  all: 'All',
+  auto: 'Auto',
+  cancel: 'Cancel',
+  clear: 'Clear',
+  close: 'Close',
+  confirm: 'Confirm',
+  delete: 'Delete',
+  deactivate: 'Disable',
+  disabled: 'Disabled',
+  edit: 'Edit',
+  enable: 'Enable',
+  enabled: 'Enabled',
+  error: 'Error',
+  loading: 'Loading...',
+  manual: 'Manual',
+  none: 'None',
+  open: 'Open',
+  preview: 'Preview',
+  refresh: 'Refresh',
+  remove: 'Remove',
+  retry: 'Retry',
+  save: 'Save',
+  saveChanges: 'Save changes',
+  search: 'Search',
+  status: 'Status',
+  success: 'Success',
+  sync: 'Sync',
+  test: 'Test',
+  update: 'Update',
+} as const;
