+++ core/i18n/resources/en/content.ts
@@ -1,0 +1,141 @@
+export const content = {
+  conversation: {
+    untitled: 'Untitled conversation',
+  },
+  export: {
+    buttonIdle: 'Export current conversation',
+    buttonRunning: 'Exporting current conversation',
+    emptyConversation: 'This page has no exportable conversation yet.',
+    formatDialogLabel: 'Choose export formats',
+    formatTitle: 'Export formats',
+    scopeLabel: 'Content scope',
+    scopeFull: 'Everything (tool calls, continuations, reasoning, injected context)',
+    scopeInputOutput: 'Inputs + final output only',
+    formatHtml: 'HTML',
+    formatMarkdown: 'Markdown',
+    formatPdf: 'PDF',
+    formatImageManifest: 'Image manifest',
+    submit: 'Export',
+    progress: 'Exporting current conversation...',
+    failed: 'Export failed',
+    cancelled: 'Export cancelled',
+    partialSuccess: 'Export completed, but some sessions could not be read. Check Export Warnings in the exported file.',
+    success: 'Current conversation exported.',
+  },
+  historyOrganizer: {
+    enhancedSearchTitle: 'DeepSeek++ search enhancements',
+    tagFilterLabel: 'Filter results by tag',
+    tagPlaceholder: 'Enter a tag',
+    currentTagsLabel: 'Tag the current chat',
+    currentTagsPlaceholder: 'Comma-separated, e.g. hk stocks, writing',
+    emptySearchStatus: 'DeepSeek++: waiting for official search results',
+    visibleStatus: 'DeepSeek++: {visible}/{total} visible',
+    loadError: 'DeepSeek++: failed to load history tags: {message}',
+    saveError: 'DeepSeek++: failed to save history tags: {message}',
+  },
+  projectSidebar: {
+    title: 'Projects',
+    empty: 'No projects yet',
+    expandProject: 'Expand project: {name}',
+    collapseProject: 'Collapse project: {name}',
+    showMore: 'Show more',
+    showLess: 'Show less',
+    moveCurrentToProject: 'Move current conversation to {name}',
+    removeCurrentFromProject: 'Remove current conversation from {name}',
+    joinProject: 'Add to project',
+    joinProjectNamed: 'Add to project: {name}',
+    moveToProjectNamed: 'Move to project: {name}',
+    currentProjectNamed: 'In project: {name}',
+    removeFromProjectNamed: 'Remove from project: {name}',
+    conversationActions: 'Conversation actions',
+    newConversationInProject: 'Start a new conversation in {name}',
+    useNextConversation: 'Use {name} for the next new conversation',
+    cancelNextConversation: 'Cancel using {name} for the next new conversation',
+    pendingNextConversation: 'The next new conversation will use this project',
+    operationFailed: 'Project action failed: {message}',
+    age: {
+      justNow: 'Just now',
+      minutesAgo: '{count}m ago',
+      hoursAgo: '{count}h ago',
+      daysAgo: '{count}d ago',
+    },
+  },
+  uxPolish: {
+    downloadCode: 'Download',
+    downloadMessageMarkdownButton: 'MD',
+    downloadMessageMarkdownTitle: 'Download message as Markdown',
+    copyMessageButton: 'Copy',
+    copyMessageTitle: 'Copy full message output',
+  },
+  toolBlock: {
+    title: 'Executed tools ({count})',
+    pythonInterpreter: 'Python interpreter',
+    summaries: {
+      saved: 'Saved',
+      updated: 'Updated',
+      deleted: 'Deleted',
+      searched: 'Searched',
+      fetched: 'Fetched',
+      running: 'Running',
+      executed: 'Executed',
+      failed: 'Execution failed',
+      messageFailed: 'Tool message failed',
+      backgroundFailed: 'Background tool execution failed',
+    },
+    missingResultDetail: 'Background did not return a tool result. Refresh the current DeepSeek page and try again.',
+    missingSandboxResultDetail: 'The sandbox tool did not return an execution result. Reload the DeepSeek++ extension and refresh the current DeepSeek page, then try again; it does not depend on MCP or Shell Local.',
+    missingMcpResultDetail: 'Background did not return an MCP tool result. Refresh the current DeepSeek page and try again. If it still fails, retest Shell Local on the MCP page.',
+    invalidResultDetail: 'Background returned an invalid tool result: {preview}',
+  },
+  agent: {
+    stop: 'Stop',
+    stopped: 'Stopped',
+    step: 'Step {index}',
+    streaming: 'streaming...',
+    complete: 'Complete',
+    completeWithTools: 'Complete ({count} tools)',
+    executingTools: 'Executing tools',
+    budgetReached: 'DeepSeek++ paused after {count} automated tool-continuation rounds to avoid presenting incomplete work as a final answer. Send "continue" in this conversation to resume the remaining work.',
+    error: 'Execution error',
+    footerComplete: 'Agent complete ({steps} steps, {tools} tool calls)',
+    process: 'Process',
+    tools: 'Tool calls',
+    results: 'Results',
+    running: 'Agent running · step {step} · {tools} tool calls',
+    footerError: 'Agent error ({steps} steps, {tools} tool calls)',
+    autoSaved: 'Long output was auto-saved as an artifact: ',
+    autoSavedDownload: 'Download',
+    concurrencyGuard: 'An agent is still running. A duplicate task from this message was skipped. To interrupt, click Stop in the agent panel.',
+  },
+  permission: {
+    webFetch: 'DeepSeek++ needs permission to access {origin} so it can fetch that page',
+    deny: 'Deny',
+    grant: 'Allow',
+    requesting: 'Requesting...',
+  },
+  skillPopup: {
+    hint: '↑↓ Navigate · Enter Select · Esc Close',
+  },
+  tokenSpeed: {
+    title: 'Accumulated tokens: {tokens}; token output speed: {speed}{idle}. Token source: {tokenSource}; speed source: {speedSource}',
+    idleSuffix: ' (idle)',
+    sourceServer: 'server',
+    sourceEstimated: 'local estimate',
+  },
+  multimodalMedia: {
+    buttonTitle: 'Add DeepSeek++ multimodal images or videos',
+    removeTitle: 'Remove {name}',
+    pastedFileName: 'Pasted media',
+    added: 'Added {count} multimodal attachment(s)',
+    analyzing: 'Analyzing {count} multimodal attachment(s)...',
+    analyzed: 'Multimodal analysis complete ({count} item(s))',
+    failed: 'Multimodal analysis failed: {message}',
+    tooMany: 'Attach at most {count} multimodal files per turn',
+    unsupported: '{name} is not a supported image or video',
+    imageTooLarge: '{name} exceeds the image limit of {limit}',
+    videoTooLarge: '{name} exceeds the inline video limit of {limit}. Use a public video URL or wait for local-path picker support.',
+    invalidRequest: 'The DeepSeek request body is invalid, so multimodal media cannot be attached.',
+    emptyPrompt: 'Type a message before sending multimodal media.',
+  },
+  extensionReloaded: 'The extension was reloaded. Refresh the current DeepSeek page and try again.',
+} as const;
