+++ core/i18n/resources/en/locale.ts
@@ -1,0 +1,7 @@
+export const locale = {
+  settingTitle: 'Language',
+  settingDescription: 'Choose the language for DeepSeek++ UI and model behavior.',
+  auto: 'Follow browser',
+  zhCN: '简体中文',
+  en: 'English',
+} as const;
