+++ core/i18n/resources/en/manifest.ts
@@ -1,0 +1,5 @@
+export const manifest = {
+  name: 'DeepSeek++',
+  description: 'Agentic memory, skills, execution, automation, and MCP tools for DeepSeek',
+  actionTitle: 'DeepSeek++',
+} as const;
