+++ core/i18n/resources/en/pet.ts
@@ -1,0 +1,12 @@
+export const pet = {
+  lines: {
+    thinking: ['Thinking...', 'Reasoning...', 'Working it through'],
+    working: ['Working', 'Crafting', 'Building'],
+    speaking: ['Explaining', 'Unfolding', 'Finding the thread'],
+    idle: ['Idle', 'Tap to check in', 'Taking a breath'],
+    confused: ['Reorienting', 'Sorting it out', 'Finding the path'],
+    success: ['Done', 'Handled', 'Finished'],
+    error: ['Stuck...', 'Something failed', 'Needs a retry'],
+    sleepy: ['Zzz...', 'Sleepy...', 'Taking a nap'],
+  },
+} as const;
