+++ core/i18n/resources/en/tool.ts
@@ -1,0 +1,215 @@
+export const tool = {
+  runtime: {
+    invalidFormat: 'Invalid tool format',
+    unknownTool: 'Unknown tool',
+    authorizationRejected: 'Tool authorization rejected',
+    authorizationMissingHint: 'The authorization is closed or was never confirmed. Re-run the action and allow this tool in the authorization prompt.',
+    authorizationStaleHint: 'The authorization expired. Re-run the action to refresh it.',
+    authorizationSessionMismatchHint: 'The authorization belongs to another chat session. Re-run the action in this conversation.',
+    authorizationDisabledHint: 'This tool is currently disabled. Enable it in tool settings and retry.',
+    authorizationIdentityMissingHint: 'Identity information required for execution is missing. Re-run the action.',
+    authorizationGenericHint: 'To continue, re-run the action and complete the authorization confirmation.',
+    sessionBindingUnavailable: 'The new conversation did not receive a browser-owned session route in time. Please wait for the chat URL to update and try again.',
+    incomplete: 'The response ended before this tool call completed.',
+  },
+  memory: {
+    providerName: 'DeepSeek++ Memory',
+    saveTitle: 'Save memory',
+    saveDescription: 'Save a new long-term memory',
+    typeDescription: 'Memory type: user=identity, role, or preference; feedback=behavior correction; topic=discussion notes; reference=external resource link',
+    nameDescription: 'Short title',
+    contentDescription: 'Content to save',
+    tagsDescription: 'Tag list',
+    updateTitle: 'Update memory',
+    updateDescription: 'Update an existing memory',
+    idDescription: 'Memory ID',
+    updatedNameDescription: 'Updated title',
+    updatedContentDescription: 'Updated content',
+    deleteTitle: 'Delete memory',
+    deleteDescription: 'Delete memory',
+    unsupported: 'Unsupported memory tool',
+    saveFailed: 'Save failed',
+    saveMissingConfirmation: 'No save confirmation received',
+    saved: 'Saved',
+    invalidPayload: 'Invalid memory payload',
+    invalidType: 'type must be user, feedback, topic, or reference',
+    invalidName: 'name must be a non-empty string',
+    invalidContent: 'content must be a non-empty string',
+    invalidTags: 'tags must be an array of strings',
+    invalidId: 'Invalid ID',
+    notFound: 'Memory not found',
+    notFoundDetail: 'ID {id} does not exist',
+    updated: 'Updated',
+    deleted: 'Deleted',
+  },
+  web: {
+    providerName: 'DeepSeek++ Web Search',
+    searchTitle: 'Search the web',
+    searchDescription: 'Search the web and return matching page titles, URLs, and snippets',
+    queryDescription: 'Search query keywords',
+    topKDescription: 'Number of results to return, default 5',
+    fetchTitle: 'Fetch web page',
+    fetchDescription: 'Download the specified URL and return visible text after removing navigation, scripts, and styles',
+    urlDescription: 'Full URL to fetch (http:// or https://)',
+    unsupported: 'Unsupported web tool',
+    emptyQuery: 'Search query cannot be empty',
+    searchComplete: 'Search completed with {count} results',
+    searchNoResults: 'No search results',
+    searchFailed: 'Search failed',
+    permissionDenied: 'The extension does not have permission to access Bing. Remove and reload the dist/chrome-mv3 extension directory, or confirm cn.bing.com is listed under site access in chrome://extensions -> DeepSeek++ details.',
+    noParseableResults: 'No parseable search results found: {error}',
+    searchFailedDetail: 'Search failed: {error}',
+    emptyUrl: 'URL cannot be empty',
+    invalidUrl: 'Invalid URL',
+    invalidUrlDetail: 'Could not parse URL: {url}',
+    contentType: 'Page type: {contentType}',
+    contentTypeDetail: 'Content-Type: {contentType}\nURL: {url}\n(Content is not text and cannot be displayed)',
+    truncated: '\n\n... [Content truncated, {count} characters total]',
+    fetchComplete: 'Fetched {url}',
+    fetchTruncatedDetail: 'Page length {length} characters, truncated to {maxLength} characters',
+    fetchLengthDetail: 'Page length {length} characters',
+    fetchFailed: 'Failed to fetch page',
+    missingHostPermission: 'Cannot access {url}; host permission is missing.',
+  },
+  artifact: {
+    createTitle: 'Create downloadable file',
+    createDescription: 'Create a downloadable UTF-8 text artifact or runnable code artifact. Use it for HTML previews and JavaScript, TypeScript, or Python runners. Never embed binary or base64 Office, PDF, image, or archive data; keep files produced by Shell or OfficeCLI at their local path and report that path.',
+    bundleTitle: 'Create downloadable project bundle',
+    bundleDescription: 'Create a ZIP bundle from multiple generated text files. Use this for multi-file projects and scaffolds.',
+    filenameDescription: 'Safe relative file name, for example report.md or src/index.ts',
+    bundleFilenameDescription: 'ZIP filename, for example project.zip',
+    contentDescription: 'Complete UTF-8 text content; never binary or base64 file data',
+    mimeTypeDescription: 'MIME type, for example text/markdown or application/json',
+    previewModeDescription: 'Artifact view mode: auto infers from content, html previews a page, code shows a run button, none keeps it download-only.',
+    languageDescription: 'Artifact language: html, javascript, typescript, python, or text. Inferred from filename and MIME type when omitted.',
+    filesDescription: 'Files to include in the bundle. Each item has path, content, and optional mimeType.',
+    failed: 'Artifact failed',
+    unsupported: 'Unsupported artifact tool',
+    fileReady: 'File ready',
+    bundleReady: 'Bundle ready',
+    result: {
+      files: '{count} files',
+      download: 'Download',
+      downloading: 'Downloading...',
+      downloaded: 'Downloaded',
+      downloadFailed: 'Download failed',
+      preview: 'Preview',
+      previewFailed: 'Preview failed',
+      previewAriaLabel: 'Artifact preview',
+      closePreview: 'Close preview',
+      run: 'Run',
+      running: 'Running...',
+      runAgain: 'Run again',
+      failed: 'Failed',
+      done: 'Done',
+      artifactNotFound: 'Artifact not found',
+      codeExecuted: 'Code executed',
+      codeFailed: 'Code failed',
+      runnerUnavailable: 'Code runner did not return a result',
+    },
+  },
+  sandbox: {
+    runTitle: 'Run code in browser sandbox',
+    runDescription: 'Execute JavaScript, TypeScript, Python, or HTML in an isolated extension sandbox, not in the DeepSeek page.',
+    languageDescription: 'Runtime language: javascript, typescript, python, or html',
+    codeDescription: 'Code to run in the isolated browser sandbox',
+    inputDescription: 'Optional input string available to the code',
+    timeoutDescription: 'Timeout in milliseconds, default 5000',
+    executed: 'Sandbox executed',
+    failed: 'Sandbox failed',
+    invalidRequest: 'Invalid sandbox request',
+    offscreenUnavailable: 'Browser sandbox unavailable',
+    offscreenUnavailableDetail: 'This browser does not expose the MV3 offscreen API needed for isolated JavaScript and TypeScript sandbox execution.',
+    runtimeUnavailableDetail: 'The browser sandbox executor is only available from the extension background runtime.',
+    pythonUnavailable: 'Python sandbox unavailable',
+    pythonUnavailableDetail: 'The bundled Pyodide runtime could not start in the isolated browser sandbox.',
+  },
+  skillCreator: {
+    title: 'Draft a Skill',
+    description: 'Create a custom Skill draft for user review. The draft is not saved until the user approves it.',
+    nameDescription: 'Lowercase Skill trigger name',
+    descriptionDescription: 'When this Skill should be used',
+    instructionsDescription: 'Markdown instructions for the Skill',
+    memoryDescription: 'Whether the Skill should receive memory context',
+    draftReady: 'Skill draft ready',
+    invalidDraft: 'Invalid Skill draft',
+    result: {
+      saveSkill: 'Save Skill',
+      saving: 'Saving...',
+      saved: 'Saved',
+      saveFailed: 'Save failed',
+      memoryOn: 'memory on',
+      memoryOff: 'memory off',
+    },
+  },
+  memoryImport: {
+    title: 'Preview memory import',
+    description: 'Convert another AI workflow export or notes into memory drafts. Nothing is saved until the user imports the preview.',
+    contentDescription: 'JSON, Markdown, or plain-text memory source',
+    typeDescription: 'Default memory type: user, feedback, topic, or reference',
+    tagsDescription: 'Tags to add to imported memories',
+    previewReady: 'Memory import preview ready ({count})',
+    previewDetail: '{duplicates} duplicates skipped, {rejected} rejected',
+    failed: 'Memory import failed',
+    result: {
+      metaSummary: '{count} memories · {duplicates} duplicates',
+      importMemories: 'Import memories',
+      importing: 'Importing...',
+      imported: 'Imported {count}',
+      importFailed: 'Import failed',
+    },
+  },
+  mcpCapability: {
+    providerName: 'MCP Capability Catalog',
+    discoverTitle: 'Discover MCP capabilities',
+    discoverDescription: 'Search available MCP tools that were not directly shown. Discover a capability before describing or invoking it.',
+    describeTitle: 'Describe an MCP capability',
+    describeDescription: 'Use a capability from discovery to inspect one MCP tool and its full input schema without executing it.',
+    invokeTitle: 'Invoke a discovered MCP capability',
+    invokeDescription: 'Invoke only a capability returned by mcp_discover. A raw tool name cannot bypass authorization.',
+    discoverQueryDescription: 'Task or capability to search for.',
+    discoverLimitDescription: 'Maximum number of candidates.',
+    capabilityDescription: 'Capability returned by mcp_discover.',
+    pointerDescription: 'Optional JSON Pointer to one schema field.',
+    argumentsDescription: 'Arguments for the resolved MCP tool.',
+  },
+  browser: {
+    providerName: 'DeepSeek++ Browser Control',
+    navigateTitle: 'Navigate',
+    navigateDescription: 'Open a URL in a new tab by default; set newTab=false only to replace the current controlled tab.',
+    goBackTitle: 'Go back',
+    goBackDescription: 'Move the controlled tab back in browser history.',
+    goForwardTitle: 'Go forward',
+    goForwardDescription: 'Move the controlled tab forward in browser history.',
+    refreshTitle: 'Refresh',
+    refreshDescription: 'Reload the controlled tab.',
+    listTabsTitle: 'List tabs',
+    listTabsDescription: 'List controllable and blocked browser tabs.',
+    selectTabTitle: 'Select tab',
+    selectTabDescription: 'Select a tab as the target for later browser tools.',
+    closeTabTitle: 'Close tab',
+    closeTabDescription: 'Close a tab; defaults to the current controlled tab.',
+    snapshotTitle: 'Read page snapshot',
+    snapshotDescription: 'Return an Accessibility Tree text snapshot with interactable element ids.',
+    clickTitle: 'Click element',
+    clickDescription: 'Click a page element by snapshot uid or CSS selector.',
+    hoverTitle: 'Hover element',
+    hoverDescription: 'Move the mouse over a page element by snapshot uid or CSS selector.',
+    fillTitle: 'Fill field',
+    fillDescription: 'Fill one input-like element by snapshot uid or CSS selector.',
+    fillFormTitle: 'Fill form',
+    fillFormDescription: 'Fill multiple input-like elements in one action.',
+    keyTitle: 'Press key',
+    keyDescription: 'Send a keyboard key to the page or a target element.',
+    typeTitle: 'Type text',
+    typeDescription: 'Insert text into the focused element or a target element.',
+    attachFileTitle: 'Attach file',
+    attachFileDescription: 'Set local file paths on a page file input.',
+    waitForTitle: 'Wait for condition',
+    waitForDescription: 'Wait until a selector, text, or JavaScript expression matches.',
+    handleDialogTitle: 'Handle dialog',
+    handleDialogDescription: 'Accept or dismiss the current JavaScript alert/confirm/prompt dialog.',
+    evaluateScriptTitle: 'Evaluate script',
+    evaluateScriptDescription: 'Run JavaScript in the controlled page and return a serializable result.',
+  },
+} as const;
