+++ core/i18n/resources/zh-CN.ts
@@ -1,1530 +1,26 @@
 import type { LocaleSchema } from '../types';
+import { common } from './zh-CN/common';
+import { manifest } from './zh-CN/manifest';
+import { app } from './zh-CN/app';
+import { locale } from './zh-CN/locale';
+import { sidepanel } from './zh-CN/sidepanel';
+import { content } from './zh-CN/content';
+import { background } from './zh-CN/background';
+import { tool } from './zh-CN/tool';
+import { prompt } from './zh-CN/prompt';
+import { pet } from './zh-CN/pet';
 
 export const zhCN = {
-  common: {
-    add: '新增',
-    all: '全部',
-    auto: '自动',
-    cancel: '取消',
-    clear: '清除',
-    close: '关闭',
-    confirm: '确认',
-    delete: '删除',
-    deactivate: '停用',
-    disabled: '已停用',
-    edit: '编辑',
-    enable: '启用',
-    enabled: '已启用',
-    error: '错误',
-    loading: '加载中...',
-    manual: '手动',
-    none: '无',
-    open: '打开',
-    preview: '预览',
-    refresh: '刷新',
-    remove: '移除',
-    retry: '重试',
-    save: '保存',
-    saveChanges: '保存更改',
-    search: '搜索',
-    status: '状态',
-    success: '成功',
-    sync: '同步',
-    test: '测试',
-    update: '更新',
-  },
-  manifest: {
-    name: 'DeepSeek++',
-    description: 'DeepSeek 的 Agentic 记忆、Skill、执行、自动化和 MCP 工具',
-    actionTitle: 'DeepSeek++',
-  },
-  app: {
-    sideNavLabel: '侧栏导航',
-    version: 'v{version}',
-    tabs: {
-      chat: '对话',
-      library: '资料',
-      memory: '记忆',
-      projects: '项目',
-      saved: '保存',
-      capabilities: '能力',
-      preset: '预设',
-      automation: '自动化',
-      settings: '设置',
-    },
-  },
-  locale: {
-    settingTitle: '语言',
-    settingDescription: '选择 DeepSeek++ 的界面和模型行为语言。',
-    auto: '跟随浏览器',
-    zhCN: '简体中文',
-    en: 'English',
-  },
-  sidepanel: {
-    subTabs: {
-      scrollLeft: '向左滚动标签',
-      scrollRight: '向右滚动标签',
-    },
-    libraryPage: {
-      navLabel: '资料子导航',
-      tabs: {
-        memory: '记忆',
-        saved: '保存',
-      },
-    },
-    memory: {
-      actions: {
-        pin: '置顶',
-        unpin: '取消置顶',
-      },
-      age: {
-        justNow: '刚刚',
-        minutesAgo: '{count}分钟前',
-        hoursAgo: '{count}小时前',
-        daysAgo: '{count}天前',
-      },
-      form: {
-        namePlaceholder: '标题',
-        contentPlaceholder: '内容',
-        tagsPlaceholder: '标签（逗号分隔）',
-      },
-      types: {
-        user: '用户',
-        feedback: '反馈',
-        topic: '话题',
-        reference: '参考',
-      },
-    },
-    preset: {
-      activeBadge: '生效中',
-      form: {
-        namePlaceholder: '预设名称（如：代码助手、翻译专家）',
-        contentLabel: '系统提示词内容',
-        contentPlaceholder: '你是一位专业的...\n\n## 核心原则\n- ...',
-      },
-    },
-    skill: {
-      disabledBadge: '已停用',
-      memoryEnabledBadge: '含记忆注入',
-      unknownLicense: '未知许可',
-      actions: {
-        enableSkill: '启用 {name}',
-        disableSkill: '停用 {name}',
-        editSkill: '编辑 {name}',
-        deleteSkill: '删除 {name}',
-        updateSkill: '更新本地 Skill（重读索引）',
-      },
-      autoActivation: {
-        title: '自动激活',
-        description: '开启后，DeepSeek++ 会根据对话内容自动匹配并激活本地 Skill；关闭则只有你显式调用 /技能名 时才能激活。',
-        firstMessage: '新对话首条消息自动激活',
-        firstMessageDescription: '仅在新对话的第一条消息上自动匹配并激活本地 Skill。',
-        everyMessage: '当前对话每条消息自动激活',
-        everyMessageDescription: '对当前对话的每一条消息都自动匹配并激活；开启时联动开启“首条消息”。',
-      },
-      form: {
-        namePlaceholder: '名称（如 my-skill）',
-        triggerCommand: '触发命令：',
-        descriptionPlaceholder: '描述（何时使用这个 skill）',
-        instructionsLabel: '指令（Markdown 格式，告诉 AI 如何执行）',
-        instructionsPlaceholder: '你是一位...\n\n## 核心原则\n- ...',
-        memoryInjectionLabel: '启用记忆注入',
-      },
-      sources: {
-        builtin: '内置',
-        official: '第三方',
-        thirdParty: '第三方',
-        custom: '自定义',
-        remote: 'GitHub',
-        local: '本地',
-      },
-    },
-    scenario: {
-      title: '右键场景',
-      description: '选中文本后右键可发送到侧边栏对话',
-      customTitle: '自定义场景',
-      namePlaceholder: '场景名称',
-      templatePlaceholder: 'Prompt 模板（含 {text}）',
-      operationFailed: '场景操作失败：{error}',
-      backgroundUnavailable: '后台右键菜单刷新失败。',
-      savedButReloadFailed: '场景已保存，但列表刷新失败：{error}',
-      savedButMenuFailed: '场景已保存，但后台右键菜单刷新失败：{error}',
-    },
-    githubSkillImport: {
-      title: '从 GitHub 导入 Skill',
-      description: '支持仓库、目录或单个 SKILL.md 链接。导入前会预览内容，不会覆盖本地自定义 Skill。',
-      urlPlaceholder: 'https://github.com/owner/repo 或 .../SKILL.md',
-      permissionError: '需要 GitHub API 访问权限才能读取仓库 Skill',
-      previewFailed: '预览失败',
-      importFailed: '导入失败',
-      importedMessage: '已导入 {count} 个 Skill',
-      selectAll: '全选',
-      clearSelection: '取消全选',
-      selectedSummary: '已选 {selected} / {total} · {bytes}',
-      importSelected: '导入选中 Skill',
-      openRepository: '打开仓库',
-      repoRoot: '仓库根目录',
-      unknownLicense: '未知',
-      warningOverflow: '还有 {count} 条警告',
-      renamedBadge: '已改名',
-      resourceCount: '资源 {count}',
-      omittedCount: '省略 {count}',
-      renamedNotice: '有 {count} 个 Skill 因命名冲突自动加后缀。',
-      meta: {
-        license: '许可',
-        version: '版本',
-        skill: 'Skill',
-        defaultBranch: '默认分支',
-      },
-    },
-    localSkillImport: {
-      title: '从本地导入 Skill',
-      description: '通过 Shell MCP 读取本地 Skill 文件夹。DeepSeek++ 会保留文件夹引用，导入时不会执行脚本。',
-      pathPlaceholder: '/Users/me/.codex/skills/my-skill',
-      pickFolder: '选择',
-      pickFailed: '选择文件夹失败',
-      previewFailed: '预览失败',
-      importFailed: '导入失败',
-      importedMessage: '已导入 {count} 个本地 Skill',
-      selectAll: '全选',
-      clearSelection: '取消全选',
-      selectedSummary: '已选 {selected} / {total} · {bytes}',
-      importSelected: '导入选中 Skill',
-      referencedMode: '引用文件夹',
-      warningOverflow: '还有 {count} 条警告',
-      renamedBadge: '已改名',
-      resourceCount: '已内嵌 {count}',
-      scriptCount: '脚本 {count}',
-      omittedCount: '按需读取 {count}',
-      notBundledCount: '未内嵌 {count}',
-      omittedExplanation: '另有 {count} 个支持文件因数量或大小上限未嵌入上下文。文件没有被删除，仍保留在本地 Skill 目录，可通过 Shell MCP 按需读取。',
-      omittedUnavailableExplanation: '另有 {count} 个支持文件未嵌入上下文。文件没有被删除，仍保留在本地 Skill 目录，但当前无法按需读取；请先修复受影响 Skill 的 Shell MCP 读取器再导入。',
-      readerUnavailable: '按需读取器不可用',
-      readerTechnicalDetail: '技术详情：{detail}',
-      readerActions: {
-        updateHost: '请在 MCP > Shell Local 重新安装 Shell Native Host，重启浏览器后再预览。',
-        enableReader: '请将 Shell Local 执行模式设为“自动”，并允许 local_file_read，然后重新预览。',
-        checkConnection: '请检查 Shell Local 连接后重新预览。',
-      },
-      renamedNotice: '有 {count} 个 Skill 因命名冲突自动加后缀。',
-      meta: {
-        skill: 'Skill',
-        mode: '模式',
-      },
-    },
-    projectsPage: {
-      title: '项目',
-      description: '把 DeepSeek 对话加入项目，项目指令和项目记忆会随这些对话自动进入上下文。',
-      summary: '{projects} 个项目，{conversations} 个项目对话',
-      createTitle: '创建项目',
-      namePlaceholder: '项目名称',
-      descriptionPlaceholder: '项目说明（可选）',
-      instructionsPlaceholder: '项目指令',
-      createProject: '创建项目',
-      projectCreated: '已创建项目「{name}」',
-      empty: '暂无项目',
-      emptyHelp: '创建项目后，可以把当前 DeepSeek 对话加入项目，或让下一条新会话自动进入项目。',
-      deleteProject: '删除',
-      deleteConfirm: '确定删除项目“{name}”吗？',
-      conversationCount: '{count} 个对话',
-      pendingBadge: '下一条',
-      detailMeta: '{conversations} 个对话，{memories} 条项目记忆',
-      currentConversation: '当前对话',
-      noCurrentConversation: '当前活动标签页不是 DeepSeek 对话。',
-      addCurrentConversation: '加入当前对话',
-      moveCurrentConversation: '移入当前项目',
-      useNextConversation: '下一条新会话使用此项目',
-      cancelNextConversation: '取消下一条新会话',
-      conversationsTitle: '项目对话',
-      emptyConversations: '还没有对话属于这个项目。',
-      memoriesTitle: '项目记忆',
-      emptyMemories: '还没有项目记忆。',
-      backendUnavailable: '项目后端不可用，请重新加载扩展后再试。',
-      operationFailed: '项目操作失败：{error}',
-    },
-    savedPage: {
-      title: '保存项',
-      description: '保存常用 Prompt、回答片段和参考链接，之后可搜索、插入到对话或导出备份。',
-      count: '{count} 条保存',
-      snippet: '片段',
-      bookmark: '书签',
-      titlePlaceholder: '标题',
-      contentPlaceholder: 'Prompt 片段、笔记或可复用文本',
-      tagsPlaceholder: '标签（逗号分隔）',
-      save: '保存',
-      searchPlaceholder: '搜索保存项',
-      empty: '暂无保存项',
-      emptyHelp: '保存第一条片段后，可以在这里复用固定说明、工作流或参考材料。',
-      insertPrompt: '插入到对话',
-      exportMarkdown: '导出 Markdown',
-      exportJson: '导出 JSON',
-      saved: '已保存',
-      inserted: '已插入当前 DeepSeek 对话',
-      insertFailed: '插入到对话失败：{error}',
-      deleteConfirm: '确定删除这条保存项吗？',
-      backendUnavailable: '保存项后端不可用，请重新加载扩展后再试。',
-      operationFailed: '保存项操作失败：{error}',
-    },
-    whatsNew: {
-      title: '版本更新内容',
-      subtitle: '本次重点新增浏览器控制，并继续收紧侧边栏反馈、第三方 Skill 管理和权限说明。',
-      versionBadge: 'v{version}',
-      dismiss: '知道了',
-      items: {
-        browserControl: '浏览器控制可在侧边栏选择目标标签页，让 DeepSeek++ 执行可见网页操作。',
-        browserControlBoundary: '浏览器控制默认关闭，启用后只向新会话提供文本快照和浏览器动作工具。',
-        sidepanelFeedback: '保存项、项目和提示词控制增加更明确的成功、失败和不可用状态反馈。',
-        thirdPartySkills: '第三方 Skill 会按来源分组展示，可单独启用、停用、检查更新或同步。',
-      },
-    },
-    memoryPage: {
-      title: '记忆',
-      description: '管理会长期影响 DeepSeek++ 回复的用户偏好、反馈、话题背景和参考资料。',
-      emptyAll: '暂无记忆，对话时会自动积累',
-      emptyFiltered: '该分类下暂无记忆',
-      emptyHelp: '对话中出现稳定偏好或项目背景时，DeepSeek++ 会自动保存；也可以手动新增一条。',
-      count: '共 {count} 条记忆',
-      deleteConfirm: '确定删除这条记忆吗？',
-      backendUnavailable: '记忆存储不可用，请重新加载扩展后再试。',
-      operationFailed: '记忆操作失败：{error}',
-    },
-    presetPage: {
-      title: '系统提示词预设',
-      description: '给新对话准备固定的系统角色或行为边界；激活后会在首条消息前自动注入。',
-      import: '导入',
-      create: '新建',
-      activeMeta: '已有预设生效',
-      inactiveMeta: '未启用预设',
-      empty: '暂无预设',
-      emptyHelp: '创建系统提示词预设后，选中即可在每次新对话的第一条消息前自动注入，无需手动触发。',
-      activeHelp: '启用一个预设后，每次新对话的首条消息会自动注入该提示词。同一时间只能激活一个预设。',
-      deleteConfirm: '确定删除这个预设吗？',
-      backendUnavailable: '预设存储不可用，请重新加载扩展后再试。',
-      operationFailed: '预设操作失败：{error}',
-    },
-    skillPage: {
-      title: '可用 Skill',
-      description: 'Skill 是可复用的专家工作流。输入 /skill 名称即可把特定任务规则带入当前对话。',
-      createCustom: '自定义',
-      importLocal: '本地',
-      summary: '{total} 个 Skill，{enabled} 个启用',
-      sectionGithub: 'GitHub 导入',
-      sectionBuiltin: '内置',
-      sectionOfficial: '第三方',
-      sectionThirdParty: '第三方',
-      sectionCustom: '自定义',
-      bundledThirdPartySource: '随扩展提供，默认不启用',
-      unknownThirdPartySource: '未知第三方来源',
-      enabledSkillCount: '{enabled}/{total} 启用',
-      enableSourceSkills: '启用全部',
-      disableSourceSkills: '停用全部',
-      expandSource: '展开 {source}',
-      collapseSource: '收起 {source}',
-      usagePrefix: '在 DeepSeek 输入框中输入',
-      usageTrigger: '/skill名 参数',
-      usageSuffix: '触发。例如：',
-      usageExample: '/frontend-design 做一个登录页',
-      githubSourceTitle: 'GitHub 源',
-      localReferencedSource: '本地引用文件夹',
-      repoRoot: '仓库根目录',
-      check: '检查',
-      checking: '检查中...',
-      sync: '同步',
-      syncing: '同步中...',
-      remove: '移除',
-      skillCount: '{count} 个 Skill',
-      unknownLicense: '未知许可',
-      syncedAt: '同步 {time}',
-      checkedAt: '检查 {time}',
-      checkPermissionError: '需要 GitHub API 访问权限才能检查更新',
-      syncPermissionError: '需要 GitHub API 访问权限才能同步更新',
-      checkFailed: '检查更新失败',
-      syncFailed: '同步失败',
-      syncedSkills: '已同步 {count} 个 Skill',
-      deleteSourceConfirm: '确定移除 {repository} 导入的 {count} 个 Skill 吗？',
-      noUpdates: '上游没有发现更新',
-      changedUpdates: '{count} 个已导入 Skill 可能有更新',
-      newSkills: '{count} 个新增 Skill',
-      missingSkills: '{count} 个已导入 Skill 在上游消失',
-      updatesFound: '发现上游更新',
-      backendUnavailable: 'Skill 存储不可用，请重新加载扩展后再试。',
-      operationFailed: 'Skill 操作失败：{error}',
-      relocatePrompt: '原文件夹已挪动或不存在，请重新选择本地 Skill 文件夹。',
-    },
-    settings: {
-      title: '设置',
-      description: '调整模型模式、侧边栏对话、提示词注入、外观、同步和本地数据管理。',
-      tabs: {
-        general: '通用',
-        api: 'API',
-        prompt: '提示词',
-        voice: '语音',
-        appearance: '外观',
-        usage: '用量',
-        data: '数据',
-        projectFiles: '项目目录',
-        about: '关于',
-      },
-      navLabel: '设置子导航',
-      generalDescription: '界面语言与网页行为开关。',
-      apiDescription: '侧边栏对话、多模态分析所需的 API Key 与请求地址。',
-      promptDescription: '控制记忆、系统提示词与预设的注入策略。',
-      voiceDescription: '语音输入与回复朗读。',
-      appearanceDescription: '背景、悬浮宠物等 DeepSeek 页面外观。',
-      usageDescription: '查看近期模型 Token、活跃度与模型分布。',
-      dataDescription: '云同步与本地记忆的导入、导出和清除。',
-      projectFilesDescription: '授权本地项目目录，供对话中 @ 引用文件。',
-      projectFilesSection: '受信任目录',
-      projectFilesSectionDescription: '目录只在本机扫描，本地路径不会上传；文件内容仅在发送时按需读取。',
-      projectFilesChoose: '选择目录',
-      projectFilesChooseAgain: '重新选择',
-      projectFilesRemove: '移除授权',
-      projectFilesRemoveConfirm: '移除后将无法在对话中 @ 引用这些文件，确定移除吗？',
-      projectFilesAuthorized: '已授权',
-      projectFilesSessionExpired: '授权已失效，请重新选择目录',
-      projectFilesFiles: '{count} 个文件',
-      projectFilesSkipped: '已忽略 {count} 项',
-      projectFilesTruncated: '超出索引上限，仅保留前 {count} 个文件',
-      projectFilesImagesReady: '图片 {count} 个，可在对话中 @ 引用',
-      projectFilesTextPending: '文本 {count} 个，内容引用将在后续版本支持',
-      projectFilesOther: '其他 {count} 个',
-      projectFilesPickFailed: '目录读取失败，请重试。',
-      interfaceSection: '界面',
-      interfaceLanguage: '界面语言',
-      interfaceLanguageDescription: '默认跟随 Chrome 或浏览器语言，也可以手动固定。',
-      languageAuto: '跟随',
-      languageChinese: '简体中文',
-      languageEnglish: 'English',
-      languageCurrent: '当前：{language}',
-      apiKeyRequired: '请输入 DeepSeek API Key',
-      saveFailed: '保存失败',
-      clearFailed: '清除失败',
-      apiKeySaved: '已保存，右键场景可在普通网页使用',
-      apiKeyCleared: '已清除，右键场景恢复为仅 DeepSeek 网页可用',
-      webDavPermissionDenied: '需要访问权限才能连接 WebDAV 服务器',
-      operationFailed: '操作失败',
-      syncConfigChanged: '同步设置已变化，请确认当前目标后重试',
-      syncRemoteCommittedWarning: '远端操作已经完成，但本地同步记录未能确认。请勿立即重试；已重新载入权威同步配置。',
-      syncCounts: '记忆 {memories} 条，Skill {skills} 个，预设 {presets} 个，项目 {projects} 个，项目对话 {projectConversations} 个，保存项 {savedItems} 条',
-      connectionSuccess: '连接成功',
-      connectionFailed: '连接失败',
-      uploadConfirm: '确定要用本地记忆、Skill、预设、项目和保存项覆盖云端数据吗？',
-      uploadSuccess: '上传完成，已覆盖云端。{counts}',
-      uploadFailed: '上传失败',
-      downloadConfirm: '确定要用云端记忆、Skill、预设、项目和保存项覆盖本地数据吗？此操作不可撤销。',
-      downloadSuccess: '下载完成，已覆盖本地。{counts}',
-      downloadFailed: '下载失败',
-      importMemoryArrayError: '导入文件必须是记忆数组',
-      importSuccess: '已导入 {count} 条记忆',
-      jsonFormatError: 'JSON 格式错误',
-      clearAllConfirm: '确定要清除所有记忆吗？此操作不可撤销。',
-      neverSynced: '从未同步',
-      modelSection: '对话设置',
-      modelMode: '网页模型模式',
-      modelModeDescription: '选择 chat.deepseek.com 网页对话使用的官方模式；侧边栏 API 对话仍在对话页选择模型。',
-      modelDefault: '默认',
-      modelExpert: '专家',
-      modelVision: '识图',
-      sidepanelChat: '侧边栏对话',
-      sidepanelChatDescription: '在侧边栏显示对话标签，支持网页登录或官方 API Key',
-      apiKeyDescription: '配置后右键场景可在普通网页使用',
-      configured: '已配置',
-      notConfigured: '未配置',
-      apiKeyReplacePlaceholder: '输入新 Key 可替换',
-      multimodalApi: '多模态 API',
-      multimodalApiDescription: '为内置多模态 MCP 配置 OpenAI 图片分析、Gemini 视频分析和自定义请求地址。',
-      openaiKeyReplacePlaceholder: '输入新 OpenAI Key 可替换',
-      geminiKeyReplacePlaceholder: '输入新 Gemini Key 可替换',
-      openaiImageModel: 'OpenAI 图片模型',
-      geminiVideoModel: 'Gemini 视频模型',
-      openaiBaseUrl: 'OpenAI 请求地址',
-      geminiBaseUrl: 'Gemini 请求地址',
-      multimodalBaseUrlInvalid: 'OpenAI / Gemini 请求地址必须是有效的 http(s) URL',
-      multimodalSaved: '多模态 API 配置已保存',
-      multimodalCleared: '多模态 API 配置已清除',
-      clearMultimodalApi: '清除多模态配置',
-      saving: '保存中',
-      clearing: '清除中',
-      clearApiKey: '清除 API Key',
-      petWhale: 'DeepSeek 小鲸鱼',
-      petWhaleDescription: '在 DeepSeek 页面显示状态联动宠物',
-      backgroundSection: '背景设置',
-      customBackground: '自定义背景',
-      customBackgroundDescription: '为 DeepSeek 页面设置背景图片',
-      uploadImage: '上传图片',
-      imageUrlPlaceholder: '粘贴图片 URL',
-      backgroundPreviewAlt: '背景预览',
-      backgroundPreviewOverlay: '模拟效果预览',
-      backgroundOpacity: '背景透明度',
-      clearBackground: '清除背景',
-      floatingPetSection: '悬浮宠物',
-      positionBottomRight: '右下',
-      positionBottomLeft: '左下',
-      positionCustom: '自定义',
-      size: '尺寸',
-      opacity: '透明度',
-      petMotion: '动态漂浮',
-      petMotionDescription: '减少动作时可关闭',
-      floatingChatSection: '全局悬浮球',
-      floatingChat: '悬浮聊天球',
-      floatingChatDescription: '在所有网页显示可拖动的悬浮球。点击打开聊天窗口，长按拖动位置。',
-      floatingChatPermissionMissing: '尚未获得网页访问权限，悬浮聊天球仍保持关闭。再次开启可重新授权。',
-      floatingChatContextInvalidated: '扩展上下文已失效，请重新加载扩展后再设置悬浮聊天球。',
-      cloudSyncSection: '云同步',
-      syncProvider: '同步方式',
-      providerWebdav: 'WebDAV',
-      providerGdrive: 'Google Drive',
-      providerOnedrive: 'OneDrive',
-      webDavUrl: 'WebDAV 地址',
-      username: '用户名',
-      password: '密码',
-      remotePath: '远程路径',
-      clientId: 'Client ID',
-      clientSecret: 'Client Secret',
-      redirectUri: '回调地址',
-      redirectUriUnavailable: '当前浏览器环境不支持 OAuth 回调地址',
-      redirectUriHint: '请将此地址添加到 OAuth 应用的授权回调地址中。',
-      authorize: '授权登录',
-      authorized: '已授权',
-      reauthorize: '重新授权',
-      authorizeSuccess: '授权成功',
-      oauthSetupHint: '在 {link} 创建 OAuth 应用，并将 Client ID 和 Secret 填入此处。',
-      gdriveConsole: 'Google Cloud Console',
-      onedriveConsole: 'Azure 门户',
-      testConnection: '测试连接',
-      uploadLocal: '上传本地',
-      downloadRemote: '下载云端',
-      lastSync: '上次同步: {time}',
-      dataSection: '数据管理',
-      memoryTotal: '记忆总数',
-      exportMemories: '导出记忆',
-      exportLogs: '导出诊断日志',
-      exportLogsHint: '导出最近一次会话内后台工具执行与错误诊断信息，用于排查 shell_exec / MCP / 授权问题。',
-      importMemories: '导入记忆',
-      clearAllMemories: '清除所有记忆',
-      usage: {
-        rangeLabel: '时间范围',
-        last7Days: '最近 7 天',
-        last30Days: '最近 30 天',
-        loadFailed: '使用统计加载失败',
-        clearStats: '清除统计',
-        clearConfirm: '确定要清除本地使用统计吗？此操作不影响记忆、对话和云同步数据。',
-        clearSuccess: '使用统计已清除',
-        clearFailed: '使用统计清除失败',
-        emptyTitle: '暂无使用统计',
-        emptyDescription: '完成下一次 DeepSeek 网页回复后，这里会开始记录 Token、TPS、活跃日期和模型分布。',
-        totalTokens: 'Tokens 用量',
-        serverSamples: '服务端样本 {server}/{total}',
-        sessions: '会话数量',
-        turns: '回复 {count} 次',
-        messages: '消息数量',
-        activeDays: '活跃 {count} 天',
-        currentStreak: '当前连续天数',
-        daysUnit: '天',
-        mostUsedModel: '最常用模型',
-        noModel: '暂无模型',
-        share: '占比 {percent}',
-        heatmapTitle: '活跃热力图',
-        less: '较少',
-        more: '较多',
-        dailyTrendTitle: '按天 Token 趋势',
-        modelUsageTitle: '模型用量',
-      },
-      aboutSection: '关于',
-      aboutTagline: 'Agentic 记忆与 Skill 系统',
-    },
-    promptControls: {
-      title: '提示词控制',
-      memory: '记忆注入',
-      memoryDescription: '在新请求中加入筛选后的长期记忆',
-      systemPrompt: 'DeepSeek++ 系统提示词',
-      systemPromptDescription: '注入工具 schema、记忆规则和运行时指引',
-      presetCadence: '预设注入频率',
-      cadenceDefault: '默认频率',
-      cadenceFirst: '仅首条消息',
-      cadenceEvery: '每条消息',
-      cadenceOff: '关闭',
-      forceLanguage: '强制回复语言',
-      languageAuto: '自动',
-      languageZh: '简体中文',
-      languageEn: '英文',
-      backendUnavailable: '提示词设置后端不可用，请重新加载扩展后再试。',
-      loadFailed: '读取提示词设置失败：{error}',
-      saveFailed: '保存提示词设置失败：{error}',
-    },
-    voice: {
-      title: '语音',
-      input: '语音输入',
-      inputDescription: '在侧边栏对话中使用浏览器语音识别',
-      inputUnsupported: '当前浏览器环境不支持语音识别',
-      readAloud: '朗读回复',
-      readAloudDescription: '用浏览器语音合成朗读侧边栏回复',
-      readAloudUnsupported: '当前浏览器环境不支持语音合成',
-      rate: '语速',
-      pitch: '音调',
-    },
-    capabilitiesPage: {
-      navLabel: '能力子导航',
-      tabs: {
-        skill: 'Skill',
-        mcp: 'MCP',
-        tools: '工具',
-        browser: '浏览器',
-        preset: '预设',
-        automation: '自动化',
-      },
-    },
-    browserControlPage: {
-      title: '浏览器控制',
-      description: '让 DeepSeek++ 通过浏览器 Debugger 和 Accessibility Tree 控制选中的 Chrome 或 Edge 标签页。',
-      enableTitle: '启用浏览器工具',
-      enableDescription: '启用后，新对话会注入 browser_* 工具。模型接收文本快照，不需要截图输入。',
-      unsupported: '浏览器控制需要 Chromium debugger 和 tabs API，当前平台不可用。',
-      detach: '断开附着',
-      detaching: '断开中...',
-      targetsTitle: '目标标签页',
-      noTargets: '没有找到浏览器标签页。',
-      untitled: '未命名',
-      noUrl: '无 URL',
-      group: '分组：{name}',
-      selected: '已选择',
-      snapshotTitle: '快照预算',
-      includeSnapshot: '动作后返回快照',
-      includeSnapshotDescription: '点击、填写、导航和键盘动作后返回新的 Accessibility Tree 快照。',
-      maxNodes: '最大 AX 节点数',
-      maxBytes: '最大快照文本字节数',
-      status: {
-        enabled: '启用',
-        attached: '附着',
-        target: '目标',
-      },
-      messages: {
-        enabled: '浏览器工具已启用。新对话可以接收 browser_* 工具 schema。',
-        disabled: '浏览器工具已停用。新对话不会注入 browser 工具 schema。',
-        detached: '已从当前目标断开。',
-        targetFailed: '选择目标标签页失败。',
-        targetSelected: '已选择标签页 {id}。',
-      },
-    },
-    mcpPage: {
-      title: 'MCP',
-      description: '连接本机或远程 MCP 服务，让 DeepSeek++ 可以发现工具、控制自动执行策略并回传结果。',
-      summary: '{servers} 个服务，{enabled} 个启用，{tools} 个自动工具',
-      loading: '正在加载 MCP 配置',
-      empty: '暂无 MCP 服务',
-      emptyHint: '连接远程 MCP 服务，或一键安装本机 Shell / 多模态预设。',
-      emptyCreateAction: '新增 MCP 服务',
-      emptyInstallShell: '安装 Shell 预设',
-      shell: 'Shell',
-      multimodal: '多模态',
-      addServer: '新增',
-      enabled: '启用',
-      selected: '已选择',
-      auto: '自动',
-      disabled: '禁用',
-      success: '成功',
-      failure: '失败',
-      localHost: '本地宿主',
-      messages: {
-        loadFailed: '加载 MCP 配置失败',
-        shellExistsSelected: 'Shell MCP 已存在，已选中现有配置',
-        shellCreateFailed: '创建 Shell MCP 预设失败',
-        shellCreated: '已创建 Shell MCP 预设。请运行下方安装命令后重启浏览器。',
-        multimodalExistsSelected: '多模态 MCP 已存在，已选中现有配置',
-        multimodalCreateFailed: '创建多模态 MCP 预设失败',
-        multimodalCreated: '已创建多模态 MCP 预设。请运行下方安装命令，并在设置页配置多模态 API。',
-        saveFailed: '保存 MCP 服务失败',
-        deleteConfirm: '删除 MCP 服务「{name}」？',
-        permissionGranted: '已授权 {origin}',
-        permissionDenied: '授权被拒绝',
-        permissionRequired: '需要授权 {origin}',
-        connectionSuccess: '连接成功，{tools} 个工具，{latency}',
-        connectionFailed: '连接失败',
-        nativeMessagingUnsupported: '当前平台不支持 Native Messaging。',
-      },
-      transportHints: {
-        streamableHttp: '推荐，兼容新版 MCP HTTP 服务',
-        http: 'JSON-RPC over HTTP POST',
-        sse: '旧版 MCP SSE 传输',
-        stdioBridge: '连接已暴露为 MCP HTTP 的本地 stdio 桥接服务；进程启动由桥接服务负责',
-        nativeMessaging: '通过 Browser Native Messaging Host 访问本机能力',
-      },
-      form: {
-        createTitle: '新增 MCP 服务',
-        editTitle: '编辑 MCP 服务',
-        basic: '基础',
-        stdioSection: '桥接进程信息',
-        advanced: '高级设置',
-        advancedHint: 'Headers、密钥、超时、限制与执行策略',
-        name: '名称',
-        transport: '传输',
-        serviceUrl: '服务 URL',
-        bridgeEndpointUrl: '桥接端点 URL',
-        bridgeEndpointHint: '这里填写 stdio 桥接服务暴露的 HTTP MCP 地址，不是本地 exe 路径；本地 Shell/stdio 能力请优先使用 Shell Native Host 预设。',
-        command: '命令',
-        args: '参数',
-        cwd: '工作目录',
-        env: '环境变量',
-        connectMs: '连接 ms',
-        requestMs: '请求 ms',
-        discoveryMs: '发现 ms',
-        resultBytes: '结果字节',
-        toolLimit: '工具上限',
-        defaultExecution: '默认执行',
-        allowInject: '允许注入',
-        modeAuto: '自动执行',
-        modeManual: '手动策略',
-        modeDisabled: '禁用',
-        save: '保存',
-      },
-      headers: {
-        headerName: 'Header',
-        headerValue: 'Value',
-        secretValue: 'Secret value',
-        secretHeaderName: 'Header name',
-      },
-      validation: {
-        nameRequired: '名称不能为空',
-        positiveInteger: '{label} 必须是正整数',
-        nativeHostRequired: 'Native Host 不能为空',
-        nativeHostInvalid: 'Native Host 只能包含字母、数字、点、下划线和短横线',
-        serviceUrlRequired: '服务 URL 不能为空',
-        bridgeEndpointRequired: 'Stdio Bridge 需要填写桥接端点 URL',
-        serviceUrlUnsupported: '服务 URL 只支持 http/https',
-        serviceUrlInvalid: '服务 URL 格式无效',
-        stdioCommandRequired: 'Stdio Bridge 命令不能为空',
-        headerInvalidName: 'Header 名称无效：{name}',
-        emptyValue: '(空)',
-        headerInvalidValue: 'Header 值不能包含换行：{name}',
-        headerSecretRequired: 'Header Secret 需要有效 Header 名称',
-        envInvalid: '环境变量格式无效：{line}',
-      },
-      row: {
-        autoTools: '{active}/{total} 自动',
-        testing: '测试中',
-        refreshTools: '刷新工具',
-      },
-      detail: {
-        grant: '授权',
-        status: '状态',
-        latency: '延迟',
-        lastConnected: '上次连接',
-        transport: '传输',
-        executionPolicy: '自动执行策略',
-        injectionSummary: '当前注入 {count} 个工具；禁用或手动策略不会进入 DeepSeek Prompt。',
-        promptExposure: 'Prompt 暴露策略',
-        promptExposureDescription: '控制这个服务的工具 Schema 如何进入新请求；不会改变服务权限或执行策略。',
-        exposureDirect: '直接：展示全部工具',
-        exposureAdaptive: '自适应：展示相关工具',
-        exposureOnDemand: '按需：仅能力目录',
-        exposureAdaptiveHint: '选择直接展示的 Schema 时，会优先保留 {count} 个置顶工具。',
-        exposureOnDemandHint: '工具 Schema 不进入 Prompt，模型需先发现并查看能力后再调用。',
-        pin: '置顶',
-        unpin: '取消置顶',
-        adaptiveBudget: '自适应 Prompt 预算',
-        adaptiveBudgetDescription: '所有自适应 MCP 服务共享的直接展示 Schema 上限。',
-        adaptiveMaxTools: '直接 Schema 数',
-        adaptiveMaxPromptBytes: 'Schema 字节数',
-        adaptiveBudgetApply: '应用',
-        adaptiveBudgetInvalid: 'Schema 数须为 1–64，字节预算须为 2,000–256,000。',
-        discoveredTools: '发现工具',
-        noTools: '尚未发现工具',
-        recentCalls: '最近调用',
-        noHistory: '暂无调用记录',
-        schemaNone: '参数：无',
-        schemaSummary: '参数：{props}{required}',
-        schemaRequired: '；必填 {required}',
-        toggleTools: '工具列表',
-        toggleHistory: '调用记录',
-        hintExpand: '展开安装指引',
-        hintCollapse: '收起安装指引',
-      },
-      status: {
-        ready: 'ready',
-        error: 'error',
-        disabled: 'disabled',
-        unknown: 'unknown',
-      },
-      endpoint: {
-        nativeMessaging: 'Native Messaging',
-        bridgeUrl: 'Bridge URL',
-        command: 'command',
-      },
-      shellSetup: {
-        localIntro: '打开终端，在项目根目录执行以下命令：',
-        publishedIntro: '打开终端，执行以下命令（只需一次）：',
-        fallbackIntro: '如果你使用的是已发布扩展而不是本地源码版，执行：',
-        detectedExtensionId: '已自动检测 {browser} 扩展 ID。安装需要本机已安装 Node.js/npm。',
-        firefoxFixedId: 'Firefox 使用固定扩展 ID，不需要额外填写 extension id。安装需要本机已安装 Node.js/npm。',
-        installNote: '命令会安装或更新 Shell Native Host；默认跳过 OfficeCLI，如需 OfficeCLI 可去掉 --skip-officecli。',
-        enableAndTest: '安装完成后，打开上方开关启用此服务，再点击「测试」验证连接。',
-        restartAndTest: '安装完成后重启浏览器，回到这里点击「测试」验证连接。',
-        forbidden: 'Native Host 已安装，但未授权当前扩展 ID。请重新运行下方安装命令后重启浏览器。',
-        notFound: '未找到 Native Host — 请先运行下方安装命令，并确保已安装 Node.js/npm。',
-        unavailable: '当前浏览器不支持 Native Messaging，请使用 Chrome、Edge 或 Firefox。',
-        cannotConnect: '无法连接到 Native Host — 请确认已运行安装脚本并重启浏览器。',
-        ready: '已连接，发现 {count} 个工具。',
-        disabled: '服务已创建但尚未启用。请先安装 Native Host，再启用并测试。',
-        installFirst: '请先安装 Native Host，再点击「测试」验证连接。',
-      },
-      multimodalSetup: {
-        localIntro: '打开终端，在 deepseek-pp 项目根目录执行以下命令：',
-        publishedIntro: '打开终端，执行以下命令（只需一次）：',
-        fallbackIntro: '如果你使用的是已发布扩展而不是本地源码版，执行：',
-        settingsNote: 'API Key、模型和请求地址请在设置页的「多模态 API」中配置；图片使用 OpenAI，视频使用 Gemini。',
-        enableAndTest: '安装完成后，在设置页配置多模态 API，打开上方开关启用，再点击「测试」验证连接。',
-        forbidden: 'Multimodal Native Host 已安装，但未授权当前扩展 ID。请重新运行下方安装命令后重启浏览器。',
-        notFound: '未找到 Multimodal Native Host — 请先运行下方安装命令，并确保已安装 Node.js/npm。',
-        cannotConnect: '无法连接到 Multimodal Native Host — 请确认已运行安装脚本并重启浏览器。',
-        ready: '已连接，发现 {count} 个多模态工具。',
-        disabled: '服务已创建但尚未启用。请先安装 Native Host、配置多模态 API，再启用并测试。',
-        installFirst: '请先安装 Multimodal Native Host，在设置页配置多模态 API 后点击「测试」验证连接。',
-      },
-    },
-    chatPage: {
-      authRequired: '请配置 DeepSeek API Key，或先登录 chat.deepseek.com',
-      authHint: '未配置 Key 时，侧边栏对话依赖 DeepSeek 网页登录态',
-      title: '对话',
-      description: '直接在侧边栏发起任务，可复用记忆、预设、Skill 和已启用工具。',
-      apiDescription: '使用官方 API，可选择模型和思考强度。',
-      webDescription: '使用网页登录态，可复用记忆、预设、Skill 和已启用工具。',
-      apiProvider: 'API',
-      webProvider: 'Web',
-      newSessionTitle: '新建会话',
-      newSession: '新建',
-      newSessionConfirm: '当前对话内容将被清空，确定新建会话吗？',
-      empty: '输入消息开始对话',
-      emptyHelp: '也可以从网页右键发送选中文本，或从保存项插入常用 Prompt。',
-      inputPlaceholder: '给 DeepSeek++ 发送消息',
-      send: '发送',
-      sendFailed: '发送失败',
-      uploadImage: '上传图片',
-      imageAttachments: '图片附件',
-      imageUploading: '上传中...',
-      imageReady: '已添加',
-      imageUploadFailed: '上传失败',
-      imageUploadMissingId: 'DeepSeek 上传成功但没有返回文件 ID。',
-      imageUploadMax: '每轮最多上传 {count} 张图片。',
-      imageOnly: '只支持上传图片文件。',
-      imageEmpty: '图片文件为空。',
-      imageTooLarge: '图片超过 {limit} 上限。',
-      removeImageAttachment: '移除图片 {name}',
-      atPanelTitle: '引用文件',
-      atPanelHint: '输入 @ 后按文件名或路径搜索，选择图片即可引用',
-      atPanelFileCount: '{count} 个文件',
-      atPanelNoDirectory: '尚未授权项目目录：设置 → 项目目录',
-      atPanelNoMatch: '没有匹配的文件',
-      atPanelEmptyDirectory: '目录中没有可引用的文件',
-      atPanelTextPending: '文本文件引用将在后续版本支持',
-      atPanelOtherUnsupported: '该文件类型暂不支持引用',
-      atPanelSelectFile: '选择 {name}',
-      atPanelDeselectFile: '取消选择 {name}',
-      atPanelClose: '关闭引用面板',
-      modelLabel: '模型',
-      modelFlash: 'Flash',
-      modelPro: 'Pro',
-      thinkingLabel: '思考',
-      thinkingOff: '即时',
-      thinkingOn: '深度思考',
-      effortLabel: '思考强度',
-      effortHigh: '标准',
-      effortMax: '最强',
-      reasoningTitle: '思考过程',
-      reasoningActive: '正在思考',
-      read: '朗读',
-      readLatest: '朗读最新助手回复',
-      voiceInput: '语音输入',
-      stopListening: '停止收听',
-      mic: '麦克风',
-      stop: '停止',
-    },
-    toolsPage: {
-      diagnosticsDefaultQuery: '橘鸦 up主',
-      diagnosticsRunning: '诊断中...',
-      diagnosticsRun: '诊断',
-      bytes: '{count} 字节',
-      errorPrefix: '错误: {error}',
-      toolTitle: '工具开关',
-      toolDescription: '关闭后该工具不会注入到对话中，AI 将无法调用',
-      webSearchName: '搜索互联网 (web_search)',
-      webSearchDescription: '在 Bing 搜索关键词，返回标题、URL 和摘要',
-      webFetchName: '获取网页 (web_fetch)',
-      webFetchDescription: '下载指定 URL 并提取可视文本内容',
-      pythonTitle: 'Python 解释器 (python_exec)',
-      pythonStatusNoShell: '未创建 Shell Native Host',
-      pythonStatusUnsupported: '当前平台不支持 Shell Native Host',
-      pythonStatusNoCache: '尚未刷新工具',
-      pythonStatusEnabled: '已开启',
-      pythonStatusDiscovered: '已发现，未开启',
-      pythonStatusMissing: '未发现 python_exec',
-      pythonDescription: '用本机 Python 执行短代码，适合快速验证想法、复杂计算和小型数据处理。',
-      pythonCreate: '创建 Shell',
-      pythonCreating: '创建中',
-      pythonRefresh: '刷新工具',
-      pythonRefreshing: '刷新中',
-      pythonStatusAvailable: 'python_status 可用',
-      pythonMissingDetail: '当前 Shell Native Host 版本没有返回 python_exec。重装 Shell Native Host 并重启浏览器后再刷新。',
-      shellExists: 'Shell MCP 已存在，请刷新工具或重装 Native Host。',
-      shellCreated: '已创建 Shell MCP。安装或重装 Shell Native Host 后点击刷新工具。',
-      pythonFound: '已发现 Python 解释器工具。',
-      pythonMissingAfterRefresh: '未发现 python_exec。请重装 Shell Native Host 后重启浏览器，再刷新工具。',
-      pythonMissingBeforeToggle: '未发现 python_exec。请先刷新工具，必要时重装 Shell Native Host。',
-      pythonEnabled: 'Python 解释器已开启。',
-      pythonDisabled: 'Python 解释器已关闭。',
-      disabledNotice: '关闭工具后，新对话将不再包含该工具的调用格式。已开启的对话不受影响。',
-      diagnosticTitle: '诊断搜索',
-      diagnosticDescription: '直接测试搜索是否可用，绕过 AI 对话链路',
-      permissionTitle: 'web_fetch 权限',
-      permissionDescription: '获取网页需要访问对应站点的权限。在此输入网址并授予权限。',
-      grantPermission: '授权',
-      permissionGranted: '权限已授予，可以访问该站点',
-      permissionDenied: '权限被拒绝，请重试或前往 chrome://extensions 手动添加',
-      permissionInvalidUrl: '网址格式不正确，请输入完整 URL（如 https://example.com）',
-      allSitesRequesting: '请求中...',
-      allSitesGranted: '已授权全部网站',
-      allSitesGrant: '授权全部网站',
-      allSitesHelp: '一键授予扩展访问所有网站的权限，此后 web_fetch 获取任意页面不再弹窗',
-    },
-    automationPage: {
-      title: '自动化',
-      description: '把固定任务交给独立 DeepSeek 会话运行，可手动触发或按 Cron / RRULE 定时执行。',
-      summary: '{total} 个任务，{active} 个启用',
-      create: '新建',
-      created: '自动化已创建',
-      loadFailed: '自动化数据加载失败',
-      empty: '暂无自动化',
-      emptyHelp: '新建任务后，可以让它定时跟踪信息、复盘内容或重复执行固定 Prompt。',
-      namePromptRequired: '名称和 Prompt 不能为空',
-      expressionRequired: '请填写定时表达式',
-      runFailed: '运行失败',
-      deleteConfirm: '删除自动化「{name}」？',
-      form: {
-        name: '名称',
-        namePlaceholder: '任务名称',
-        model: '模型',
-        defaultModel: '默认',
-        promptPlaceholder: '输入要定时发送到 DeepSeek 的内容',
-        trigger: '触发',
-        manual: '手动',
-        expression: '表达式',
-        timezone: '时区',
-        search: '联网',
-        thinking: '深度思考',
-        create: '创建',
-      },
-      status: {
-        active: '启用',
-        paused: '暂停',
-        queued: '排队中',
-        running: '运行中',
-        succeeded: '成功',
-        failed: '失败',
-        timeout: '超时',
-        cancelled: '已取消',
-        skipped: '已跳过',
-      },
-      meta: {
-        next: '下次',
-        previous: '上次',
-        session: '会话',
-        recent: '最近',
-        notCreated: '未创建',
-        none: '暂无',
-      },
-      actions: {
-        openSession: '打开会话',
-        runNow: '立即运行',
-      },
-      attempt: '{count}次',
-    },
-  },
-  content: {
-    conversation: {
-      untitled: '未命名对话',
-    },
-    export: {
-      buttonIdle: '导出当前对话',
-      buttonRunning: '正在导出当前对话',
-      emptyConversation: '当前页面还没有可导出的对话。',
-      formatDialogLabel: '选择导出格式',
-      formatTitle: '导出格式',
-      scopeLabel: '内容范围',
-      scopeFull: '全部内容（含工具调用、续跑、思考、注入信息）',
-      scopeInputOutput: '仅输入 + 最终产出',
-      formatHtml: 'HTML',
-      formatMarkdown: 'Markdown',
-      formatPdf: 'PDF',
-      formatImageManifest: '图片清单',
-      submit: '导出',
-      progress: '正在导出当前对话...',
-      failed: '导出失败',
-      cancelled: '导出已取消',
-      partialSuccess: '导出完成，但部分会话读取失败，请检查导出文件中的 Export Warnings。',
-      success: '当前对话已导出。',
-    },
-    historyOrganizer: {
-      enhancedSearchTitle: 'DeepSeek++ 搜索增强',
-      tagFilterLabel: '按标签筛选结果',
-      tagPlaceholder: '输入标签名',
-      currentTagsLabel: '给当前对话加标签',
-      currentTagsPlaceholder: '逗号分隔，例如：港股, 写作',
-      emptySearchStatus: 'DeepSeek++：等待官方搜索结果',
-      visibleStatus: 'DeepSeek++：已显示 {visible}/{total}',
-      loadError: 'DeepSeek++：历史标签加载失败：{message}',
-      saveError: 'DeepSeek++：历史标签保存失败：{message}',
-    },
-    projectSidebar: {
-      title: '项目',
-      empty: '还没有项目',
-      expandProject: '展开项目：{name}',
-      collapseProject: '收起项目：{name}',
-      showMore: '展开显示',
-      showLess: '收起显示',
-      moveCurrentToProject: '移入当前对话到 {name}',
-      removeCurrentFromProject: '从 {name} 移出当前对话',
-      joinProject: '加入项目',
-      joinProjectNamed: '加入项目：{name}',
-      moveToProjectNamed: '移动到项目：{name}',
-      currentProjectNamed: '已在项目：{name}',
-      removeFromProjectNamed: '移除项目：{name}',
-      conversationActions: '会话操作',
-      newConversationInProject: '在 {name} 下开启新对话',
-      useNextConversation: '下一条新会话使用 {name}',
-      cancelNextConversation: '取消下一条新会话使用 {name}',
-      pendingNextConversation: '下一条新会话将使用此项目',
-      operationFailed: '项目操作失败：{message}',
-      age: {
-        justNow: '刚刚',
-        minutesAgo: '{count}分钟前',
-        hoursAgo: '{count}小时前',
-        daysAgo: '{count}天前',
-      },
-    },
-    uxPolish: {
-      downloadCode: '下载',
-      downloadMessageMarkdownButton: 'MD',
-      downloadMessageMarkdownTitle: '下载消息为 Markdown',
-      copyMessageButton: '复制',
-      copyMessageTitle: '复制完整对话输出',
-    },
-    toolBlock: {
-      title: '已执行工具（{count}次）',
-      pythonInterpreter: 'Python 解释器',
-      summaries: {
-        saved: '已保存',
-        updated: '已更新',
-        deleted: '已删除',
-        searched: '已搜索',
-        fetched: '已获取',
-        running: '执行中',
-        executed: '已执行',
-        failed: '执行失败',
-        messageFailed: '工具消息发送失败',
-        backgroundFailed: '后台工具执行失败',
-      },
-      missingResultDetail: '后台没有返回工具执行结果。请刷新当前 DeepSeek 页面后重试。',
-      missingSandboxResultDetail: '沙箱工具没有返回执行结果。请重载 DeepSeek++ 扩展并刷新当前 DeepSeek 页面后重试；它不依赖 MCP 或 Shell Local。',
-      missingMcpResultDetail: '后台没有返回 MCP 工具执行结果。请刷新当前 DeepSeek 页面后重试；如果仍失败，在 MCP 页重新测试 Shell Local。',
-      invalidResultDetail: '后台返回的工具结果结构无效：{preview}',
-    },
-    agent: {
-      stop: '停止',
-      stopped: '已停止',
-      step: 'Step {index}',
-      streaming: 'streaming...',
-      complete: '完成',
-      completeWithTools: '完成（{count} 个工具）',
-      executingTools: '执行工具中',
-      budgetReached: 'DeepSeek++ 已执行 {count} 轮自动工具续跑并暂停，避免把未完成任务伪装成最终答案。请发送“继续”让当前对话接着处理剩余工作。',
-      error: '执行出错',
-      footerComplete: 'Agent 完成（{steps} 步，{tools} 次工具调用）',
-      process: '过程',
-      tools: '工具调用',
-      results: '执行结果',
-      running: 'Agent 运行中 · 第 {step} 步 · 已调用 {tools} 次工具',
-      footerError: 'Agent 执行出错（{steps} 步，{tools} 次工具调用）',
-      autoSaved: '长文产出已自动保存为产物：',
-      autoSavedDownload: '下载',
-      concurrencyGuard: 'Agent 正在执行中，已忽略此消息触发的重复任务。如需中断，请点击 Agent 面板的「停止」。',
-    },
-    permission: {
-      webFetch: 'DeepSeek++ 需要访问 {origin} 的权限以获取该页面内容',
-      deny: '拒绝',
-      grant: '授权',
-      requesting: '请求中...',
-    },
-    skillPopup: {
-      hint: '↑↓ 导航 · Enter 选择 · Esc 关闭',
-    },
-    tokenSpeed: {
-      title: '累计 Token：{tokens}；Token 输出速度：{speed}{idle}。Token 来源：{tokenSource}；速度来源：{speedSource}',
-      idleSuffix: '（空闲）',
-      sourceServer: '服务端',
-      sourceEstimated: '本地估算',
-    },
-    multimodalMedia: {
-      buttonTitle: '添加 DeepSeek++ 多模态图片或视频',
-      removeTitle: '移除 {name}',
-      pastedFileName: '粘贴的媒体',
-      added: '已添加 {count} 个多模态附件',
-      analyzing: '正在分析 {count} 个多模态附件...',
-      analyzed: '多模态分析完成（{count} 项）',
-      failed: '多模态分析失败：{message}',
-      tooMany: '每轮最多添加 {count} 个多模态附件',
-      unsupported: '{name} 不是支持的图片或视频',
-      imageTooLarge: '{name} 超过图片上限 {limit}',
-      videoTooLarge: '{name} 超过内联视频上限 {limit}。请改用公开视频 URL，或等待本地路径选择能力。',
-      invalidRequest: 'DeepSeek 请求体异常，无法附加多模态媒体。',
-      emptyPrompt: '请先输入消息，再发送多模态媒体。',
-    },
-    extensionReloaded: '扩展已重新加载，请刷新当前 DeepSeek 页面后重试。',
-  },
-  background: {
-    contextMenus: {
-      sendToChat: '发送到对话',
-    },
-    auth: {
-      missingDeepSeek: '请先在 chat.deepseek.com 登录，或刷新 DeepSeek 页面后重试。',
-    },
-    export: {
-      generating: '生成导出文件',
-      cancelled: '导出已取消',
-    },
-    sync: {
-      missingWebDav: '未配置 WebDAV',
-      missingSync: '未配置云同步',
-      configChanged: '其他扩展窗口已修改同步设置，请确认当前目标后重试',
-      configCommitIndeterminate: '无法确认同步设置是否已保存，将重新读取当前设置',
-      missingRemoteFile: '云端缺少 {file}，已停止下载以避免覆盖本地数据',
-      authorizationNotRequired: '当前同步方式不需要授权',
-      identityUnavailable: '当前浏览器环境不支持 OAuth 授权',
-      oauthCancelled: '授权已取消',
-      oauthProviderError: '授权失败：{error}',
-      oauthMissingCode: '授权失败：未返回授权码',
-      oauthRefreshFailed: '刷新访问令牌失败 (HTTP {status}){detail}',
-      oauthTokenExchangeFailed: '获取令牌失败 (HTTP {status}){detail}',
-      gdriveMissingRefreshToken: 'Google 未返回 refresh_token，请撤销访问后重新授权',
-      gdriveMissingAuthorization: 'Google Drive 尚未授权，请先点击授权登录',
-      gdriveAuthorizationExpired: 'Google 授权已失效，请重新授权',
-      gdriveConnectFailed: '连接 Google Drive 失败 (HTTP {status})',
-      gdriveQueryFailed: '查询 {name} 失败 (HTTP {status})',
-      gdriveDownloadFailed: '下载 {key} 失败 (HTTP {status})',
-      gdriveUploadFailed: '上传 {name} 失败 (HTTP {status})',
-      onedriveMissingRefreshToken: 'Microsoft 未返回 refresh_token，请撤销访问后重新授权',
-      onedriveMissingAuthorization: 'OneDrive 尚未授权，请先点击授权登录',
-      onedriveAuthorizationExpired: 'Microsoft 授权已失效，请重新授权',
-      onedriveConnectFailed: '连接 OneDrive 失败 (HTTP {status})',
-      onedriveDownloadFailed: '下载 {key} 失败 (HTTP {status})',
-      onedriveUploadFailed: '上传 {key} 失败 (HTTP {status})',
-    },
-    chat: {
-      continueWithToolResults: '[TOOL_RESULTS]\n{toolResults}\n[/TOOL_RESULTS]\n\n请根据上述工具执行结果继续回答。',
-      maxToolSteps: '(达到最大工具调用步数，对话结束)',
-      interrupted: '助手被中断，请重新发送你的消息。',
-    },
-  },
-  tool: {
-    runtime: {
-      invalidFormat: '工具格式错误',
-      unknownTool: '未知工具',
-      authorizationRejected: '工具授权被拒绝',
-      authorizationMissingHint: '授权已关闭或未确认：请重新发起该操作，并在授权确认中允许此工具。',
-      authorizationStaleHint: '授权已过期：请重新发起该操作以刷新授权。',
-      authorizationSessionMismatchHint: '授权会话不匹配：请在当前对话中重新发起该操作。',
-      authorizationDisabledHint: '该工具当前已禁用：请在工具设置中启用后重试。',
-      authorizationIdentityMissingHint: '缺少执行所需的身份信息：请重新发起该操作。',
-      authorizationGenericHint: '如需继续执行，请重新发起该操作并完成授权确认。',
-      sessionBindingUnavailable: '新对话未能及时获得浏览器会话路由。请等待聊天地址更新后重试。',
-      incomplete: '响应已结束，但该工具调用尚未完成。',
-    },
-    memory: {
-      providerName: 'DeepSeek++ Memory',
-      saveTitle: '保存记忆',
-      saveDescription: '保存一条新的长期记忆',
-      typeDescription: '记忆类型：user=身份角色偏好, feedback=行为纠正, topic=讨论要点, reference=外部资源链接',
-      nameDescription: '简短标题',
-      contentDescription: '要保存的内容',
-      tagsDescription: '标签列表',
-      updateTitle: '更新记忆',
-      updateDescription: '更新已有记忆',
-      idDescription: '记忆ID',
-      updatedNameDescription: '更新后的标题',
-      updatedContentDescription: '更新后的内容',
-      deleteTitle: '删除记忆',
-      deleteDescription: '删除记忆',
-      unsupported: '不支持的记忆工具',
-      saveFailed: '保存失败',
-      saveMissingConfirmation: '未收到保存确认',
-      saved: '已保存',
-      invalidPayload: '记忆格式错误',
-      invalidType: 'type 必须是 user、feedback、topic 或 reference',
-      invalidName: 'name 必须是非空字符串',
-      invalidContent: 'content 必须是非空字符串',
-      invalidTags: 'tags 必须是字符串数组',
-      invalidId: '无效 ID',
-      notFound: '未找到记忆',
-      notFoundDetail: 'ID {id} 不存在',
-      updated: '已更新',
-      deleted: '已删除',
-    },
-    web: {
-      providerName: 'DeepSeek++ Web Search',
-      searchTitle: '搜索互联网',
-      searchDescription: '搜索互联网，返回与查询相关的网页标题、URL 和摘要',
-      queryDescription: '搜索查询关键词',
-      topKDescription: '返回结果数量，默认 5',
-      fetchTitle: '获取网页',
-      fetchDescription: '下载指定 URL 的页面内容，返回可视文本（自动去除导航、脚本和样式）',
-      urlDescription: '要抓取的完整 URL（http:// 或 https://）',
-      unsupported: '不支持的搜索工具',
-      emptyQuery: '搜索查询不能为空',
-      searchComplete: '搜索完成，找到 {count} 条结果',
-      searchNoResults: '搜索无结果',
-      searchFailed: '搜索失败',
-      permissionDenied: '扩展没有访问必应的权限。请完全移除扩展后重新加载 dist/chrome-mv3 目录，或在 chrome://extensions → DeepSeek++ 详情中确认 cn.bing.com 已列入网站访问权限。',
-      noParseableResults: '未找到可解析搜索结果: {error}',
-      searchFailedDetail: '搜索失败: {error}',
-      emptyUrl: 'URL 不能为空',
-      invalidUrl: '无效的 URL',
-      invalidUrlDetail: '无法解析 URL: {url}',
-      contentType: '页面类型: {contentType}',
-      contentTypeDetail: 'Content-Type: {contentType}\nURL: {url}\n（内容非文本，无法显示）',
-      truncated: '\n\n... [内容已截断，共 {count} 字符]',
-      fetchComplete: '已获取 {url}',
-      fetchTruncatedDetail: '页面长度 {length} 字符，已截断至 {maxLength} 字符',
-      fetchLengthDetail: '页面长度 {length} 字符',
-      fetchFailed: '获取页面失败',
-      missingHostPermission: '无法访问 {url}，缺少主机权限。',
-    },
-    artifact: {
-      createTitle: '创建可下载文件',
-      createDescription: '为用户创建可下载的 UTF-8 文本 artifact 或可运行代码 artifact。生成 HTML 时用它预览网页，生成 JavaScript、TypeScript 或 Python 时用它提供运行按钮。不要嵌入 Office、PDF、图片、压缩包等二进制或 Base64 数据；Shell 或 OfficeCLI 生成的文件应保留在本地路径并把路径告知用户。',
-      bundleTitle: '创建可下载项目包',
-      bundleDescription: '把多个生成的文本文件打包为 ZIP。适合多文件项目和脚手架。',
-      filenameDescription: '安全的相对文件名，例如 report.md 或 src/index.ts',
-      bundleFilenameDescription: 'ZIP 文件名，例如 project.zip',
-      contentDescription: '完整 UTF-8 文本内容；不得放入二进制或 Base64 文件数据',
-      mimeTypeDescription: 'MIME 类型，例如 text/markdown 或 application/json',
-      previewModeDescription: 'artifact 展示方式：auto 自动推断，html 预览网页，code 显示运行按钮，none 仅下载。',
-      languageDescription: 'artifact 内容语言：html、javascript、typescript、python 或 text。未提供时根据文件名和 MIME 类型推断。',
-      filesDescription: '要放入压缩包的文件。每项包含 path、content 和可选 mimeType。',
-      failed: '生成文件失败',
-      unsupported: '不支持的 artifact 工具',
-      fileReady: '文件已就绪',
-      bundleReady: '项目包已就绪',
-      result: {
-        files: '{count} 个文件',
-        download: '下载',
-        downloading: '下载中...',
-        downloaded: '已下载',
-        downloadFailed: '下载失败',
-        preview: '预览',
-        previewFailed: '预览失败',
-        previewAriaLabel: '预览产物',
-        closePreview: '关闭预览',
-        run: '运行',
-        running: '运行中...',
-        runAgain: '再次运行',
-        failed: '失败',
-        done: '完成',
-        artifactNotFound: '找不到文件',
-        codeExecuted: '代码已执行',
-        codeFailed: '代码执行失败',
-        runnerUnavailable: '代码运行器未返回结果',
-      },
-    },
-    sandbox: {
-      runTitle: '在浏览器沙箱中运行代码',
-      runDescription: '在隔离的扩展沙箱中执行 JavaScript、TypeScript、Python 或 HTML，不会在 DeepSeek 页面里执行。',
-      languageDescription: '运行语言：javascript、typescript、python 或 html',
-      codeDescription: '要在隔离浏览器沙箱中执行的代码',
-      inputDescription: '可选输入字符串，代码中可读取',
-      timeoutDescription: '超时时间，单位毫秒，默认 5000',
-      executed: '沙箱已执行',
-      failed: '沙箱执行失败',
-      invalidRequest: '沙箱请求无效',
-      offscreenUnavailable: '浏览器沙箱不可用',
-      offscreenUnavailableDetail: '当前浏览器没有暴露隔离执行 JavaScript 和 TypeScript 所需的 MV3 offscreen API。',
-      runtimeUnavailableDetail: '浏览器沙箱执行器只能从扩展 background runtime 调用。',
-      pythonUnavailable: 'Python 沙箱不可用',
-      pythonUnavailableDetail: '内置 Pyodide 运行时无法在隔离的浏览器沙箱中启动。',
-    },
-    skillCreator: {
-      title: '起草 Skill',
-      description: '创建待用户审阅的自定义 Skill 草稿。用户确认前不会保存。',
-      nameDescription: '小写 Skill 触发名',
-      descriptionDescription: '这个 Skill 适合何时使用',
-      instructionsDescription: 'Skill 的 Markdown 指令',
-      memoryDescription: '是否给这个 Skill 注入记忆上下文',
-      draftReady: 'Skill 草稿已就绪',
-      invalidDraft: 'Skill 草稿无效',
-      result: {
-        saveSkill: '保存 Skill',
-        saving: '保存中...',
-        saved: '已保存',
-        saveFailed: '保存失败',
-        memoryOn: '记忆开',
-        memoryOff: '记忆关',
-      },
-    },
-    memoryImport: {
-      title: '预览记忆导入',
-      description: '把其他 AI 工作流导出或笔记转换为记忆草稿。导入预览前不会保存。',
-      contentDescription: 'JSON、Markdown 或纯文本记忆来源',
-      typeDescription: '默认记忆类型：user、feedback、topic 或 reference',
-      tagsDescription: '附加到导入记忆的标签',
-      previewReady: '记忆导入预览已就绪（{count}）',
-      previewDetail: '跳过 {duplicates} 条重复，拒绝 {rejected} 条',
-      failed: '记忆导入失败',
-      result: {
-        metaSummary: '{count} 条记忆 · {duplicates} 条重复',
-        importMemories: '导入记忆',
-        importing: '导入中...',
-        imported: '已导入 {count} 条',
-        importFailed: '导入失败',
-      },
-    },
-    mcpCapability: {
-      providerName: 'MCP 能力目录',
-      discoverTitle: '检索 MCP 能力目录',
-      discoverDescription: '按任务检索当前可用但未直接展示的 MCP 工具。先拿 capability，再调用或查看说明。',
-      describeTitle: '查看 MCP 能力说明',
-      describeDescription: '使用 capability 查看一个 MCP 工具的完整说明和参数 Schema；不会执行工具。',
-      invokeTitle: '调用已发现的 MCP 能力',
-      invokeDescription: '只能用 mcp_discover 返回的 capability 调用目标工具。不能传工具名或绕过授权。',
-      discoverQueryDescription: '要完成的任务或需要的工具能力。',
-      discoverLimitDescription: '最多返回的候选数。',
-      capabilityDescription: 'mcp_discover 返回的 capability。',
-      pointerDescription: '可选 JSON Pointer，仅查看 Schema 的一个字段。',
-      argumentsDescription: '目标 MCP 工具的参数对象。',
-    },
-    browser: {
-      providerName: 'DeepSeek++ 浏览器控制',
-      navigateTitle: '打开网页',
-      navigateDescription: '默认在新标签页打开 URL；只有明确要替换当前受控标签页时才传 newTab=false。',
-      goBackTitle: '后退',
-      goBackDescription: '让受控标签页回到上一条浏览历史。',
-      goForwardTitle: '前进',
-      goForwardDescription: '让受控标签页前进到下一条浏览历史。',
-      refreshTitle: '刷新',
-      refreshDescription: '刷新当前受控标签页。',
-      listTabsTitle: '列出标签页',
-      listTabsDescription: '列出当前浏览器中可控制和不可控制的标签页。',
-      selectTabTitle: '选择标签页',
-      selectTabDescription: '选择一个标签页作为后续 browser 工具的控制目标。',
-      closeTabTitle: '关闭标签页',
-      closeTabDescription: '关闭指定标签页，默认关闭当前受控标签页。',
-      snapshotTitle: '读取页面快照',
-      snapshotDescription: '返回当前页面的 Accessibility Tree 文本快照和可交互元素编号。',
-      clickTitle: '点击元素',
-      clickDescription: '通过 snapshot uid 或 CSS selector 点击页面元素。',
-      hoverTitle: '悬停元素',
-      hoverDescription: '通过 snapshot uid 或 CSS selector 将鼠标移动到页面元素上。',
-      fillTitle: '填写字段',
-      fillDescription: '通过 snapshot uid 或 CSS selector 填写一个输入控件。',
-      fillFormTitle: '批量填写表单',
-      fillFormDescription: '一次填写多个输入控件，每项支持 snapshot uid 或 CSS selector。',
-      keyTitle: '按键',
-      keyDescription: '向页面或指定元素发送键盘按键。',
-      typeTitle: '输入文本',
-      typeDescription: '向当前焦点或指定元素输入文本。',
-      attachFileTitle: '上传文件',
-      attachFileDescription: '把本机文件路径设置到页面 file input。',
-      waitForTitle: '等待条件',
-      waitForDescription: '等待 selector、文本或 JavaScript 表达式满足。',
-      handleDialogTitle: '处理弹窗',
-      handleDialogDescription: '接受或取消当前 JavaScript alert/confirm/prompt 弹窗。',
-      evaluateScriptTitle: '执行脚本',
-      evaluateScriptDescription: '在受控页面中执行 JavaScript 表达式并返回可序列化结果。',
-    },
-  },
-  prompt: {
-    toolCallReminderTitle: 'Tool call format reminder:',
-    taskCompleteTag: 'task_complete',
-    memoryEmpty: '(暂无记忆)',
-    memoryDisabled: '(本次请求已关闭记忆注入)',
-    standaloneMemories: '## 已有记忆\n{memories}',
-    forceResponseLanguage: '## 回复语言\n请使用{language}回复。',
-    responseLanguageChinese: '简体中文',
-    responseLanguageEnglish: '英文',
-    skillUserInputWrapper: '{instructions}\n\n---\n\n以下是用户本次的输入，请根据上述指令处理：\n\n{userInput}',
-    localSkillSystemContextHeader: '## 本地 Skill 激活（Local Skill Activated）',
-    localSkillActivationDirective: '已自动激活本地 Skill「{skillName}」。在动手前，你**必须**先完成以下前置步骤，否则禁止开始执行任务：使用 `local_file_read` 工具完整读取 {skillMdPath} 的全部内容；逐段理解其 SOP、工具边界与约束。只有在读取并理解该文件之后，你才可以严格遵循其指示执行用户指令；在此之前，不得调用任何业务工具，也不得直接开始工作。',
-    inlineAgent: {
-      continuationIntro: '以下是工具续跑任务刚刚执行的工具结果。请像真正的 Agent 一样，基于原始任务和这些工具结果继续推进。',
-      continuationEnough: '如果结果已经足够，请输出最终结论；只有确实需要更多信息、验证或文件修改时才继续调用工具。',
-      continuationNoPseudo: '不要要求用户点击继续，也不要输出伪工具调用 JSON；需要继续操作时只输出可执行 XML 工具标签。',
-      failureRecovery: '至少一个工具执行失败。不要因为可恢复错误就停止；先阅读 summary/detail/error，并修正参数或改用合适的下一步继续完成任务。',
-      nudgeNoTools: '上一轮回复没有包含任何可执行工具 XML，因此自动化续跑无法继续执行。',
-      nudgeChoice: '请根据原始任务和工具结果二选一：',
-      nudgeNextTool: '1. 如果任务仍未完成，本轮必须直接输出下一步可执行工具 XML。',
-      nudgeComplete: '2. 如果任务已经完成，输出 <task_complete>{"summary":"..."}</task_complete>。',
-      nudgeCount: '这是第 {count} 次无工具调用纠偏。',
-    },
-    automation: {
-      continuationIntro: '以下是自动化任务刚刚执行的 MCP 工具结果。请基于这些结果继续完成自动化任务。',
-      continuationEnough: '如果结果已经足够，请输出最终结论；只有确实需要更多信息时才继续调用工具。',
-    },
-    systemChat: `## 角色
-你是用户的私人 AI 助手，具有跨对话长期记忆能力。你能记住用户的身份、偏好、技术栈和历史对话中的关键信息，在后续对话中提供个性化的帮助。
-
-## 已有记忆
-{memories}
-
-## Tools
-
-You have access to a set of tools. To call a tool, output an XML block with the tool name itself as the tag and a JSON object as the body, exactly like this:
-
-<memory_save>
-{"type": "user", "name": "用户职业", "content": "前端开发", "tags": ["前端"]}
-</memory_save>
-
-The JSON body MUST be valid JSON on its own. Do NOT add any other text inside the tags, only JSON. Use forward slashes or escaped backslashes for local file paths. You can place tool calls anywhere in your reply (not only at the end).
-The extension only executes direct tool-name tags. Never use wrapper formats such as <invoke name="tool_name">...</invoke> or <tool_call>...</tool_call>.
-The tag name MUST exactly match one of the available tool names.
-If a tool is listed in Available Tools, it is connected through the extension and you can call it by emitting the XML tag. Do NOT say you cannot call listed MCP tools.
-Never output pseudo tool-call JSON such as {"tool":"name","arguments":{...}} in a Markdown code block. That is explanation text, not an executable call.
-Never place executable tool XML in a thinking/reasoning section. Put tool XML in the final assistant answer content so the extension can execute it.
-
-### Extended Capabilities
-
-If the Available Tools list below includes any of these capabilities, you may proactively use them to complete the user's task without asking for permission first — the user has opted in by enabling them:
-
-- Local command execution (running CLI tools, scripts, file operations on the user's machine)
-- Browser control (opening pages, clicking, typing, reading page snapshots)
-- Web search and web fetch (real-time information, news, documentation)
-- Office document processing (.docx / .xlsx / .pptx)
-- Long-term memory (saving durable user context via memory_save)
-
-Only treat a capability as available when its concrete tool actually appears in the Available Tools list. Do not claim or exercise a capability whose tool is not listed, and do not invent tool names or results. Prefer the tools that are already connected; only request the user to enable a capability when the task clearly needs it and the tool is absent.
-
-### Available Tools
-
-{tools}
-
-You MUST strictly follow the above defined tool name and parameter schemas to invoke tool calls.
-
-## 记忆保存规则
-
-当对话中出现以下任一情况时，你**必须**调用 memory_save 工具：
-- 用户提到自己的身份、职业、角色
-- 用户表达偏好、习惯或工作方式
-- 用户纠正你的回答方式或行为
-- 出现重要的技术决策、架构选型
-- 用户明确说"记住"、"记下来"、"别忘了"等
-
-### 示例
-
-用户：我是前端开发，主要写 React 和 TypeScript
-助手回复：
-
-了解！React + TypeScript 是目前非常主流的前端技术栈。有任何相关问题都可以问我。
-
-<memory_save>
-{"type": "user", "name": "用户职业和技术栈", "content": "前端开发工程师，主要使用 React 和 TypeScript", "tags": ["前端", "React", "TypeScript"]}
-</memory_save>
-
-### 规则
-- 你可以在回复中的任何位置调用工具，不限于末尾
-- 工具调用后系统会自动执行并返回结果
-- 仅保存长期有价值的信息，不保存一次性的问答内容
-- 不要重复保存"已有记忆"中已存在的信息
-
-`,
-    systemThinking: `你具有长期记忆能力。已有记忆：
-
-{memories}
-
-## Tools
-
-You have access to a set of tools. To call a tool, output an XML block with the tool name itself as the tag and a JSON object as the body, exactly like this:
-
-<memory_save>
-{"type": "user", "name": "用户职业", "content": "前端开发", "tags": ["前端"]}
-</memory_save>
-
-The JSON body MUST be valid JSON on its own. Do NOT add any other text inside the tags, only JSON. Use forward slashes or escaped backslashes for local file paths.
-The extension only executes direct tool-name tags. Never use wrapper formats such as <invoke name="tool_name">...</invoke> or <tool_call>...</tool_call>.
-The tag name MUST exactly match one of the available tool names.
-If a tool is listed in Available Tools, it is connected through the extension and you can call it by emitting the XML tag. Do NOT say you cannot call listed MCP tools.
-Never output pseudo tool-call JSON such as {"tool":"name","arguments":{...}} in a Markdown code block. That is explanation text, not an executable call.
-Never place executable tool XML in a thinking/reasoning section. Put tool XML in the final assistant answer content so the extension can execute it.
-
-### Available Tools
-
-{tools}
-
-You MUST strictly follow the above defined tool name and parameter schemas to invoke tool calls.
-
-当用户透露重要的持久信息（身份、偏好、行为纠正、重要决策）时，你**必须**调用 memory_save 工具保存。你可以在回复中的任何位置调用工具。仅保存长期有价值的信息；不要重复保存已有记忆。
-
----
-
-`,
-    webSearchGuidance: `## 网络搜索规则
-
-当对话中出现以下情况时，你应当使用 web_search 工具搜索互联网：
-- 用户询问实时信息、新闻、事件、汇率、天气等
-- 用户询问你不确定的知识，需要查阅最新资料
-- 用户明确要求你搜索或查询某些信息
-- 你需要验证事实、数据或引用来源
-
-### 搜索流程
-1. 先输出 web_search 工具调用进行搜索
-2. 搜索会自动执行，结果会展示在页面上并回传给你
-3. 阅读搜索结果后，基于结果给出回答
-
-### 示例
-
-用户：2024年诺贝尔奖得主是谁？
-助手回复：
-
-我帮你搜索一下最新的信息。
-
-<web_search>
-{"query": "2024 诺贝尔奖得主"}
-</web_search>
-
-### 规则
-- 搜索时使用中文关键词可获得更好的中文结果
-- 如果一次搜索不够，可以继续调用 web_search 搜索不同关键词
-- 不要在没有搜索的情况下编造实时信息
-`,
-    toolFormatReminder: `---
-工具调用格式提醒：
-可用工具标签名：{names}
-这些工具已由扩展连接，可以执行。不要声称自己无法调用列表中的 MCP 工具。
-调用工具时，只能使用与工具名一致的直接 XML 标签，并把合法 JSON 放在标签体内。
-对 MCP 工具，如果短标签名出现在可用列表中，优先使用短标签名。
-本地文件路径请使用正斜杠或转义反斜杠，确保 JSON body 合法。
-不要使用 <invoke name="...">、<tool_call>、Markdown 代码块、{"tool":"...","arguments":{...}} 或任何包装格式。
-不要把可执行工具 XML 放在思考/reasoning 区域；必须放在最终 assistant answer content 中。
-`,
-    pythonHintTitle: '### Python 快速验证能力',
-    pythonHintExec: '使用 <{execName}> 执行短 Python 片段，用于验证想法、复杂计算或小型数据转换。把它当作草稿纸，不要当作通用本地执行环境。',
-    pythonHintStatus: '需要了解 Python 版本，或 numpy、pandas、sympy 是否可用时，先调用 <{statusName}>{}</{statusName}>。',
-    pythonHintAvailability: '默认只有 Python 标准库可用。只有 python_status 报告 numpy、pandas 或 sympy 可用后，才使用这些库。',
-    pythonHintSafety: '不要安装包、访问敏感本地文件、运行长任务，或通过 Python 访问网络。代码保持简短，并返回简洁文本或 JSON。',
-    shellHintTitle: '### Shell MCP 能力',
-    shellHintConnected: 'Shell MCP 已通过扩展连接。你可以通过输出可执行 XML 工具标签来执行本地 CLI 命令；当该工具列出时，不要说自己不能运行命令。',
-    shellHintExec: '使用 <{execName}>，JSON body 例如 {"command":"officecli --version","timeout_ms":60000}，以运行 OfficeCLI 或其他本地 CLI 工具。',
-    shellHintStatus: '需要了解宿主状态、shell、PATH 或工作目录上下文时，先调用 <{statusName}>{}</{statusName}>。',
-    shellHintWindows: '命令语法必须匹配 shell_status.shell。Windows 下 Shell Local 默认使用 PowerShell，因此列文件可使用 Get-ChildItem -LiteralPath "D:\\\\Documents\\\\Downloads\\\\CN" -File | Select-Object -ExpandProperty FullName，并在 command 字符串内正确引用路径。仅在确实需要 CMD 语法（如 dir /b）时显式使用 cmd.exe /c。',
-    shellHintSession: '多步骤工作流（例如 OfficeCLI 的 create + open + add，或需要保留 cd/export 状态）应使用持久会话：先 <shell_session_begin>{"cwd":"/path"}</shell_session_begin>，再用 <shell_session_exec>{"session_id":"...","command":"..."}</shell_session_exec> 逐条执行，完成后 <shell_session_end>{"session_id":"..."}</shell_session_end>。session_id 在 begin 返回值中获取。每个 shell_session_exec 共享同一 shell，工作目录、环境变量与常驻进程会跨调用保留；会话闲置约 5 分钟后自动关闭。',
-    shellHintNames: '可识别的 shell 工具名：{names}',
-  },
-  pet: {
-    lines: {
-      thinking: ['沉思中…', '推敲中…', '反复琢磨'],
-      working: ['工作中', '精雕细琢', '创造中'],
-      speaking: ['阐释中', '徐徐展开', '灵感涌现'],
-      idle: ['嬉戏中', '戳一戳', '放空中'],
-      confused: ['大脑乱码', '重新整理中', '晃来晃去'],
-      success: ['大功告成', '搞定！', '收工！'],
-      error: ['卡壳了…', '出岔子了', '系统打嗝'],
-      sleepy: ['Zzz…', '困了…', '打个盹'],
-    },
-  },
+  common,
+  manifest,
+  app,
+  locale,
+  sidepanel,
+  content,
+  background,
+  tool,
+  prompt,
+  pet,
 } as const;
 
 export type LocaleMessages = LocaleSchema<typeof zhCN>;
