+++ core/i18n/resources/zh-CN/app.ts
@@ -1,0 +1,15 @@
+export const app = {
+  sideNavLabel: '侧栏导航',
+  version: 'v{version}',
+  tabs: {
+    chat: '对话',
+    library: '资料',
+    memory: '记忆',
+    projects: '项目',
+    saved: '保存',
+    capabilities: '能力',
+    preset: '预设',
+    automation: '自动化',
+    settings: '设置',
+  },
+} as const;
