+++ core/i18n/resources/zh-CN/background.ts
@@ -1,0 +1,44 @@
+export const background = {
+  contextMenus: {
+    sendToChat: '发送到对话',
+  },
+  auth: {
+    missingDeepSeek: '请先在 chat.deepseek.com 登录，或刷新 DeepSeek 页面后重试。',
+  },
+  export: {
+    generating: '生成导出文件',
+    cancelled: '导出已取消',
+  },
+  sync: {
+    missingWebDav: '未配置 WebDAV',
+    missingSync: '未配置云同步',
+    configChanged: '其他扩展窗口已修改同步设置，请确认当前目标后重试',
+    configCommitIndeterminate: '无法确认同步设置是否已保存，将重新读取当前设置',
+    missingRemoteFile: '云端缺少 {file}，已停止下载以避免覆盖本地数据',
+    authorizationNotRequired: '当前同步方式不需要授权',
+    identityUnavailable: '当前浏览器环境不支持 OAuth 授权',
+    oauthCancelled: '授权已取消',
+    oauthProviderError: '授权失败：{error}',
+    oauthMissingCode: '授权失败：未返回授权码',
+    oauthRefreshFailed: '刷新访问令牌失败 (HTTP {status}){detail}',
+    oauthTokenExchangeFailed: '获取令牌失败 (HTTP {status}){detail}',
+    gdriveMissingRefreshToken: 'Google 未返回 refresh_token，请撤销访问后重新授权',
+    gdriveMissingAuthorization: 'Google Drive 尚未授权，请先点击授权登录',
+    gdriveAuthorizationExpired: 'Google 授权已失效，请重新授权',
+    gdriveConnectFailed: '连接 Google Drive 失败 (HTTP {status})',
+    gdriveQueryFailed: '查询 {name} 失败 (HTTP {status})',
+    gdriveDownloadFailed: '下载 {key} 失败 (HTTP {status})',
+    gdriveUploadFailed: '上传 {name} 失败 (HTTP {status})',
+    onedriveMissingRefreshToken: 'Microsoft 未返回 refresh_token，请撤销访问后重新授权',
+    onedriveMissingAuthorization: 'OneDrive 尚未授权，请先点击授权登录',
+    onedriveAuthorizationExpired: 'Microsoft 授权已失效，请重新授权',
+    onedriveConnectFailed: '连接 OneDrive 失败 (HTTP {status})',
+    onedriveDownloadFailed: '下载 {key} 失败 (HTTP {status})',
+    onedriveUploadFailed: '上传 {key} 失败 (HTTP {status})',
+  },
+  chat: {
+    continueWithToolResults: '[TOOL_RESULTS]\n{toolResults}\n[/TOOL_RESULTS]\n\n请根据上述工具执行结果继续回答。',
+    maxToolSteps: '(达到最大工具调用步数，对话结束)',
+    interrupted: '助手被中断，请重新发送你的消息。',
+  },
+} as const;
