+++ core/i18n/resources/zh-CN/common.ts
@@ -1,0 +1,32 @@
+export const common = {
+  add: '新增',
+  all: '全部',
+  auto: '自动',
+  cancel: '取消',
+  clear: '清除',
+  close: '关闭',
+  confirm: '确认',
+  delete: '删除',
+  deactivate: '停用',
+  disabled: '已停用',
+  edit: '编辑',
+  enable: '启用',
+  enabled: '已启用',
+  error: '错误',
+  loading: '加载中...',
+  manual: '手动',
+  none: '无',
+  open: '打开',
+  preview: '预览',
+  refresh: '刷新',
+  remove: '移除',
+  retry: '重试',
+  save: '保存',
+  saveChanges: '保存更改',
+  search: '搜索',
+  status: '状态',
+  success: '成功',
+  sync: '同步',
+  test: '测试',
+  update: '更新',
+} as const;
