+++ core/i18n/resources/zh-CN/content.ts
@@ -1,0 +1,141 @@
+export const content = {
+  conversation: {
+    untitled: '未命名对话',
+  },
+  export: {
+    buttonIdle: '导出当前对话',
+    buttonRunning: '正在导出当前对话',
+    emptyConversation: '当前页面还没有可导出的对话。',
+    formatDialogLabel: '选择导出格式',
+    formatTitle: '导出格式',
+    scopeLabel: '内容范围',
+    scopeFull: '全部内容（含工具调用、续跑、思考、注入信息）',
+    scopeInputOutput: '仅输入 + 最终产出',
+    formatHtml: 'HTML',
+    formatMarkdown: 'Markdown',
+    formatPdf: 'PDF',
+    formatImageManifest: '图片清单',
+    submit: '导出',
+    progress: '正在导出当前对话...',
+    failed: '导出失败',
+    cancelled: '导出已取消',
+    partialSuccess: '导出完成，但部分会话读取失败，请检查导出文件中的 Export Warnings。',
+    success: '当前对话已导出。',
+  },
+  historyOrganizer: {
+    enhancedSearchTitle: 'DeepSeek++ 搜索增强',
+    tagFilterLabel: '按标签筛选结果',
+    tagPlaceholder: '输入标签名',
+    currentTagsLabel: '给当前对话加标签',
+    currentTagsPlaceholder: '逗号分隔，例如：港股, 写作',
+    emptySearchStatus: 'DeepSeek++：等待官方搜索结果',
+    visibleStatus: 'DeepSeek++：已显示 {visible}/{total}',
+    loadError: 'DeepSeek++：历史标签加载失败：{message}',
+    saveError: 'DeepSeek++：历史标签保存失败：{message}',
+  },
+  projectSidebar: {
+    title: '项目',
+    empty: '还没有项目',
+    expandProject: '展开项目：{name}',
+    collapseProject: '收起项目：{name}',
+    showMore: '展开显示',
+    showLess: '收起显示',
+    moveCurrentToProject: '移入当前对话到 {name}',
+    removeCurrentFromProject: '从 {name} 移出当前对话',
+    joinProject: '加入项目',
+    joinProjectNamed: '加入项目：{name}',
+    moveToProjectNamed: '移动到项目：{name}',
+    currentProjectNamed: '已在项目：{name}',
+    removeFromProjectNamed: '移除项目：{name}',
+    conversationActions: '会话操作',
+    newConversationInProject: '在 {name} 下开启新对话',
+    useNextConversation: '下一条新会话使用 {name}',
+    cancelNextConversation: '取消下一条新会话使用 {name}',
+    pendingNextConversation: '下一条新会话将使用此项目',
+    operationFailed: '项目操作失败：{message}',
+    age: {
+      justNow: '刚刚',
+      minutesAgo: '{count}分钟前',
+      hoursAgo: '{count}小时前',
+      daysAgo: '{count}天前',
+    },
+  },
+  uxPolish: {
+    downloadCode: '下载',
+    downloadMessageMarkdownButton: 'MD',
+    downloadMessageMarkdownTitle: '下载消息为 Markdown',
+    copyMessageButton: '复制',
+    copyMessageTitle: '复制完整对话输出',
+  },
+  toolBlock: {
+    title: '已执行工具（{count}次）',
+    pythonInterpreter: 'Python 解释器',
+    summaries: {
+      saved: '已保存',
+      updated: '已更新',
+      deleted: '已删除',
+      searched: '已搜索',
+      fetched: '已获取',
+      running: '执行中',
+      executed: '已执行',
+      failed: '执行失败',
+      messageFailed: '工具消息发送失败',
+      backgroundFailed: '后台工具执行失败',
+    },
+    missingResultDetail: '后台没有返回工具执行结果。请刷新当前 DeepSeek 页面后重试。',
+    missingSandboxResultDetail: '沙箱工具没有返回执行结果。请重载 DeepSeek++ 扩展并刷新当前 DeepSeek 页面后重试；它不依赖 MCP 或 Shell Local。',
+    missingMcpResultDetail: '后台没有返回 MCP 工具执行结果。请刷新当前 DeepSeek 页面后重试；如果仍失败，在 MCP 页重新测试 Shell Local。',
+    invalidResultDetail: '后台返回的工具结果结构无效：{preview}',
+  },
+  agent: {
+    stop: '停止',
+    stopped: '已停止',
+    step: 'Step {index}',
+    streaming: 'streaming...',
+    complete: '完成',
+    completeWithTools: '完成（{count} 个工具）',
+    executingTools: '执行工具中',
+    budgetReached: 'DeepSeek++ 已执行 {count} 轮自动工具续跑并暂停，避免把未完成任务伪装成最终答案。请发送“继续”让当前对话接着处理剩余工作。',
+    error: '执行出错',
+    footerComplete: 'Agent 完成（{steps} 步，{tools} 次工具调用）',
+    process: '过程',
+    tools: '工具调用',
+    results: '执行结果',
+    running: 'Agent 运行中 · 第 {step} 步 · 已调用 {tools} 次工具',
+    footerError: 'Agent 执行出错（{steps} 步，{tools} 次工具调用）',
+    autoSaved: '长文产出已自动保存为产物：',
+    autoSavedDownload: '下载',
+    concurrencyGuard: 'Agent 正在执行中，已忽略此消息触发的重复任务。如需中断，请点击 Agent 面板的「停止」。',
+  },
+  permission: {
+    webFetch: 'DeepSeek++ 需要访问 {origin} 的权限以获取该页面内容',
+    deny: '拒绝',
+    grant: '授权',
+    requesting: '请求中...',
+  },
+  skillPopup: {
+    hint: '↑↓ 导航 · Enter 选择 · Esc 关闭',
+  },
+  tokenSpeed: {
+    title: '累计 Token：{tokens}；Token 输出速度：{speed}{idle}。Token 来源：{tokenSource}；速度来源：{speedSource}',
+    idleSuffix: '（空闲）',
+    sourceServer: '服务端',
+    sourceEstimated: '本地估算',
+  },
+  multimodalMedia: {
+    buttonTitle: '添加 DeepSeek++ 多模态图片或视频',
+    removeTitle: '移除 {name}',
+    pastedFileName: '粘贴的媒体',
+    added: '已添加 {count} 个多模态附件',
+    analyzing: '正在分析 {count} 个多模态附件...',
+    analyzed: '多模态分析完成（{count} 项）',
+    failed: '多模态分析失败：{message}',
+    tooMany: '每轮最多添加 {count} 个多模态附件',
+    unsupported: '{name} 不是支持的图片或视频',
+    imageTooLarge: '{name} 超过图片上限 {limit}',
+    videoTooLarge: '{name} 超过内联视频上限 {limit}。请改用公开视频 URL，或等待本地路径选择能力。',
+    invalidRequest: 'DeepSeek 请求体异常，无法附加多模态媒体。',
+    emptyPrompt: '请先输入消息，再发送多模态媒体。',
+  },
+  extensionReloaded: '扩展已重新加载，请刷新当前 DeepSeek 页面后重试。',
+} as const;
