+++ core/i18n/resources/zh-CN/locale.ts
@@ -1,0 +1,7 @@
+export const locale = {
+  settingTitle: '语言',
+  settingDescription: '选择 DeepSeek++ 的界面和模型行为语言。',
+  auto: '跟随浏览器',
+  zhCN: '简体中文',
+  en: 'English',
+} as const;
