+++ core/i18n/resources/zh-CN/manifest.ts
@@ -1,0 +1,5 @@
+export const manifest = {
+  name: 'DeepSeek++',
+  description: 'DeepSeek 的 Agentic 记忆、Skill、执行、自动化和 MCP 工具',
+  actionTitle: 'DeepSeek++',
+} as const;
