+++ core/i18n/resources/zh-CN/pet.ts
@@ -1,0 +1,12 @@
+export const pet = {
+  lines: {
+    thinking: ['沉思中…', '推敲中…', '反复琢磨'],
+    working: ['工作中', '精雕细琢', '创造中'],
+    speaking: ['阐释中', '徐徐展开', '灵感涌现'],
+    idle: ['嬉戏中', '戳一戳', '放空中'],
+    confused: ['大脑乱码', '重新整理中', '晃来晃去'],
+    success: ['大功告成', '搞定！', '收工！'],
+    error: ['卡壳了…', '出岔子了', '系统打嗝'],
+    sleepy: ['Zzz…', '困了…', '打个盹'],
+  },
+} as const;
