+++ core/i18n/resources/zh-CN/tool.ts
@@ -1,0 +1,215 @@
+export const tool = {
+  runtime: {
+    invalidFormat: '工具格式错误',
+    unknownTool: '未知工具',
+    authorizationRejected: '工具授权被拒绝',
+    authorizationMissingHint: '授权已关闭或未确认：请重新发起该操作，并在授权确认中允许此工具。',
+    authorizationStaleHint: '授权已过期：请重新发起该操作以刷新授权。',
+    authorizationSessionMismatchHint: '授权会话不匹配：请在当前对话中重新发起该操作。',
+    authorizationDisabledHint: '该工具当前已禁用：请在工具设置中启用后重试。',
+    authorizationIdentityMissingHint: '缺少执行所需的身份信息：请重新发起该操作。',
+    authorizationGenericHint: '如需继续执行，请重新发起该操作并完成授权确认。',
+    sessionBindingUnavailable: '新对话未能及时获得浏览器会话路由。请等待聊天地址更新后重试。',
+    incomplete: '响应已结束，但该工具调用尚未完成。',
+  },
+  memory: {
+    providerName: 'DeepSeek++ Memory',
+    saveTitle: '保存记忆',
+    saveDescription: '保存一条新的长期记忆',
+    typeDescription: '记忆类型：user=身份角色偏好, feedback=行为纠正, topic=讨论要点, reference=外部资源链接',
+    nameDescription: '简短标题',
+    contentDescription: '要保存的内容',
+    tagsDescription: '标签列表',
+    updateTitle: '更新记忆',
+    updateDescription: '更新已有记忆',
+    idDescription: '记忆ID',
+    updatedNameDescription: '更新后的标题',
+    updatedContentDescription: '更新后的内容',
+    deleteTitle: '删除记忆',
+    deleteDescription: '删除记忆',
+    unsupported: '不支持的记忆工具',
+    saveFailed: '保存失败',
+    saveMissingConfirmation: '未收到保存确认',
+    saved: '已保存',
+    invalidPayload: '记忆格式错误',
+    invalidType: 'type 必须是 user、feedback、topic 或 reference',
+    invalidName: 'name 必须是非空字符串',
+    invalidContent: 'content 必须是非空字符串',
+    invalidTags: 'tags 必须是字符串数组',
+    invalidId: '无效 ID',
+    notFound: '未找到记忆',
+    notFoundDetail: 'ID {id} 不存在',
+    updated: '已更新',
+    deleted: '已删除',
+  },
+  web: {
+    providerName: 'DeepSeek++ Web Search',
+    searchTitle: '搜索互联网',
+    searchDescription: '搜索互联网，返回与查询相关的网页标题、URL 和摘要',
+    queryDescription: '搜索查询关键词',
+    topKDescription: '返回结果数量，默认 5',
+    fetchTitle: '获取网页',
+    fetchDescription: '下载指定 URL 的页面内容，返回可视文本（自动去除导航、脚本和样式）',
+    urlDescription: '要抓取的完整 URL（http:// 或 https://）',
+    unsupported: '不支持的搜索工具',
+    emptyQuery: '搜索查询不能为空',
+    searchComplete: '搜索完成，找到 {count} 条结果',
+    searchNoResults: '搜索无结果',
+    searchFailed: '搜索失败',
+    permissionDenied: '扩展没有访问必应的权限。请完全移除扩展后重新加载 dist/chrome-mv3 目录，或在 chrome://extensions → DeepSeek++ 详情中确认 cn.bing.com 已列入网站访问权限。',
+    noParseableResults: '未找到可解析搜索结果: {error}',
+    searchFailedDetail: '搜索失败: {error}',
+    emptyUrl: 'URL 不能为空',
+    invalidUrl: '无效的 URL',
+    invalidUrlDetail: '无法解析 URL: {url}',
+    contentType: '页面类型: {contentType}',
+    contentTypeDetail: 'Content-Type: {contentType}\nURL: {url}\n（内容非文本，无法显示）',
+    truncated: '\n\n... [内容已截断，共 {count} 字符]',
+    fetchComplete: '已获取 {url}',
+    fetchTruncatedDetail: '页面长度 {length} 字符，已截断至 {maxLength} 字符',
+    fetchLengthDetail: '页面长度 {length} 字符',
+    fetchFailed: '获取页面失败',
+    missingHostPermission: '无法访问 {url}，缺少主机权限。',
+  },
+  artifact: {
+    createTitle: '创建可下载文件',
+    createDescription: '为用户创建可下载的 UTF-8 文本 artifact 或可运行代码 artifact。生成 HTML 时用它预览网页，生成 JavaScript、TypeScript 或 Python 时用它提供运行按钮。不要嵌入 Office、PDF、图片、压缩包等二进制或 Base64 数据；Shell 或 OfficeCLI 生成的文件应保留在本地路径并把路径告知用户。',
+    bundleTitle: '创建可下载项目包',
+    bundleDescription: '把多个生成的文本文件打包为 ZIP。适合多文件项目和脚手架。',
+    filenameDescription: '安全的相对文件名，例如 report.md 或 src/index.ts',
+    bundleFilenameDescription: 'ZIP 文件名，例如 project.zip',
+    contentDescription: '完整 UTF-8 文本内容；不得放入二进制或 Base64 文件数据',
+    mimeTypeDescription: 'MIME 类型，例如 text/markdown 或 application/json',
+    previewModeDescription: 'artifact 展示方式：auto 自动推断，html 预览网页，code 显示运行按钮，none 仅下载。',
+    languageDescription: 'artifact 内容语言：html、javascript、typescript、python 或 text。未提供时根据文件名和 MIME 类型推断。',
+    filesDescription: '要放入压缩包的文件。每项包含 path、content 和可选 mimeType。',
+    failed: '生成文件失败',
+    unsupported: '不支持的 artifact 工具',
+    fileReady: '文件已就绪',
+    bundleReady: '项目包已就绪',
+    result: {
+      files: '{count} 个文件',
+      download: '下载',
+      downloading: '下载中...',
+      downloaded: '已下载',
+      downloadFailed: '下载失败',
+      preview: '预览',
+      previewFailed: '预览失败',
+      previewAriaLabel: '预览产物',
+      closePreview: '关闭预览',
+      run: '运行',
+      running: '运行中...',
+      runAgain: '再次运行',
+      failed: '失败',
+      done: '完成',
+      artifactNotFound: '找不到文件',
+      codeExecuted: '代码已执行',
+      codeFailed: '代码执行失败',
+      runnerUnavailable: '代码运行器未返回结果',
+    },
+  },
+  sandbox: {
+    runTitle: '在浏览器沙箱中运行代码',
+    runDescription: '在隔离的扩展沙箱中执行 JavaScript、TypeScript、Python 或 HTML，不会在 DeepSeek 页面里执行。',
+    languageDescription: '运行语言：javascript、typescript、python 或 html',
+    codeDescription: '要在隔离浏览器沙箱中执行的代码',
+    inputDescription: '可选输入字符串，代码中可读取',
+    timeoutDescription: '超时时间，单位毫秒，默认 5000',
+    executed: '沙箱已执行',
+    failed: '沙箱执行失败',
+    invalidRequest: '沙箱请求无效',
+    offscreenUnavailable: '浏览器沙箱不可用',
+    offscreenUnavailableDetail: '当前浏览器没有暴露隔离执行 JavaScript 和 TypeScript 所需的 MV3 offscreen API。',
+    runtimeUnavailableDetail: '浏览器沙箱执行器只能从扩展 background runtime 调用。',
+    pythonUnavailable: 'Python 沙箱不可用',
+    pythonUnavailableDetail: '内置 Pyodide 运行时无法在隔离的浏览器沙箱中启动。',
+  },
+  skillCreator: {
+    title: '起草 Skill',
+    description: '创建待用户审阅的自定义 Skill 草稿。用户确认前不会保存。',
+    nameDescription: '小写 Skill 触发名',
+    descriptionDescription: '这个 Skill 适合何时使用',
+    instructionsDescription: 'Skill 的 Markdown 指令',
+    memoryDescription: '是否给这个 Skill 注入记忆上下文',
+    draftReady: 'Skill 草稿已就绪',
+    invalidDraft: 'Skill 草稿无效',
+    result: {
+      saveSkill: '保存 Skill',
+      saving: '保存中...',
+      saved: '已保存',
+      saveFailed: '保存失败',
+      memoryOn: '记忆开',
+      memoryOff: '记忆关',
+    },
+  },
+  memoryImport: {
+    title: '预览记忆导入',
+    description: '把其他 AI 工作流导出或笔记转换为记忆草稿。导入预览前不会保存。',
+    contentDescription: 'JSON、Markdown 或纯文本记忆来源',
+    typeDescription: '默认记忆类型：user、feedback、topic 或 reference',
+    tagsDescription: '附加到导入记忆的标签',
+    previewReady: '记忆导入预览已就绪（{count}）',
+    previewDetail: '跳过 {duplicates} 条重复，拒绝 {rejected} 条',
+    failed: '记忆导入失败',
+    result: {
+      metaSummary: '{count} 条记忆 · {duplicates} 条重复',
+      importMemories: '导入记忆',
+      importing: '导入中...',
+      imported: '已导入 {count} 条',
+      importFailed: '导入失败',
+    },
+  },
+  mcpCapability: {
+    providerName: 'MCP 能力目录',
+    discoverTitle: '检索 MCP 能力目录',
+    discoverDescription: '按任务检索当前可用但未直接展示的 MCP 工具。先拿 capability，再调用或查看说明。',
+    describeTitle: '查看 MCP 能力说明',
+    describeDescription: '使用 capability 查看一个 MCP 工具的完整说明和参数 Schema；不会执行工具。',
+    invokeTitle: '调用已发现的 MCP 能力',
+    invokeDescription: '只能用 mcp_discover 返回的 capability 调用目标工具。不能传工具名或绕过授权。',
+    discoverQueryDescription: '要完成的任务或需要的工具能力。',
+    discoverLimitDescription: '最多返回的候选数。',
+    capabilityDescription: 'mcp_discover 返回的 capability。',
+    pointerDescription: '可选 JSON Pointer，仅查看 Schema 的一个字段。',
+    argumentsDescription: '目标 MCP 工具的参数对象。',
+  },
+  browser: {
+    providerName: 'DeepSeek++ 浏览器控制',
+    navigateTitle: '打开网页',
+    navigateDescription: '默认在新标签页打开 URL；只有明确要替换当前受控标签页时才传 newTab=false。',
+    goBackTitle: '后退',
+    goBackDescription: '让受控标签页回到上一条浏览历史。',
+    goForwardTitle: '前进',
+    goForwardDescription: '让受控标签页前进到下一条浏览历史。',
+    refreshTitle: '刷新',
+    refreshDescription: '刷新当前受控标签页。',
+    listTabsTitle: '列出标签页',
+    listTabsDescription: '列出当前浏览器中可控制和不可控制的标签页。',
+    selectTabTitle: '选择标签页',
+    selectTabDescription: '选择一个标签页作为后续 browser 工具的控制目标。',
+    closeTabTitle: '关闭标签页',
+    closeTabDescription: '关闭指定标签页，默认关闭当前受控标签页。',
+    snapshotTitle: '读取页面快照',
+    snapshotDescription: '返回当前页面的 Accessibility Tree 文本快照和可交互元素编号。',
+    clickTitle: '点击元素',
+    clickDescription: '通过 snapshot uid 或 CSS selector 点击页面元素。',
+    hoverTitle: '悬停元素',
+    hoverDescription: '通过 snapshot uid 或 CSS selector 将鼠标移动到页面元素上。',
+    fillTitle: '填写字段',
+    fillDescription: '通过 snapshot uid 或 CSS selector 填写一个输入控件。',
+    fillFormTitle: '批量填写表单',
+    fillFormDescription: '一次填写多个输入控件，每项支持 snapshot uid 或 CSS selector。',
+    keyTitle: '按键',
+    keyDescription: '向页面或指定元素发送键盘按键。',
+    typeTitle: '输入文本',
+    typeDescription: '向当前焦点或指定元素输入文本。',
+    attachFileTitle: '上传文件',
+    attachFileDescription: '把本机文件路径设置到页面 file input。',
+    waitForTitle: '等待条件',
+    waitForDescription: '等待 selector、文本或 JavaScript 表达式满足。',
+    handleDialogTitle: '处理弹窗',
+    handleDialogDescription: '接受或取消当前 JavaScript alert/confirm/prompt 弹窗。',
+    evaluateScriptTitle: '执行脚本',
+    evaluateScriptDescription: '在受控页面中执行 JavaScript 表达式并返回可序列化结果。',
+  },
+} as const;
