+++ core/memory/selector.ts
@@ -1,6 +1,6 @@
 import type { Memory } from '../types';
 import { MEMORY_TOKEN_BUDGET, STOP_WORDS } from '../constants';
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import { estimateTokens } from '../token/estimator';
 
 export { estimateTokens } from '../token/estimator';
