+++ core/prompt/augmentation.ts
@@ -1,6 +1,6 @@
 import { SHELL_TOOL_NAMES } from '../shell/contracts';
 import type { Memory, ToolDescriptor } from '../types';
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import {
   DEFAULT_TOOL_DESCRIPTORS,
   createDefaultToolDescriptors,
