+++ core/sandbox/tool.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import type { ToolCall, ToolDescriptor, ToolProviderIdentity, ToolResult } from '../tool/types';
 import type { SandboxLanguage, SandboxRunRequest } from './types';
