+++ core/tool/memory.ts
@@ -1,5 +1,5 @@
 import type { Memory, MemoryType, NewMemory } from '../types';
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import type {
   JsonValue,
   ToolCall,
