+++ core/tool/runtime.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import { appendToolCallHistory } from './history';
 import { ToolPostEffectPersistenceError } from './execution-error';
 import type {
