+++ core/tool/web-search.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n';
+import { DEFAULT_LOCALE, translate, type SupportedLocale } from '../i18n/background';
 import { createAbortScope } from '../network/abort';
 import type {
   JsonValue,
