+++ scripts/i18n-coverage-audit.mjs
@@ -22,7 +22,7 @@ const pathAllowlist = new Map([
   ['core/export/normalize.ts', 'generated export data fallback title; user-authored conversation content is preserved'],
   ['core/export/schema.ts', 'conversation export validation diagnostics are internal data-contract errors'],
   ['core/export/service.ts', 'conversation export progress copy is surfaced through background export flow and covered by the explicit export exception until export progress takes a locale parameter'],
-  ['core/i18n/resources/zh-CN.ts', 'zh-CN locale resource'],
+  ['core/i18n/resources/zh-CN', 'zh-CN locale resources (per-section modules; background bundle imports only its section subset)'],
   ['core/mcp/client.ts', 'MCP low-level tool-result diagnostics remain in the protocol layer; localized MCP UI state is covered separately'],
   ['core/mcp/discovery.ts', 'MCP low-level execution diagnostics remain in the protocol layer; localized MCP UI state is covered separately'],
   ['core/skill/builtin.ts', 'canonical built-in Skill source remains zh-CN; English projection is generated by getLocalizedBuiltinSkills'],
@@ -45,7 +45,7 @@ const pathAllowlist = new Map([
 
 const lineAllowlist = [
   {
-    path: 'core/i18n/resources/en.ts',
+    path: 'core/i18n/resources/en/locale.ts',
     includes: "zhCN: '简体中文'",
     reason: 'language selector uses the native name for Simplified Chinese',
   },
