+++ scripts/sidepanel-chunk-budget.mjs
@@ -37,9 +37,18 @@ if (requestedBrowsers.some((browser) => !browser)) {
 // was re-measured after merging main (#504): post-merge Node-25 measurement
 // 375456 raw / 112742 gzip; the initialShell below is set to the actual
 // post-merge build measurement so the budget guardrail stays meaningful.
+// Refreshed for the background i18n slimming (#505/#506): background-bundled
+// modules now import the slim core/i18n/background slice instead of the full
+// resource tree, which changed the shared-module graph and rollup's chunk
+// grouping. Content is essentially unchanged (+216 raw, +0.06%); the i18n
+// resources moved from their own chunk into the entry chunk and a new shared
+// runtime chunk appeared, so the gzip total regroups too. Local Node-25
+// measurement: 375672 raw / 114387 gzip; CI Node-22 zlib measures 114659 gzip
+// (+272 encoder variance); the baseline below uses the CI measurement so the
+// guardrail stays green on both runtimes.
 // The initial shell is sidepanel.html's entry script plus every static modulepreload.
 const BASELINE = Object.freeze({
-  initialShell: { raw: 375_456, gzip: 112_742 },
+  initialShell: { raw: 375_672, gzip: 114_659 },
   routeChunks: {
     ChatPage: { raw: 134_938, gzip: 40_056 },
     CapabilitiesPage: { raw: 160_137, gzip: 35_259 },
@@ -67,6 +76,10 @@ const WAVE_2_CHAT_BASELINE = Object.freeze({
 // entry (#495) also flow into the first chat screen graph. Local Node-25
 // measurement: 402288 raw / 121796 gzip (+2199 raw over the #475 refresh);
 // gzip stays within the existing 122000 cap.
+// Raised 405800 -> 406000 (gzip 123100 -> 125000) for the background i18n
+// slimming (#505/#506): the initial-shell chunk regrouping above (+216 raw)
+// also flows into the first chat screen graph (ChatPage itself unchanged).
+// Local Node-25 measurement: 405928 raw / 124407 gzip.
 
 // gzip size may vary slightly between supported Node/zlib releases even when
 // the built JavaScript bytes are identical. Keep the raw baseline exact and
@@ -78,7 +91,7 @@ const BUDGET = Object.freeze({
     raw: BASELINE.initialShell.raw,
     gzip: BASELINE.initialShell.gzip + GZIP_ENCODER_VARIANCE_BYTES,
   },
-  firstChatScreen: { raw: 405_800, gzip: 123_100 },
+  firstChatScreen: { raw: 406_000, gzip: 125_000 },
   richRendererIncrement: { raw: 120_000, gzip: 36_000 },
   routeChunks: {
     ChatPage: { raw: 25_000, gzip: 8_000 },
